### 2020-02-25 1922c79
Fix SciPy doi
