+++ paper.bib
@@ -35,7 +35,7 @@
              Computing in Python},
     journal = {Nature Methods},
     year = "2020",
-    doi = {https://doi.org/10.1038/s41592-019-0686-2},
+    doi = {10.1038/s41592-019-0686-2},
 }
 
 @article{NumPy,
