### 2020-02-25 2bfd60d
Typo fixes in docstrings
