### 2020-02-25 263c277
Fix bug in kA equation of condenser
