### 2020-02-25 40c1b31
Update doctests
