### 2020-02-25 94f6f2f
Adapt test to bug fixes and changes from PR 156
