+++ tests/heat_pump_test.py
@@ -280,9 +280,6 @@ class test_heat_pump:
 
         self.nw.solve('design')
         self.nw.save('tmp')
-        self.nw.save('tmp2')
-
-        self.nw.print_results()
 
         # input values from ebsilon
         T = [105, 100, 90, 80]
