### 2020-02-25 da7c717
Make PR template more clear
