+++ PULL_REQUEST_TEMPLATE.md
@@ -1,13 +1,27 @@
-* Describe your pull request as transparent as possible
-  * What functionality does it implement?
-  * Where is it located?
-  * How does the API look?
-  ...
+**General**
 
-* Related issues?
+Describe your pull request as transparent as possible:
 
-* Share your knowledge: Insights/Remarks
+* What functionality does it implement?
+* Where is it located?
+* How does the API look?
+* Which bugs are fixed
+* Are there related issues?
 
-* Other comments and questions
+**Documentation**
 
-* [ ] For new features: Remember the documentation!
+Please document all changes and/or new features:
+
+* Update the API documentation
+* Update the corresponding sections in the online documentation
+* Add inline comments to your code, if required
+* Update the corresponding What's New file
+
+**Testing**
+
+* Adapt the software tests (including doc-tests), if the PR does affect existing tests
+* Add new tests for new features
+
+**Questions and Comments**
+
+You are welcome to ask for help or start a discussion on a feature. We very much appreciate your contribution.
