### 2020-02-25 34db150
Add bus printout "eta"
Modified the bus_func methods of all components with a flag
to calculate the base value of the bus. The efficiency is
printed in the results overview of the busses.
