+++ tespy/components/turbomachinery.py
@@ -264,7 +264,7 @@ class turbomachine(component):
             s_mix = s_mix_pT(i, T_mix)
             return h_mix_ps(o, s_mix, T0=self.outl[0].T.val_SI)
 
-    def bus_func(self, bus, calc_base=False):
+    def bus_func(self, bus, calc_efficiency=False):
         r"""
         Calculate the residual value of the bus function.
 
@@ -273,7 +273,7 @@ class turbomachine(component):
         bus : tespy.connections.bus
             TESPy bus object.
 
-        calc_base : boolean
+        calc_efficiency : boolean
             Calculate bus base value without applying characteristcs.
 
         Returns
@@ -291,11 +291,13 @@ class turbomachine(component):
         val = i[0] * (o[2] - i[2])
         if np.isnan(bus.P_ref):
             expr = 1
-        elif calc_base is True:
-            return val
         else:
             expr = abs(val / bus.P_ref)
-        return val * bus.char.evaluate(expr)
+
+        if calc_efficiency is True:
+            return bus.char.evaluate(expr)
+        else:
+            return val * bus.char.evaluate(expr)
 
     def bus_deriv(self, bus):
         r"""
