+++ tespy/components/heat_exchangers.py
@@ -631,7 +631,7 @@ class heat_exchanger_simple(component):
 
         return i[0] * (o[2] - i[2]) + self.kA.val * fkA * td_log
 
-    def bus_func(self, bus, calc_base=False):
+    def bus_func(self, bus, calc_efficiency=False):
         r"""
         Calculate the residual value of the bus function.
 
@@ -640,7 +640,7 @@ class heat_exchanger_simple(component):
         bus : tespy.connections.bus
             TESPy bus object.
 
-        calc_base : boolean
+        calc_efficiency : boolean
             Calculate bus base value without applying characteristcs.
 
         Returns
@@ -660,11 +660,13 @@ class heat_exchanger_simple(component):
         val = i[0] * (o[2] - i[2])
         if np.isnan(bus.P_ref):
             expr = 1
-        elif calc_base is True:
-            return val
         else:
             expr = abs(val / bus.P_ref)
-        return val * bus.char.evaluate(expr)
+
+        if calc_efficiency is True:
+            return bus.char.evaluate(expr)
+        else:
+            return val * bus.char.evaluate(expr)
 
     def bus_deriv(self, bus):
         r"""
@@ -1756,7 +1758,7 @@ class heat_exchanger(component):
         return (self.ttd_l.val - T_mix_ph(o1, T0=self.outl[0].T.val_SI) +
                 T_mix_ph(i2, T0=self.inl[1].T.val_SI))
 
-    def bus_func(self, bus, calc_base=False):
+    def bus_func(self, bus, calc_efficiency=False):
         r"""
         Calculate the residual value of the bus function.
 
@@ -1765,7 +1767,7 @@ class heat_exchanger(component):
         bus : tespy.connections.bus
             TESPy bus object.
 
-        calc_base : boolean
+        calc_efficiency : boolean
             Calculate bus base value without applying characteristcs.
 
         Returns
@@ -1785,11 +1787,13 @@ class heat_exchanger(component):
         val = i[0] * (o[2] - i[2])
         if np.isnan(bus.P_ref):
             expr = 1
-        elif calc_base is True:
-            return val
         else:
             expr = abs(val / bus.P_ref)
-        return val * bus.char.evaluate(expr)
+
+        if calc_efficiency is True:
+            return bus.char.evaluate(expr)
+        else:
+            return val * bus.char.evaluate(expr)
 
     def bus_deriv(self, bus):
         r"""
