### 2020-02-26 5331a79
Add extrapolation for characteristics line
