### 2020-03-01 daee21e
Remove broken link in solar collector equations
