### 2020-03-03 405ba25
Update docs and simplify extrapolation implementation
