### 2020-03-03 300a85e
Add characteristic line inter- and extrapolationtests
