### 2020-03-03 2a3bef5
Add char_map and compressor_map tests
