+++ tests/tools_tests/characteristics_tests.py
@@ -12,7 +12,7 @@ SPDX-License-Identifier: MIT
 
 from nose.tools import eq_
 
-from tespy.tools.characteristics import (char_line, compressor_map,
+from tespy.tools.characteristics import (char_line, char_map, compressor_map,
                                          load_default_char, load_custom_char)
 
 from tespy.tools.helpers import extend_basic_path
@@ -112,8 +112,8 @@ class characteristics_loader_tests:
         eq_(True, z2_cond, msg)
 
 
-def test_characteristic_line():
-    """Test the characteristc line without extrapolation."""
+def test_char_line_evaluation():
+    """Test the characteristc line evaluation."""
 
     # create a characteristc line with values of y=(x-2)^2
     line = char_line(x=[0, 1, 2, 3, 4], y=[4, 1, 0, 1, 4])
@@ -121,62 +121,156 @@ def test_characteristic_line():
     # test evaluation at given x value (x=3, y=1)
     x = 3
     y = line.evaluate(x)
-    msg = ("The evaluation of x=" + str(x) + " must be 1.0, but is" +
+    msg = ("The evaluation of x=" + str(x) + " must be 1.0, but is " +
            str(y) + ".")
     eq_(1.0, round(y, 1), msg)
 
     # test evaluation at x=0.5 to force interpolation, result: y=2.5
     x = 0.5
     y = line.evaluate(x)
-    msg = ("The evaluation of x=" + str(x) + " must be 2.5, but is" +
+    msg = ("The evaluation of x=" + str(x) + " must be 2.5, but is " +
            str(y) + ".")
     eq_(2.5, round(y, 1), msg)
 
     # test evaluation at x=-1 to check lower limits, result: y=4
     x = -1
     y = line.evaluate(x)
-    msg = ("The evaluation of x=" + str(x) + " must be 4, but is" +
+    msg = ("The evaluation of x=" + str(x) + " must be 4, but is " +
            str(y) + ".")
     eq_(4, round(y, 1), msg)
 
     # test evaluation at x=5 to check upper limits, result: y=4
     x = 5
     y = line.evaluate(x)
-    msg = ("The evaluation of x=" + str(x) + " must be 4, but is" +
+    msg = ("The evaluation of x=" + str(x) + " must be 4, but is " +
            str(y) + ".")
     eq_(4, round(y, 1), msg)
 
 
-def test_extrapolation():
+def test_char_line_extrapolation():
     """Test the characteristc line with extrapolation."""
 
     # create a characteristc line with values of y=(x-2)^2
     line = char_line(x=[0, 1, 2, 3, 4], y=[4, 1, 0, 1, 4], extrapolate=True)
 
-    # test evaluation at given x value (x=3, y=1)
-    x = 1
-    y = line.evaluate(x)
-    msg = ("The evaluation of x=" + str(x) + " must be 1.0, but is" +
-           str(y) + ".")
-    eq_(1.0, round(y, 1), msg)
-
-    # test evaluation at x=0.5 to force interpolation, result: y=2.5
-    x = 0.5
-    y = line.evaluate(x)
-    msg = ("The evaluation of x=" + str(x) + " must be 2.5, but is" +
-           str(y) + ".")
-    eq_(2.5, round(y, 1), msg)
-
     # test evaluation at x=-1 to check lower limits, result: y=7
     x = -1
[…]
