+++ doc/whats_new/v0-2-2.rst
@@ -9,6 +9,11 @@ New Features
   disable this feature, you need to state
   :code:`mynetwork.solve(..., init_previous=False)`
   (`PR #156 <https://github.com/oemof/tespy/pull/156>`_).
+- Extrapolation for characteristic lines is available. In default state, the
+  upper or lower value range limit is used when a characteristic line is
+  evaluated outside of the available x-value range. The :code:`extrapolate`
+  parameter allows linear extrapolation, for an example see the section below
+  (`PR #159 <https://github.com/oemof/tespy/pull/159>`_).
 
 Documentation
 #############
@@ -23,6 +28,8 @@ Testing
   (`PR #153 <https://github.com/oemof/tespy/pull/153>`_).
 - Improve coverage of the networks module
   (`PR #153 <https://github.com/oemof/tespy/pull/153>`_).
+- Add tests for characteristic line and map evaluation
+  (`PR #159 <https://github.com/oemof/tespy/pull/159>`_).
 
 Bug fixes
 #########
@@ -45,3 +52,4 @@ Contributors
 ############
 
 - Francesco Witte (@fwitte)
+- @maltefritz
