### 2020-03-03 3cc6b5b
Update characteristics tests
