### 2020-03-04 5f58e3d
Add missing kwargs for zeta3 specification
