+++ doc/whats_new/v0-2-2.rst
@@ -5,7 +5,7 @@ New Features
 ############
 - Add a new component evaporator for geothermal organic rankine cycle. The
   component has inlets for geothermal steam brine. On the cold side, the orc
-  working fluid is evaporated. Read more about this component in the API 
+  working fluid is evaporated. Read more about this component in the API
   documentation: :py:class:`tespy.components.customs.orc_evaporator`
   (`PR #148 <https://github.com/oemof/tespy/pull/148>`_).
 
@@ -39,3 +39,4 @@ Contributors
 ############
 
 - Francesco Witte (@fwitte)
+- @ChaofanChen
