### 2020-03-04 9689648
Readd missing orc imports and tests
