### 2020-03-04 3cf023c
Fix typo in fluid initialisation
