### 2020-03-04 dc29800
Add results printout to heat_pump_test
