+++ tests/heat_pump_test.py
@@ -280,6 +280,7 @@ class test_heat_pump:
 
         self.nw.solve('design')
         self.nw.save('tmp')
+        self.nw.print_results()
 
         # input values from ebsilon
         T = [105, 100, 90, 80]
