### 2020-03-04 616508f
Fix typos in What's New
