### 2020-03-04 #2109 @fwitte
Hi @kyleniemeyer,
is it possible to update the software version submitted to v0.2.2? I would like to publish v0.2.2 within the next days as some bugs in the code and the documentation have been fixed in the meantime. Thank you very much!
Best regards
