+++ CONTRIBUTING.md
@@ -1,0 +1,1 @@
+The developer rules can be found in the chapter [developing TESPy](https://tespy.readthedocs.io/en/latest/developing_tespy.html) of the tespy.documentation.
