+++ doc/whats_new/v0-2-2.rst
@@ -1,5 +1,5 @@
-v0.2.2 - (March, someday, 2020)
-+++++++++++++++++++++++++++++++
+v0.2.2 Rankine's Realm - (March, someday, 2020)
++++++++++++++++++++++++++++++++++++++++++++++++
 
 New Features
 ############
