### 2020-03-06 ae76689
Update characteristic map documentation
