### 2020-03-06 79cf3c2
Add note for compressor maps
