### 2020-03-06 076aca8
Add bibtex package for sphinx
