### 2020-03-06 14a3249
Add references bibtex file
