+++ doc/references.bib
@@ -1,0 +1,72 @@
+@book{Baehr2016,
+    doi = {10.1007/978-3-662-49568-1},
+    year = {2016},
+    publisher = {Springer Berlin Heidelberg},
+    author = {Baehr, Hans Dieter and Kabelac. Stephan},
+    title = {Thermodynamik}
+}
+
+@article{Bell2014,
+    author = {Bell, Ian H. and Wronski, Jorrit and Quoilin, Sylvain and
+        Lemort, Vincent},
+    title = {Pure and Pseudo-pure Fluid Thermophysical Property Evaluation
+        and the Open-Source Thermophysical Property Library CoolProp},
+    journal = {Industrial \& Engineering Chemistry Research},
+    volume = {53},
+    number = {6},
+    pages = {2498-2508},
+    year = {2014},
+    doi = {10.1021/ie4033999}
+}
+
+@book{Bswirth2012,
+  doi = {10.1007/978-3-8348-8647-7},
+  year = {2012},
+  publisher = {Vieweg$\mathplus$Teubner Verlag},
+  author = {B\"{o}swirth, Leopold and Bschorer, Sabine},
+  title = {Technische Str\"{o}mungslehre}
+}
+
+@book{Epple2012,
+    doi = {10.1007/978-3-7091-1182-6},
+    year = {2012},
+    publisher = {Springer Vienna},
+    editor = {Epple, Bernd and Leithner, Reinhard and Linzer, Wladimir and
+    Walter, Heimo},
+    title = {Simulation von Kraftwerken und Feuerungen}
+}
+
+@misc{GasTurb2018,
+    author = {GasTurb GmbH},
+    title = {GasTurb 13: Design and Off-Design Performance of Gas Turbines},
+    year = {2018}
+}
+
+@article{Herning1936,
+    author = {Herning, Fritz and Zipperer, L},
+    title = {Calculation of the Viscosity of Technical Gas Mixtures From the
+        Viscosity of the Individual Gases},
+    journal = {Gas-und Wasserfach},
+    volume = {79},
+    pages = {69-73},
+    year = {2015}
+}
+
+@article{Pils2016,
+    author = {Plis, Marcin and Rusinowski, Henryk},
+    title = {Mathematical Modeling of an Axial Compressor in a Gas Turbine
+        Cycle.},
+    journal = {Journal of Power Technologies},
+    volume = {96},
+    number = {3},
+    pages = {194-199},
+    year = {2016}
+}
+
+@book{Traupel2001,
+    doi = {10.1007/978-3-642-17465-0},
+    year = {2001},
+    publisher = {Springer Berlin Heidelberg},
+    author = {Walter Traupel},
+    title = {Thermische Turbomaschinen}
+}
