### 2020-03-06 8b37699
Add bibliography file
