+++ doc/references.bib
@@ -22,7 +22,7 @@
 @book{Bswirth2012,
   doi = {10.1007/978-3-8348-8647-7},
   year = {2012},
-  publisher = {Vieweg$\mathplus$Teubner Verlag},
+  publisher = {Vieweg+Teubner Verlag},
   author = {B\"{o}swirth, Leopold and Bschorer, Sabine},
   title = {Technische Str\"{o}mungslehre}
 }
@@ -61,6 +61,13 @@
     number = {3},
     pages = {194-199},
     year = {2016}
+}
+
+@book{Quaschning2013,
+    year = {2013},
+    publisher = {Carl Hanser Verlag M\"{u}nchen},
+    author = {Quaschning, Volker},
+    title = {Regenerative Energiesysteme}
 }
 
 @book{Traupel2001,
