### 2020-03-06 35aba05
Add reference fiel for characteristics
