+++ doc/references.bib
@@ -52,7 +52,7 @@
     year = {2015}
 }
 
-@article{Pils2016,
+@article{Plis2016,
     author = {Plis, Marcin and Rusinowski, Henryk},
     title = {Mathematical Modeling of an Axial Compressor in a Gas Turbine
         Cycle.},
@@ -76,4 +76,11 @@
     publisher = {Springer Berlin Heidelberg},
     author = {Walter Traupel},
     title = {Thermische Turbomaschinen}
+}
+
+@misc{Vissmann2020,
+    author = {Viessmann Werke GmbH&Co KG},
+    title = {Adapted from Technische Beschreibung VITOBLOC 200},
+    year = {2020},
+    howpublished = {\url{https://www.viessmann.de/de/gewerbe/kraft-waerme-kopplung/blockheizkraftwerk/vitobloc-200-em-530-660.html}}
 }
