### 2020-03-06 4f1c362
Move references to corresponding equations
