### 2020-03-06 7cf7384
Fix image rendering
