+++ doc/api/modules.rst
@@ -1,6 +1,12 @@
 tespy
 =====
 
+All component and connection property equations derive from balance equations
+for fluid composition, mass flow and energy in regarding thermal as well as
+hydraulic state and thermodynamic fluid property equations respectively.
+Standard literature is for example :cite:`Baehr2016,Epple2012,Bswirth2012`
+(german). Equations and properties from other sources are cited individually.
+
 .. toctree::
    :maxdepth: 4
 
