+++ doc/references.bib
@@ -77,10 +77,3 @@
     author = {Walter Traupel},
     title = {Thermische Turbomaschinen}
 }
-
-@misc{Vissmann2020,
-    author = {Viessmann Werke GmbH&Co KG},
-    title = {Adapted from Technische Beschreibung VITOBLOC 200},
-    year = {2020},
-    howpublished = {\url{https://www.viessmann.de/de/gewerbe/kraft-waerme-kopplung/blockheizkraftwerk/vitobloc-200-em-530-660.html}}
-}
