### 2020-03-06 0f8a6cd
Add all flag
