+++ doc/zliterature.rst
@@ -2,3 +2,4 @@ Literature
 ==========
 
 .. bibliography:: references.bib
+    :all:
