+++ doc/index.rst
@@ -1,8 +1,3 @@
-.. TESPY documentation master file, created by
-   sphinx-quickstart on Sat Nov 18 11:20:45 2017.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
 Welcome to TESPy's documentation!
 =================================
 
