+++ doc/whats_new/v0-2-3.rst
@@ -1,0 +1,23 @@
+v0.2.3 - (somemonth, someday, 2020)
++++++++++++++++++++++++++++++++++++
+
+New Features
+############
+
+Documentation
+#############
+
+Parameter renaming
+##################
+
+Testing
+#######
+
+Bug fixes
+#########
+
+Other changes
+#############
+
+Contributors
+############
