### 2020-03-09 58e3633
Adjust tests to work with pytest
