+++ tests/error_tests.py
@@ -9,8 +9,7 @@ tests/error_tests.py
 
 SPDX-License-Identifier: MIT
 """
-
-from nose.tools import eq_, raises
+import pytest
 
 from tespy.connections import connection, bus, ref
 from tespy.components import (basics, combustion, components, heat_exchangers,
@@ -41,66 +40,66 @@ class specification_error_tests:
         self.bus = bus('mybus')
         self.sub = subsystems.subsystem('MySub')
 
-    @raises(ValueError)
     def cmp_instanciation_ValueError(self, label):
-        combustion.combustion_engine(label)
+        with pytest.raises(ValueError):
+            combustion.combustion_engine(label)
 
-    @raises(ValueError)
     def set_attr_ValueError(self, instance, **kwargs):
-        instance.set_attr(**kwargs)
+        with pytest.raises(ValueError):
+            instance.set_attr(**kwargs)
 
-    @raises(ValueError)
     def create_connection_ValueError(self, pos):
-        if pos == 'source':
-            connection(self.comp, 'out6', self.pipe, 'in1')
-        elif pos == 'target':
-            connection(self.comp, 'out1', self.pipe, 'in5')
+        with pytest.raises(ValueError):
+            if pos == 'source':
+                connection(self.comp, 'out6', self.pipe, 'in1')
+            elif pos == 'target':
+                connection(self.comp, 'out1', self.pipe, 'in5')
 
-    @raises(TypeError)
     def set_attr_TypeError(self, instance, **kwargs):
-        instance.set_attr(**kwargs)
+        with pytest.raises(TypeError):
+            instance.set_attr(**kwargs)
 
-    @raises(TypeError)
     def bus_add_comps_TypeError(self, c):
-        self.bus.add_comps(c)
+        with pytest.raises(TypeError):
+            self.bus.add_comps(c)
 
-    @raises(TypeError)
     def test_create_conn_TypeError(self):
-        connection(self.comp, 'out1', self.conn, 'in1')
+        with pytest.raises(TypeError):
+            connection(self.comp, 'out1', self.conn, 'in1')
 
-    @raises(TypeError)
     def create_ref_TypeError(self, params):
-        ref(params[0], params[1], params[2])
+        with pytest.raises(TypeError):
+            ref(params[0], params[1], params[2])
 
-    @raises(KeyError)
-    def set_attr_KeyError(self, instance, **kwargs):
-        instance.set_attr(**kwargs)
-
-    @raises(KeyError)
-    def get_attr_KeyError(self, instance, key):
-        instance.get_attr(key)
-
-    @raises(TESPyConnectionError)
-    def test_con_TESPyConnectionError(self):
-        connection(self.comp, 'out1', self.comp, 'in1')
-
-    @raises(TESPyConnectionError)
-    def test_bus_TESPyConnectionError(self):
-        self.bus.add_comps(self.comp)
-
-    @raises(TESPyNetworkError)
-    def test_network_bus_duplicate(self):
-        self.nw.add_busses(self.bus, self.bus)
-
-    @raises(TESPyNetworkError)
-    def test_network_buslabel_duplicate(self):
-        b = bus('mybus')
-        self.nw.add_busses(self.bus)
-        self.nw.add_busses(b)
-
-    @raises(TypeError)
     def test_network_bus_type(self):
-        self.nw.add_busses(self.conn)
+        with pytest.raises(TypeError):
+            self.nw.add_busses(self.conn)
+
+    def set_attr_KeyError(self, instance, **kwargs):
+        with pytest.raises(KeyError):
+            instance.set_attr(**kwargs)
+
+    def get_attr_KeyError(self, instance, key):
+        with pytest.raises(KeyError):
[…]
