### 2020-03-09 89b20fe
Adjust test to work with pytest
