+++ .travis.yml
@@ -10,14 +10,14 @@ matrix:
       python: 3.7
       dist: xenial
       sudo: true
-#    - name: "python 3.8 on xenial Linux"
-#      python: 3.8
-#      dist: xenial
-#      sudo: true
-#    - name: "Python 3.7.2 on macOS"
-#      os: osx
-#      osx_image: xcode10.2
-#      language: shell
+    - name: "python 3.8 on xenial Linux"
+      python: 3.8
+      dist: xenial
+      sudo: true
+    - name: "Python 3.7.2 on macOS"
+      os: osx
+      osx_image: xcode10.2
+      language: shell
 #    - name: "Python 3.7.3 on Windows"
 #      os: windows
 #      language: shell
@@ -25,14 +25,13 @@ matrix:
 #      env: PATH=/c/Python37:/c/Python37/Scripts:$PATH
 
 install:
-  - pip3 install .
-  - pip3 install coveralls
-  - pip3 install nose
+  - pip install .
+  - pip install pytest-cov
+  - pip install coveralls
 
 # command to run tests
 script:
-  - python3 -m nose --with-doctest --with-coverage -c nose.cfg || python -m nose --with-doctest --with-coverage -c nose.cfg
+  - python -m pytest
 
 after_success:
-- coveralls
-
+  - coveralls
