### 2020-03-09 a58ebb5
Adjust test handling
