### 2020-03-09 c9e59e9
Add tests and doctests
