### 2020-03-10 bbdea65
Update test configuration
