+++ .travis.yml
@@ -31,7 +31,7 @@ install:
 
 # command to run tests
 script:
-  - python -m pytest --cov=tests --doctest-modules
+  - python -m pytest .tespy/ --doctest-modules ./tests --cov=tespy
 
 after_success:
   - coveralls
