+++ tests/test_errors.py
@@ -1,0 +1,640 @@
+# -*- coding: utf-8
+
+"""Module for testing program errors.
+
+This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
+by the contributors recorded in the version control history of the file,
+available from its original location
+tests/error_tests.py
+
+SPDX-License-Identifier: MIT
+"""
+from pytest import raises
+
+from tespy.connections import connection, bus, ref
+from tespy.components import (basics, combustion, components, heat_exchangers,
+                              nodes, piping, reactors, subsystems,
+                              turbomachinery)
+from tespy.networks.network_reader import load_network
+from tespy.networks.networks import network
+from tespy.tools.helpers import (TESPyComponentError, TESPyConnectionError,
+                                 TESPyNetworkError)
+from tespy.tools.data_containers import data_container, dc_cc, dc_cp, dc_flu
+from tespy.tools.fluid_properties import tespy_fluid
+from tespy.tools.characteristics import char_map, char_line, load_custom_char
+
+import shutil
+import csv
+
+##############################################################################
+# test errors of set_attr and get_attr methods
+
+
+def set_attr_ValueError(instance, **kwargs):
+    with raises(ValueError):
+        instance.set_attr(**kwargs)
+
+
+def set_attr_TypeError(instance, **kwargs):
+    with raises(TypeError):
+        instance.set_attr(**kwargs)
+
+
+def set_attr_KeyError(instance, **kwargs):
+    with raises(KeyError):
+        instance.set_attr(**kwargs)
+
+
+def get_attr_KeyError(instance, key):
+    with raises(KeyError):
+        instance.get_attr(key)
+
+
+def test_set_attr_errors():
+    """Test errors of set_attr methods."""
+    nw = network(['water', 'air'])
+    comb = combustion.combustion_engine('combustion engine')
+    pipeline = piping.pipe('pipeline')
+    conn = connection(comb, 'out1', pipeline, 'in1')
+    mybus = bus('mybus')
+    sub = subsystems.subsystem('MySub')
+
+    # ValueErrors
+    set_attr_ValueError(comb, offdesign=['Q'])
+
+    set_attr_ValueError(conn, offdesign=['f'])
+    set_attr_ValueError(conn, state='f')
+
+    set_attr_ValueError(nw, m_unit='kg')
+    set_attr_ValueError(nw, h_unit='kg')
+    set_attr_ValueError(nw, p_unit='kg')
+    set_attr_ValueError(nw, T_unit='kg')
+    set_attr_ValueError(nw, v_unit='kg')
+
+    # TypeErrors
+    set_attr_TypeError(comb, P=[5])
+    set_attr_TypeError(comb, tiP_char=None)
+    set_attr_TypeError(comb, design='f')
+    set_attr_TypeError(comb, lamb=dc_cc())
+    set_attr_TypeError(comb, design_path=7)
+    set_attr_TypeError(comb, local_design=5)
+    set_attr_TypeError(comb, local_offdesign=5)
+    set_attr_TypeError(pipeline, hydro_group=5)
+    set_attr_TypeError(comb, printout=5)
+
+    set_attr_TypeError(conn, design='h')
+    set_attr_TypeError(conn, fluid_balance=1)
+    set_attr_TypeError(conn, h0=[4])
+    set_attr_TypeError(conn, fluid=5)
+    set_attr_TypeError(conn, state=5)
+    set_attr_TypeError(conn, design_path=5)
+    set_attr_TypeError(conn, local_design=5)
+    set_attr_TypeError(conn, local_offdesign=5)
+    set_attr_TypeError(conn, printout=5)
+
[…]
