+++ tests/error_tests.py
@@ -1,658 +1,0 @@
-# -*- coding: utf-8
-
-"""Module for testing program errors.
-
-This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
-by the contributors recorded in the version control history of the file,
-available from its original location
-tests/error_tests.py
-
-SPDX-License-Identifier: MIT
-"""
-import pytest
-
-from tespy.connections import connection, bus, ref
-from tespy.components import (basics, combustion, components, heat_exchangers,
-                              nodes, piping, reactors, subsystems,
-                              turbomachinery)
-from tespy.networks.network_reader import load_network
-from tespy.networks.networks import network
-from tespy.tools.helpers import (TESPyComponentError, TESPyConnectionError,
-                                 TESPyNetworkError)
-from tespy.tools.data_containers import data_container, dc_cc, dc_cp, dc_flu
-from tespy.tools.fluid_properties import tespy_fluid
-from tespy.tools.characteristics import char_map, char_line, load_custom_char
-
-import shutil
-import csv
-
-
-# %% bulk tests
-
-
-class specification_error_tests:
-
-    def setup(self):
-        self.nw = network(['water', 'air'])
-        self.comp = combustion.combustion_engine('combustion engine')
-        self.pipe = piping.pipe('pipe')
-        self.conn = connection(self.comp, 'out1', self.pipe, 'in1')
-        self.bus = bus('mybus')
-        self.sub = subsystems.subsystem('MySub')
-
-    def cmp_instanciation_ValueError(self, label):
-        with pytest.raises(ValueError):
-            combustion.combustion_engine(label)
-
-    def set_attr_ValueError(self, instance, **kwargs):
-        with pytest.raises(ValueError):
-            instance.set_attr(**kwargs)
-
-    def create_connection_ValueError(self, pos):
-        with pytest.raises(ValueError):
-            if pos == 'source':
-                connection(self.comp, 'out6', self.pipe, 'in1')
-            elif pos == 'target':
-                connection(self.comp, 'out1', self.pipe, 'in5')
-
-    def set_attr_TypeError(self, instance, **kwargs):
-        with pytest.raises(TypeError):
-            instance.set_attr(**kwargs)
-
-    def bus_add_comps_TypeError(self, c):
-        with pytest.raises(TypeError):
-            self.bus.add_comps(c)
-
-    def test_create_conn_TypeError(self):
-        with pytest.raises(TypeError):
-            connection(self.comp, 'out1', self.conn, 'in1')
-
-    def create_ref_TypeError(self, params):
-        with pytest.raises(TypeError):
-            ref(params[0], params[1], params[2])
-
-    def test_network_bus_type(self):
-        with pytest.raises(TypeError):
-            self.nw.add_busses(self.conn)
-
-    def set_attr_KeyError(self, instance, **kwargs):
-        with pytest.raises(KeyError):
-            instance.set_attr(**kwargs)
-
-    def get_attr_KeyError(self, instance, key):
-        with pytest.raises(KeyError):
-            instance.get_attr(key)
-
-    def test_con_TESPyConnectionError(self):
-        with pytest.raises(TESPyConnectionError):
-            connection(self.comp, 'out1', self.comp, 'in1')
-
-    def test_bus_TESPyConnectionError(self):
-        with pytest.raises(TESPyConnectionError):
-            self.bus.add_comps(self.comp)
-
-    def test_network_bus_duplicate(self):
-        with pytest.raises(TESPyNetworkError):
-            self.nw.add_busses(self.bus, self.bus)
-
[…]
