### 2020-03-10 b37b832
Remove python3.8 and mac tests
