### 2020-03-10 123a17b
Update error messages of stoichiometric combustion chamber
