### 2020-03-10 fd2cb0e
Update tests to work with pytest
