+++ tests/test_components/test_heat_exchangers.py
@@ -5,12 +5,10 @@
 This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
 by the contributors recorded in the version control history of the file,
 available from its original location
-tests/component_tests/heat_exchanger_tests.py
+tests/test_components/test_heat_exchangers.py
 
 SPDX-License-Identifier: MIT
 """
-
-from nose.tools import eq_
 
 from tespy.components.basics import sink, source
 from tespy.components.heat_exchangers import (heat_exchanger_simple,
@@ -29,7 +27,7 @@ import shutil
 def convergence_check(lin_dep):
     """Check convergence status of a simulation."""
     msg = 'Calculation did not converge!'
-    eq_(lin_dep, False, msg)
+    assert lin_dep is False, msg
 
 
 class heat_exchanger_tests:
@@ -76,9 +74,9 @@ class heat_exchanger_tests:
         instance.kA_group.is_set = True
         self.nw.solve('design', init_only=True)
         msg = ('Hydro group must no be set, if one parameter is missing!')
-        eq_(instance.hydro_group.is_set, False, msg)
+        assert instance.hydro_group.is_set is False, msg
         msg = ('kA group must no be set, if one parameter is missing!')
-        eq_(instance.kA_group.is_set, False, msg)
+        assert instance.kA_group.is_set is False, msg
 
         # test diameter calculation from specified dimensions (as pipe)
         # with Hazen-Williams method
@@ -92,7 +90,7 @@ class heat_exchanger_tests:
         pr = round(self.c2.p.val_SI / self.c1.p.val_SI, 3)
         msg = ('Value of pressure ratio must be ' + str(pr) + ', is ' +
                str(instance.pr.val) + '.')
-        eq_(pr, round(instance.pr.val, 3), msg)
+        assert pr == round(instance.pr.val, 3), msg
 
         # make zeta system variable and use previously calculated diameter
         # to calculate zeta. The value for zeta must not change
@@ -103,7 +101,7 @@ class heat_exchanger_tests:
         convergence_check(self.nw.lin_dep)
         msg = ('Value of zeta must be ' + str(zeta) + ', is ' +
                str(round(instance.zeta.val, 0)) + '.')
-        eq_(zeta, round(instance.zeta.val, 0), msg)
+        assert zeta == round(instance.zeta.val, 0), msg
 
         # same test with pressure ratio as sytem variable
         pr = round(instance.pr.val, 3)
@@ -112,7 +110,7 @@ class heat_exchanger_tests:
         convergence_check(self.nw.lin_dep)
         msg = ('Value of pressure ratio must be ' + str(pr) +
                ', is ' + str(round(instance.pr.val, 3)) + '.')
-        eq_(pr, round(instance.pr.val, 3), msg)
+        assert pr == round(instance.pr.val, 3), msg
 
         # test heat transfer coefficient as variable of the system (ambient
         # temperature required)
@@ -125,7 +123,7 @@ class heat_exchanger_tests:
         # somewhere near to that (actual value is 677)
         msg = ('Value of heat transfer coefficient must be 667, is ' +
                str(instance.kA.val) + '.')
-        eq_(677, round(instance.kA.val, 0), msg)
+        assert 677 == round(instance.kA.val, 0), msg
 
         # test heat transfer as variable of the system
         instance.set_attr(Q='var', kA=np.nan)
@@ -135,7 +133,7 @@ class heat_exchanger_tests:
         convergence_check(self.nw.lin_dep)
         msg = ('Value of heat transfer must be ' + str(Q) +
[…]
