+++ tests/test_errors.py
@@ -5,7 +5,7 @@
 This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
 by the contributors recorded in the version control history of the file,
 available from its original location
-tests/error_tests.py
+tests/test_errors.py
 
 SPDX-License-Identifier: MIT
 """
