### 2020-03-10 2aaeb5d
Change error type in bus.add_comps
