### 2020-03-10 69d30bc
Fix assertions of ttd
