### 2020-03-10 ee7260f
Modify tests to keep custom characteristics files
