### 2020-03-10 310a88f
Delete deprecated config
