+++ nose.cfg
@@ -1,3 +1,0 @@
-[nosetests]
-with-coverage=1
-cover-package=tespy
