### 2020-03-10 85eca74
Update documentation for Darcy friction formulae
