+++ doc/references.bib
@@ -52,6 +52,26 @@
     year = {2015}
 }
 
+@Inbook{Nirschl2018,
+    author={Nirschl, Hermann},
+    editor={Stephan, Peter
+    and Mewes, Dieter
+    and Kabelac, Stephan
+    and Kind, Matthias
+    and Schaber, Karlheinz
+    and Wetzel, Thomas},
+    title={Druckverlust in durchstr{\"o}mten Rohren},
+    bookTitle={
+        VDI-W{\"a}rmeatlas : Fachlicher Tr{\"a}ger VDI-Gesellschaft
+        Verfahrenstechnik und Chemieingenieurwesen
+    },
+    year={2018},
+    publisher={Springer Berlin Heidelberg},
+    address={Berlin, Heidelberg},
+    pages={1-8},
+    doi={10.1007/978-3-662-52991-1_75-2}
+}
+
 @article{Plis2016,
     author = {Plis, Marcin and Rusinowski, Henryk},
     title = {Mathematical Modeling of an Axial Compressor in a Gas Turbine
