### 2020-03-10 5634760
Fix typo in equation documentation
