+++ doc/whats_new/v0-2-2.rst
@@ -1,4 +1,4 @@
-v0.2.2 Rankine's Realm - (March, 6, 2020)
+v0.2.2 - Rankine's Realm (March, 6, 2020)
 +++++++++++++++++++++++++++++++++++++++++
 
 New Features
