+++ doc/whats_new/v0-2-3.rst
@@ -1,5 +1,5 @@
-v0.2.3 - (somemonth, someday, 2020)
-+++++++++++++++++++++++++++++++++++
+v0.2.3 - Someversion (somemonth, someday, 2020)
++++++++++++++++++++++++++++++++++++++++++++++++
 
 New Features
 ############
