+++ tespy/networks/networks.py
@@ -897,8 +897,9 @@ class network:
         Note
         ----
         The methods
-        :func:`tespy.networks.network.init_comp_design_params` (components) and
-        the :func:`tespy.networks.network.init_conn_design_params`
+        :func:`tespy.networks.networks.network.init_comp_design_params`
+        (components) and the
+        :func:`tespy.networks.networks.network.init_conn_design_params`
         (connections) handle the parameter specification.
         """
         # components without any parameters
