### 2020-03-12 191df94
Add linkcheck script
