+++ .travis.yml
@@ -18,15 +18,10 @@ install:
   - pip install .
   - pip install pytest-cov
   - pip install coveralls
-  - pip install sphinx
-  - pip install sphinx_rtd_theme
-  - pip install sphinx_autoapi
 
 # command to run tests
 script:
   - python -m pytest ./tespy --doctest-modules ./tests --cov=tespy
-  - ./linkcheck.sh
-
 
 after_success:
   - coveralls
