### 2020-03-12 bd65292
Add missing equation for heat transfer in parameter list
