### 2020-03-27 f7a33f2
Add connection property output for plotting
Change the connection properties "s", "s_id", "t" and "t_id"
to "source" and "target" respectively.
