+++ tespy/networks/networks.py
@@ -208,11 +208,23 @@ class network:
             'kJ / kg': 1e3,
             'MJ / kg': 1e6
         }
+        # specific volume
+        self.vol = {
+            'm3 / kg': 1,
+            'l / kg': 1e-3
+        }
+        # entropy
+        self.s = {
+            'J / kgK': 1,
+            'kJ / kgK': 1e3,
+            'MJ / kgK': 1e6
+        }
         # temperature
         self.T = {
             'C': [273.15, 1],
             'F': [459.67, 5 / 9],
-            'K': [0, 1]
+            'K': [0, 1],
+            'R': [0, 5 / 9]
         }
         # volumetric flow
         self.v = {
@@ -221,14 +233,34 @@ class network:
             'm3 / h': 1 / 3600,
             'l / h': 1 / 3.6
         }
+        # vapor mass fraction
+        self.x = {
+            '-': 1,
+            '%': 1e-2
+        }
         # SI unit specification
         self.SI_units = {
               'm': 'kg / s',
+              'v': 'm3 / s',
               'p': 'Pa',
               'h': 'J / kg',
               'T': 'K',
-              'v': 'm3 / s'
+              'vol': 'm3 / kg',
+              'x': '-',
+              's': 'J / kgK'
               }
+
+        # written propteries
+        self.props = {
+            'm': 'mass flow',
+            'v': 'volumetric flow',
+            'p': 'pressure',
+            'h': 'enthalpy',
+            'T': 'temperature',
+            'vol': 'specific volume',
+            'x': 'vapour mass fraction',
+            's': 'entropy'
+        }
 
         # iterinfo
         self.iterinfo = True
@@ -239,12 +271,18 @@ class network:
         self.h_unit = self.SI_units['h']
         self.T_unit = self.SI_units['T']
         self.v_unit = self.SI_units['v']
+        self.vol_unit = self.SI_units['vol']
+        self.s_unit = self.SI_units['s']
+        self.x_unit = self.SI_units['x']
 
         msg = ('Default unit specifications: '
                'mass flow: ' + self.m_unit + ', ' +
                'pressure: ' + self.p_unit + ', ' +
                'enthalpy: ' + self.h_unit + ', ' +
                'temperature: ' + self.T_unit + ', ' +
+               'specific volume: ' + self.vol_unit + ', ' +
+               'entropy: ' + self.s_unit + ', ' +
+               'vapour mass fraction: ' + self.x_unit + ', ' +
                'volumetric flow: ' + self.v_unit + '.')
         logging.debug(msg)
 
@@ -314,56 +352,19 @@ class network:
             Print convergence progress to console.
         """
         # unit sets
-        if 'm_unit' in kwargs.keys():
-            if kwargs['m_unit'] not in self.m.keys():
-                msg = ('Allowed units for mass flow are: ' +
-                       str(self.m.keys()))
-                logging.error(msg)
-                raise ValueError(msg)
-            else:
-                self.m_unit = kwargs['m_unit']
-                msg = 'Setting mass flow unit: ' + self.m_unit
-                logging.debug(msg)
-
-        if 'p_unit' in kwargs.keys():
-            if kwargs['p_unit'] not in self.p.keys():
-                msg = ('Allowed units for pressure are: ' + str(self.p.keys()))
-                logging.error(msg)
-                raise ValueError(msg)
-            else:
-                self.p_unit = kwargs['p_unit']
-                msg = 'Setting pressure unit: ' + self.p_unit
-                logging.debug(msg)
-
-        if 'h_unit' in kwargs.keys():
-            if kwargs['h_unit'] not in self.h.keys():
-                msg = ('Allowed units for pressure are: ' + str(self.h.keys()))
[…]
