+++ tespy/connections.py
@@ -233,10 +233,10 @@ class connection:
             raise ValueError(msg)
 
         # set specified values
-        self.s = comp1
-        self.s_id = outlet_id
-        self.t = comp2
-        self.t_id = inlet_id
+        self.source = comp1
+        self.source_id = outlet_id
+        self.target = comp2
+        self.target_id = inlet_id
 
         # defaults
         self.new_design = True
@@ -255,8 +255,9 @@ class connection:
 
         self.set_attr(**kwargs)
 
-        msg = ('Created connection ' + self.s.label + ' (' + self.s_id +
-               ') -> ' + self.t.label + ' (' + self.t_id + ').')
+        msg = (
+            'Created connection ' + self.source.label + ' (' + self.source_id +
+            ') -> ' + self.target.label + ' (' + self.target_id + ').')
         logging.debug(msg)
 
     def set_attr(self, **kwargs):
@@ -524,6 +525,24 @@ class connection:
             logging.error(msg)
             raise KeyError(msg)
 
+    def get_plotting_props(self):
+        r"""
+        Get the connection's fluid properties for plotting.
+
+        Returns
+        -------
+        out : dict
+            Fluid properties of connection.
+        """
+        return {
+            'p': self.p.val,
+            'h': self.h.val,
+            'T': self.T.val,
+            's': self.s.val,
+            'v': self.vol.val,
+            'Q': self.x.val
+        }
+
     @staticmethod
     def attr():
         r"""
@@ -535,7 +554,8 @@ class connection:
             List of available attributes of a connection.
         """
         return {'m': dc_prop(), 'p': dc_prop(), 'h': dc_prop(), 'T': dc_prop(),
-                'x': dc_prop(), 'v': dc_prop(),
+                'x': dc_prop(), 'v': dc_prop(), 'vol': dc_prop(),
+                's': dc_prop(),
                 'fluid': dc_flu(), 'Td_bp': dc_prop(), 'state': dc_simple()}
 
     def to_flow(self):
@@ -936,8 +956,8 @@ class ref:
 
         msg = ('Created reference object with factor ' + str(self.f) +
                ' and delta ' + str(self.d) + ' referring to connection ' +
-               ref_obj.s.label + ' (' + ref_obj.s_id + ') -> ' +
-               ref_obj.t.label + ' (' + ref_obj.t_id + ').')
+               ref_obj.source.label + ' (' + ref_obj.source_id + ') -> ' +
+               ref_obj.target.label + ' (' + ref_obj.target_id + ').')
         logging.debug(msg)
 
     def get_attr(self, key):
