### 2020-04-06 7938ddd
Fix bug in logging
