### 2020-04-06 531813c
Rename vairables and attributes to clarify naming
