+++ tespy/networks/networks.py
@@ -154,7 +154,8 @@ class network:
         # in case of a design calculation after an offdesign calculation
         self.redesign = False
         # connection dataframe
-        self.conns = pd.DataFrame(columns=['s', 's_id', 't', 't_id'])
+        self.conns = pd.DataFrame(
+            columns=['source', 'source_id', 'target', 'target_id'])
         # list for busses
         self.busses = OrderedDict()
         # default design_path value
@@ -543,11 +544,13 @@ class network:
 
     def check_conns(self):
         r"""Check connections for multiple usage of inlets or outlets."""
-        dub = self.conns.loc[self.conns.duplicated(['s', 's_id']) == True]
+        dub = self.conns.loc[
+            self.conns.duplicated(['source', 'source_id']) == True]
         for c in dub.index:
             targets = ''
-            for conns in self.conns[(self.conns['s'] == c.source) &
-                                    (self.conns['s_id'] == c.source_id)].index:
+            for conns in self.conns[
+                    (self.conns['source'] == c.source) &
+                    (self.conns['source_id'] == c.source_id)].index:
                 targets += conns.target.label + ' (' + conns.target_id + '); '
 
             msg = (
@@ -558,11 +561,13 @@ class network:
             logging.error(msg)
             raise hlp.TESPyNetworkError(msg)
 
-        dub = self.conns.loc[self.conns.duplicated(['t', 't_id']) == True]
+        dub = self.conns.loc[
+            self.conns.duplicated(['target', 'target_id']) == True]
         for c in dub.index:
             sources = ''
-            for conns in self.conns[(self.conns['t'] == c.target) &
-                                    (self.conns['t_id'] == c.target_id)].index:
+            for conns in self.conns[
+                    (self.conns['target'] == c.target) &
+                    (self.conns['target_id'] == c.target_id)].index:
                 sources += conns.source.label + ' (' + conns.source_id + '); '
 
             msg = (
@@ -638,14 +643,14 @@ class network:
         r"""Check if components are connected properly within the network."""
         self.check_conns()
         # get unique components in connections dataframe
-        comps = pd.unique(self.conns[['s', 't']].values.ravel())
+        comps = pd.unique(self.conns[['source', 'target']].values.ravel())
         # build the dataframe for components
         self.init_components(comps)
         # count number of incoming and outgoing connections and compare to
         # expected values
         for comp in self.comps.index:
-            num_o = (self.conns[['s', 't']] == comp).sum().s
-            num_i = (self.conns[['s', 't']] == comp).sum().t
+            num_o = (self.conns[['source', 'target']] == comp).sum().source
+            num_i = (self.conns[['source', 'target']] == comp).sum().target
             if num_o != comp.num_o:
                 msg = (comp.label + ' is missing ' + str(comp.num_o - num_o) +
                        ' outgoing connections. Make sure all outlets are '
@@ -680,20 +685,20 @@ class network:
         connections. Thus it does not hold any additional information, the
         dataframe is used to simplify the code, only.
         """
-        self.comps = pd.DataFrame(index=comps, columns=['i', 'o'])
+        self.comps = pd.DataFrame(index=comps, columns=['inlets', 'outlets'])
 
         labels = []
         for comp in self.comps.index:
             # get for incoming and outgoing connections of a component
-            s = self.conns[self.conns['s'] == comp]
-            s = s.s_id.sort_values().index
-            t = self.conns[self.conns['t'] == comp]
-            t = t.t_id.sort_values().index
-            self.comps.loc[comp] = [t, s]
+            sources = self.conns[self.conns['source'] == comp]
+            sources = sources['source_id'].sort_values().index
[…]
