+++ tespy/networks/network_reader.py
@@ -93,7 +93,7 @@ def load_network(path):
     The structure of the path must be as follows:
 
     - Folder: path (e. g. 'mynetwork')
-    - Subfolder: comps (e. g. 'mynetwork/comps') containing
+    - Subfolder: components (e. g. 'mynetwork/components') containing
 
         - bus.csv*
         - char.csv*
@@ -212,7 +212,7 @@ def load_network(path):
     if path[-1] != '/' and path[-1] != '\\':
         path += '/'
 
-    path_comps = modify_path_os(path + 'comps/')
+    path_comps = modify_path_os(path + 'components/')
     path = modify_path_os(path)
 
     msg = 'Reading network data from base path ' + path + '.'
@@ -280,7 +280,7 @@ def load_network(path):
     nw.imp_comps = comps.to_dict()['instance']
 
     # load connections
-    fn = path + 'conn.csv'
+    fn = path + 'connections.csv'
     conns = pd.read_csv(fn, sep=';', decimal='.',
                         converters={'design': ast.literal_eval,
                                     'offdesign': ast.literal_eval})
@@ -454,7 +454,7 @@ def construct_network(path):
         TESPy network object.
     """
     # read network .csv-file
-    netw = pd.read_csv(path + 'netw.csv', sep=';', decimal='.',
+    netw = pd.read_csv(path + 'network.json', sep=';', decimal='.',
                        converters={'fluids': ast.literal_eval})
     f_list = netw['fluids'][0]
 
@@ -493,8 +493,8 @@ def construct_conns(c, *args):
         TESPy connection object.
     """
     # create connection
-    conn = connection(args[0].instance[c.s], c.s_id,
-                      args[0].instance[c.t], c.t_id)
+    conn = connection(args[0].instance[c.source], c.source_id,
+                      args[0].instance[c.target], c.target_id)
 
     kwargs = {}
     # read basic properties
