### 2020-04-06 1e63e40
Update tests according to renaming in networks module
