### 2020-04-06 11887f8
tespy/networks/network_reader.py
