+++ doc/whats_new/v0-2-3.rst
@@ -1,26 +1,0 @@
-v0.2.3 - Someversion (somemonth, someday, 2020)
-+++++++++++++++++++++++++++++++++++++++++++++++
-
-New Features
-############
-
-Documentation
-#############
-
-Parameter renaming
-##################
-
-Testing
-#######
-- Move from nose to pytest
-  (`PR #165 <https://github.com/oemof/tespy/pull/165>`_).
-
-Bug fixes
-#########
-
-Other changes
-#############
-
-Contributors
-############
-- Francesco Witte (`@fwitte <https://github.com/fwitte>`_)
