+++ tespy/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8
 
-__version__ = '0.2.3 - dev'
+__version__ = '0.3.0 - dev'
 
 # tespy networks imports
 from .networks import (
