### 2020-04-06 d5b897e
Add label attribute to connections
