### 2020-04-06 7bc2ae6
Handle connection and component labels for fast access
