### 2020-04-06 2321029
Add error test for connection label specification
