### 2020-04-06 1b7b3e3
Add doctest
