### 2020-04-06 1f53d13
Add label keyword in set_attr method
