### 2020-04-06 33df083
Add label field in network export and import
