+++ tespy/networks/network_reader.py
@@ -105,12 +105,14 @@ def load_network(path):
 
     The imported network has the following additional features:
 
-    - Connections are accessible by their target's label and id, e. g. for a
-      connection going to 'condenser' at inlet 'in2' use
-      :code:`myimportednetwork.imp_conns['condenser:in2']`.
-    - Components are accessible by label, e. g. for a component
-      'heat exchanger' :code:`myimportednetwork.imp_comps['heat exchanger']`.
-    - Busses are accessible by label, e. g. for a bus 'power input'
+    - Connections are accessible by label, e. g.
+      :code:`myimportednetwork.connections['myconnection']`. The default label
+      logic is :code:`source:source_id_target:target_id`, where the source
+      means the label of the component the connection originates from and
+      target means the label of the component, the connections targets on.
+    - Components are accessible by label as well, e. g. for a component
+      'heat exchanger' :code:`myimportednetwork.components['heat exchanger']`.
+    - Busses are accessible by label, too, e. g. for a bus 'power input'
       :code:`myimportednetwork.busses['power input']`.
 
     Example
@@ -137,7 +139,7 @@ def load_network(path):
     >>> comb = combustion_chamber('combustion')
     >>> t = turbine('turbine')
     >>> si = sink('sink')
-    >>> inc = connection(air, 'out1', c, 'in1')
+    >>> inc = connection(air, 'out1', c, 'in1', label='ambient air')
     >>> cc = connection(c, 'out1', comb, 'in1')
     >>> fc = connection(f, 'out1', comb, 'in2')
     >>> ct = connection(comb, 'out1', t, 'in1')
@@ -170,6 +172,7 @@ def load_network(path):
     >>> power.add_comps({'c': c}, {'c': t})
     >>> nw.add_busses(power)
     >>> nw.solve('design')
+    >>> mass_flow = round(nw.connections['ambient air'].m.val_SI, 1)
     >>> nw.save('exported_nwk')
     >>> c.set_attr(igva='var')
     >>> nw.solve('offdesign', design_path='exported_nwk',
@@ -193,19 +196,21 @@ def load_network(path):
     >>> imported_nwk = load_network('exported_nwk')
     >>> imported_nwk.set_attr(iterinfo=False)
     >>> imported_nwk.solve('design')
-    >>> round(imported_nwk.imp_comps['turbine'].eta_s.val, 3)
+    >>> round(imported_nwk.connections['ambient air'].m.val_SI, 1) == mass_flow
+    True
+    >>> round(imported_nwk.components['turbine'].eta_s.val, 3)
     0.9
-    >>> imported_nwk.imp_comps['compressor'].set_attr(igva='var')
+    >>> imported_nwk.components['compressor'].set_attr(igva='var')
     >>> imported_nwk.solve('offdesign', design_path='exported_nwk',
     ... init_path='exported_nwk')
-    >>> round(imported_nwk.imp_comps['turbine'].eta_s.val, 3)
+    >>> round(imported_nwk.components['turbine'].eta_s.val, 3)
     0.9
     >>> imported_nwk.busses['total power output'].set_attr(P=-0.75e6)
     >>> imported_nwk.solve('offdesign', design_path='exported_nwk',
     ... init_path='exported_nwk')
-    >>> round(imported_nwk.imp_comps['turbine'].eta_s.val, 3) == eta_s_t
+    >>> round(imported_nwk.components['turbine'].eta_s.val, 3) == eta_s_t
     True
-    >>> round(imported_nwk.imp_comps['compressor'].igva.val, 3) == igva
+    >>> round(imported_nwk.components['compressor'].igva.val, 3) == igva
     True
     >>> shutil.rmtree('./exported_nwk', ignore_errors=True)
     """
[…]
