### 2020-04-06 bd9a483
Update example according to new API
