### 2020-04-06 f7ad66c
Update network reader documentation
