+++ doc/tespy_modules/networks.rst
@@ -605,10 +605,11 @@ created network.
 
 .. note::
 
-    - Imported connections are accessible by the connections' source and source
-      id as well as target and target id following this principle:
-      :code:`imported_plant.imp_conns['source-label:source-id_target-label:target-id']`,
-      e. g. :code:`imported_plant.imp_conns['condenser:out1_condensate pump:in1']`.
-    - Imported components and busses are accessible by their label, e. g.
-      :code:`imported_plant.imp_comps['condenser']` and
-      :code:`imported_plant.imp_busses['total heat output']` respectively.
+    Imported busses, components and connections are accessible by their label,
+    e. g. :code:`imported_plant.busses['total heat output']`,
+    :code:`imported_plant.components['condenser']` and
+    :code:`imported_plant.connections['myconnectionlabel']` respectively. If
+    you did not provide labels for your connections, by default, the
+    connection's label will be according to this principle:
+    :code:`'source-label:source-id_target-label:target-id'`, where source and
+    target are the labels of the connected components.
