+++ doc/tespy_modules/networks.rst
@@ -22,7 +22,7 @@ The solver will solve for these variables. As stated in the introduction the
 list of fluids is passed to your network on creation. If your **system**
 **includes fluid mixtures**, you should **always make use of the value ranges**
 for the system variables. This improves the stability of the algorithm. Try to
-fit the boundaries as tight as possible, for instance, if you kwow that the
+fit the boundaries as tight as possible, for instance, if you know that the
 maximum pressure in the system will be at 10 bar, use it as upper boundary.
 
 .. note::
@@ -135,7 +135,7 @@ choose :code:`offdesign` as calculation mode the specification of a
 The usage of an initialisation path is always optional but highly recommended,
 as the convergence of the solution process will be improved, if you provide
 good starting values. If you do not specify an :code:`init_path`, the
-initialisation from priorly saved results will be skipped.
+initialisation from saved results will be skipped.
 :code:`init_only=True` usually is used for debugging. Or, you could use this
 feature to export a not solved network, if you want to do the parametrisation
 in .csv-files rather than your python script.
@@ -157,7 +157,7 @@ Specifying a value for the efficiency is of course possible.
 
 Offdesign mode
 ++++++++++++++
-The offdesign mode is used to **calulate the performance of your plant, if**
+The offdesign mode is used to **calculate the performance of your plant, if**
 **parameters deviate from the plant's design point**. This can be partload
 operation, operation at different temperature or pressure levels etc.. Thus,
 before starting an offdesing calculation you have to design your plant first.
@@ -217,7 +217,7 @@ variables n** is :math:`n = num_{conn} \cdot (3 + num_{fluids})`.
 The algorithm requires starting values for all variables of the system, thus an
 initialisation of the system is run prior to calculating the solution. **High**
 **quality initial values are crutial for convergence speed and stability**, bad
-starting values might lead to instabilty and diverging calculation can be the
+starting values might lead to instability and diverging calculation can be the
 result. There are different levels for the initialisation.
 
 Initialisation
@@ -241,7 +241,7 @@ The initialisation is performed in the following steps.
    argument).
 
 The network check is used to find errors in the network topology, the
-calulation can not start without a successful check. For components, a
+calculation can not start without a successful check. For components, a
 preprocessing of some parameters is necessary. It is performed by the
 :code:`comp_init` method of the components. You will find the methods in the
 :py:mod:`components module <tespy.components>`. The design/offdesign switch is
@@ -288,9 +288,9 @@ file.
     The files do not need to contain all connections of your network. You can
     build your network step by step and initialise the existing parts of your
     network from the :code:`init_path`. Be aware that a change within the fluid
-    vector does not alow this practice! If you plan to use additional fluids in
-    parts of the network you have not touched until now, you will need to state
-    all fluids from the beginning.
+    vector does not allow this practice! If you plan to use additional fluids
+    in parts of the network you have not touched until now, you will need to
+    state all fluids from the beginning.
 
 
 Algorithm
@@ -301,7 +301,7 @@ implemented.
 Newton–Raphson method
 +++++++++++++++++++++
 The Newton–Raphson method requires the calculation of residual values for the
-equations and of the partial derivatives to all system variables (jacobian
+equations and of the partial derivatives to all system variables (Jacobian
 matrix). In the next step the matrix is inverted and multiplied with the
 residual vector to calculate the increment for the system variables. This
 process is repeated until every equation's result in the system is "correct",
@@ -383,7 +383,7 @@ applied:
 
  * Cut off fluid mass fractions smaller than 0 and larger than 1. This way a
    mass fraction of a single fluid components never exceeds these boundaries.
- * Check, wheather the fluid properties of pure fluids are within the available
[…]
