+++ doc/tespy_modules/other.rst
@@ -1,7 +1,14 @@
 How can TESPy contribute to your energy system calculations?
 ============================================================
 
-In this part you learn how you can use TESPy for your energy system calculations: In energy system calculations, for instance in oemof-solph, plants are usually modelled as abstract components on a much lower level of detail.
-In order to represent a plant within an abstract component it is possible to supply characteristics establishing a connection between your energy system model and a specific plant model.
-Thus the characteristics are a representation of a specific plant layout in terms of topology and process parameters. In the examples section we have an example of a heat pump COP at different loads and ambient temperatures
-as well as a CHP unit with backpressure turbine operating at different loads and varying feed flow temperatures of a heating system.
+In this part you learn how you can use TESPy for your energy system
+calculations: In energy system calculations, for instance in oemof-solph,
+plants are usually modelled as abstract components on a much lower level of
+detail. In order to represent a plant within an abstract component it is
+possible to supply characteristics establishing a connection between your
+energy system model and a specific plant model. Thus the characteristics are a
+representation of a specific plant layout in terms of topology and process
+parameters. In the examples section we have an example of a heat pump COP at
+different loads and ambient temperatures as well as a CHP unit with
+backpressure turbine operating at different loads and varying feed flow
+temperatures of a heating system.
