### 2020-04-07 ccdcbfc
Add description of connection access through label
