+++ doc/tespy_modules/connections.rst
@@ -92,8 +92,31 @@ If you want to specify the fluid vector you can do it in the following way:
     # unset values
     myconn.fluid.set_attr(val_set={'water': False, 'air': False})
 
-References can not be used for fluid composition at the moment!
+.. note::
 
+    References can not be used for fluid composition at the moment!
+
+You may want to access the network's connections other than using the variable
+names, for example in an imported network or connections from a subsystem. It
+is possible to access these using the connection's label. By default, the label
+is generated by this logic:
+
+:code:`source:source_id_target:target_id`, where
+
+- :code:`source` and :code:`target` are the  labels of the components that are
+  connected.
+- :code:`source_id` and :code:`target_id` are e. g. :code:`out1` and
+  :code:`in2` respectively.
+
+.. code-block:: python
+
+    mynetwork.connections['myconnectionlabel'].set_attr(p=1e5)
+
+.. note::
+
+    Once added to the network, changing the label of a connection will not
+    affect the network. Thus, you should label your connections before adding
+    them to the network!
 
 .. _tespy_busses_label:
 
