### 2020-04-07 1abdd57
Add check for duplicate connection labels
