### 2020-04-07 9709f56
Add duplicate connection label error test
