### 2020-04-07 d96b290
Replace single quotes with double quotes in logging messages
