+++ tespy/components/components.py
@@ -89,7 +89,7 @@ class component:
             raise ValueError(msg)
 
         elif len([x for x in [';', ',', '.'] if x in label]) > 0:
-            msg = ('Can\'t use ' + str([';', ',', '.']) + ' in label (' +
+            msg = ('You must not use ' + str([';', ',', '.']) + ' in label (' +
                    str(self.component()) + ').')
             logging.error(msg)
             raise ValueError(msg)
