### 2020-04-07 205a2f4
Rename to value_range in memorisation
