### 2020-04-07 a027f81
Update stoichiometric combustion chamber according to CoolProp backend changes
