+++ tespy/components/combustion.py
@@ -400,9 +400,8 @@ class combustion_chamber(component):
         if self.ti.is_set:
             # stoichiometric combustion chamber
             if isinstance(self, combustion_chamber_stoich):
-                pos = 3 + self.nw_fluids.index(
-                    'TESPy::' + self.fuel_alias.val)
-                fuel = 'TESPy::' + self.fuel_alias.val
+                fuel = self.fuel_alias.val
+                pos = 3 + self.nw_fluids.index(fuel)
 
                 for i in range(2):
                     self.mat_deriv[k, i, 0] = -self.inl[i].fluid.val[fuel]
@@ -1221,16 +1220,14 @@ class combustion_chamber_stoich(combustion_chamber):
         Fuel composition, e. g. :code:`{'CH4': 0.96, 'CO2': 0.04}`.
 
     fuel_alias : str
-        Alias for the fuel, name of fuel for usage in network will be
-        TESPy::fuel_alias.
+        Alias for the fuel.
 
     air : dict
         Fresh air composition,
         e. g. :code:`{'N2': 0.76, 'O2': 0.23, 'Ar': 0.01}`.
 
     air_alias : str
-        Alias for the fresh air, name of air for usage in network will be
-        TESPy::air_alias.
+        Alias for the fresh air.
 
     path : str
         Path to existing fluid property table.
@@ -1255,14 +1252,14 @@ class combustion_chamber_stoich(combustion_chamber):
 
     If you choose 'Air' or 'air' as alias for the fresh air, TESPy will use
     the fluid properties from CoolProp's air. Else, a custom fluid
-    'TESPy::yourairalias' will be created.
+    'yourairalias' will be created.
 
-    The name of the flue gas will be: 'TESPy::yourfuelalias_fg'. It is also
+    The name of the flue gas will be: 'yourfuelalias_fg'. It is also
     possible to use fluid mixtures for the fuel, e. g.
     :code:`fuel={CH4: 0.9, 'CO2': 0.1}`. If you specify a fluid mixture for
-    the fuel, TESPy will automatically create a custom fluid called:
-    'TESPy::yourfuelalias'. For more information see the examples section
-    or look for the combustion chamber tutorials at tespy.readthedocs.io.
+    the fuel, TESPy will automatically create a custom fluid called. For more
+    information see the examples section or look for the combustion chamber
+    tutorials at tespy.readthedocs.io.
 
     Example
     -------
@@ -1281,7 +1278,7 @@ class combustion_chamber_stoich(combustion_chamber):
     >>> from tespy.tools.fluid_properties import T_bp_p
     >>> import numpy as np
     >>> import shutil
-    >>> fluid_list = ['TESPy::myAir', 'TESPy::myFuel', 'TESPy::myFuel_fg']
+    >>> fluid_list = ['myAir', 'myFuel', 'myFuel_fg']
     >>> nw = network(fluids=fluid_list, p_unit='bar', T_unit='C',
     ... p_range=[0.001, 10], T_range=[10, 2000], iterinfo=False)
     >>> amb = source('ambient air')
@@ -1304,10 +1301,10 @@ class combustion_chamber_stoich(combustion_chamber):
     >>> comb.set_attr(fuel={'CH4': 0.96, 'CO2': 0.03, 'H2': 0.01},
     ... air={'Ar': 0.0129, 'N2': 0.7553, 'H2O': 0, 'CH4': 0, 'CO2': 0.0004,
     ... 'O2': 0.2314}, fuel_alias='myFuel', air_alias='myAir', ti=500000)
-    >>> amb_comb.set_attr(T=20, p=1, fluid={'TESPy::myAir': 1,
-    ... 'TESPy::myFuel': 0,'TESPy::myFuel_fg': 0})
[…]
