### 2020-04-07 9e18e5f
Implement backend selection for fluids
