+++ tespy/tools/fluid_properties.py
@@ -31,6 +31,7 @@ import numpy as np
 from scipy import interpolate
 import pandas as pd
 import os
+from collections import OrderedDict
 
 import logging
 
@@ -63,9 +64,6 @@ class tespy_fluid:
     path : str
         Path to importing tespy fluid from.
 
-    plot : boolean
-        Plot the lookup tables after creation?
-
     Note
     ----
     Creates lookup tables for
@@ -89,9 +87,7 @@ class tespy_fluid:
     >>> from tespy.components.basics import sink, source
     >>> from tespy.tools.fluid_properties import (tespy_fluid, h_mix_pT,
     ... s_mix_pT, v_mix_pT, visc_mix_pT)
-    >>> from tespy.tools.global_vars import molar_masses
     >>> from tespy.networks.networks import network
-    >>> from CoolProp.CoolProp import PropsSI as CP
     >>> import shutil
     >>> fluidvec = {'N2': 0.7552, 'O2': 0.2314, 'CO2': 0.0005, 'Ar': 0.0129}
     >>> p_arr = np.array([0.1, 10]) * 1e5
@@ -106,9 +102,7 @@ class tespy_fluid:
     viscosity. Specific volume and viscosity are absolute values, thus no
     difference is calculated.
 
-    >>> molar_masses['air'] = CP('M', 'air')
-    >>> gas_constants['air'] = CP('GAS_CONSTANT', 'air')
-    >>> memorise.add_fluids(['air'])
+    >>> memorise.add_fluids({'air': 'HEOS'})
 
     >>> type(myfluid)
     <class 'tespy.tools.fluid_properties.tespy_fluid'>
@@ -149,23 +143,21 @@ class tespy_fluid:
     >>> shutil.rmtree('./LUT', ignore_errors=True)
     """
 
-    def __init__(self, alias, fluid, p_range, T_range, path=None, plot=False):
+    def __init__(self, alias, fluid, p_range, T_range, path=None):
 
         if not isinstance(alias, str):
             msg = 'Alias must be of type String.'
             logging.error(msg)
             raise TypeError(msg)
 
-        if 'IDGAS::' in alias:
-            msg = 'You are not allowed to use "IDGAS::" within your alias.'
+        # process parameters
+        if '::' in alias:
+            msg = 'The sequence "::" is not allowed in your fluid alias.'
             logging.error(msg)
             raise ValueError(msg)
-
-        # process parameters
-        if 'TESPy::' in alias:
-            self.alias = alias
         else:
-            self.alias = 'TESPy::' + alias
+            self.alias = alias
+
         self.fluid = fluid
 
         # adjust value ranges according to specified unit system
@@ -176,24 +168,29 @@ class tespy_fluid:
         self.p = np.linspace(self.p_range[0], self.p_range[1])
         self.T = np.linspace(self.T_range[0], self.T_range[1])
 
-        # plotting
-        self.plot = plot
-
         # path for loading
         self.path = path
 
-        # calculate molar mass and gas constant
-        for f in self.fluid:
-            molar_masses[f] = CPPSI('M', f)
-            gas_constants[f] = CPPSI('GAS_CONSTANT', f)
-
-        molar_masses[self.alias] = 1 / molar_mass_flow(self.fluid)
-        gas_constants[self.alias] = (gas_constants['uni'] /
-                                     molar_masses[self.alias])
-
         # create look up tables
-        tespy_fluid.fluids[self.alias] = {}
-        memorise.add_fluids(self.fluid.keys())
+        tespy_fluid.fluids[self.alias] = self
+
+        self.fluids_backends = OrderedDict()
+
+        for f in sorted(list(self.fluid.keys()) + [self.alias]):
+            try:
+                data = f.split('::')
+                backend = data[0]
[…]
