### 2020-04-07 bb9055a
Update tests according to changes
