+++ tests/test_components/test_turbomachinery.py
@@ -48,7 +48,7 @@ class TestTurbomachinery:
         self.setup_network(instance)
 
         # compress NH3, other fluids in network are for turbine, pump, ...
-        fl = {'N2': 1, 'O2': 0, 'Ar': 0, 'INCOMP::DowQ': 0, 'NH3': 0}
+        fl = {'N2': 1, 'O2': 0, 'Ar': 0, 'DowQ': 0, 'NH3': 0}
         self.c1.set_attr(fluid=fl, v=1, p=1, T=5)
         self.c2.set_attr(p=6)
         instance.set_attr(eta_s=0.8)
@@ -156,7 +156,7 @@ class TestTurbomachinery:
         """Test component properties of pumps."""
         instance = pump('pump')
         self.setup_network(instance)
-        fl = {'N2': 0, 'O2': 0, 'Ar': 0, 'INCOMP::DowQ': 1, 'NH3': 0}
+        fl = {'N2': 0, 'O2': 0, 'Ar': 0, 'DowQ': 1, 'NH3': 0}
         self.c1.set_attr(fluid=fl, v=1, p=5, T=50)
         self.c2.set_attr(p=7)
         instance.set_attr(eta_s=1)
@@ -246,7 +246,7 @@ class TestTurbomachinery:
         """Test component properties of turbines."""
         instance = turbine('turbine')
         self.setup_network(instance)
-        fl = {'N2': 0.7556, 'O2': 0.2315, 'Ar': 0.0129, 'INCOMP::DowQ': 0,
+        fl = {'N2': 0.7556, 'O2': 0.2315, 'Ar': 0.0129, 'DowQ': 0,
               'NH3': 0}
         self.c1.set_attr(fluid=fl, m=15, p=10)
         self.c2.set_attr(p=1, T=20)
@@ -344,7 +344,7 @@ class TestTurbomachinery:
                instance.component() + '.')
         assert 'turbomachine' == instance.component(), msg
         self.setup_network(instance)
-        fl = {'N2': 0.7556, 'O2': 0.2315, 'Ar': 0.0129, 'INCOMP::DowQ': 0,
+        fl = {'N2': 0.7556, 'O2': 0.2315, 'Ar': 0.0129, 'DowQ': 0,
               'H2O': 0, 'NH3': 0, 'CO2': 0, 'CH4': 0}
         self.c1.set_attr(fluid=fl, m=10, p=1, h=1e5)
         self.c2.set_attr(p=3, h=2e5)
