### 2020-04-07 1947ddc
Fix SyntaxError in Q_ph function
