+++ doc/tespy_modules/fluid_properties.rst
@@ -1,15 +1,30 @@
-.. _tespy_fluid_properties_label:
-
 Fluid properties
 ================
-
 The basic fluid properties are handled by
 `CoolProp <http://www.coolprop.org/>`_. All available fluids can be found on
 their homepage. Also see :cite:`Bell2014`.
 
+CoolProp back-ends
+------------------
+CoolProp provides multiple back-ends for fluid property calculation. The
+back-ends vary in calculation speed and calculation accuracy. It is possible
+to choose from the following back-ends:
+
+- :code:`HEOS`: Helmhotz Equation Of State with highest accuracy and lowest
+  calculation speed. **This is the default back-end!**
+- :code:`BICUBIC`: Tabular back-end with high accuracy and very high
+  calculation speed.
+- :code:`TTSE`: Tabular back-end with lowest accuracy and very high calculation
+  speed.
+- :code:`INCOMP`: Back-end for incompressible fluids.
+- :code:`IF97`: Back-end for the IAPWS-IF97 of water, very accurate and much
+  higher calculation speed than :code:`HEOS`.
+
+For more information on the back-ends please visit the CoolProp online
+documentation.
+
 Pure and pseudo-pure fluids
 ---------------------------
-
 If you use pure fluids, TESPy directly uses CoolProp functions to gather all
 fluid properties. CoolProp covers the most important fluids such as water, air
 as a pseudo-pure fluid as well as its components, several fuels and
@@ -20,7 +35,6 @@ two-phase region.
 
 Incompressible fluids
 ---------------------
-
 If you are looking for heat transfer fluids, the list of incompressible
 `fluids <http://www.coolprop.org/fluid_properties/Incompressibles.html>`_
 might be interesting for you. In contrast to the pure fluids, the properties
@@ -28,13 +42,11 @@ cover liquid state only.
 
 Fluid mixtures
 --------------
-
 CoolProp provides fluid properties for two component mixtures. BUT: These are
-NOT integrated in TESPy! Nevertheless, you can use fluid mixtures for gases:
+NOT integrated in TESPy! Nevertheless, you can use fluid mixtures for gases.
 
 Ideal mixtures of gaseous fluids
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
 TESPy can handle mixtures of gaseous fluids, by using the single fluid
 properties from CoolProp together with corresponding equations for mixtures.
 The equations can be found in the
@@ -49,7 +61,6 @@ for more information.
 
 Other mixtures
 ^^^^^^^^^^^^^^
-
 It is **not possible** to use mixtures of liquid and other liquid or gaseous
 fluids **at the moment**! If you try to use a mixture of two liquid or gaseous
 fluids and liquid fluids, e. g. water and methanol or liquid water and air, the
