### 2020-04-07 f0aa7d0
Update combustion_chamber_stoich tutorial
