### 2020-04-07 c9e69f6
Fix s_mix_pT function and remove some residuals
