### 2020-04-08 3d99627
Update some logging messages
