### 2020-04-08 5f82f6c
Improve variable naming in many modules
