+++ tespy/networks/networks.py
@@ -13,39 +13,27 @@ available from its original location tespy/networks/networks.py
 
 SPDX-License-Identifier: MIT
 """
-# reading .csv
 import ast
-# ordered dicts for fluid composition vector
 from collections import Counter, OrderedDict
-# calculation of molar masses and gas constants
 from CoolProp.CoolProp import PropsSI as CPPSI
-# logging messages
 import logging
-# numpy functions
 import numpy as np
 from numpy.linalg import inv
 from numpy.linalg import norm
-# checking/creating folders
 import os
-# DataFrames for connections and components
 import pandas as pd
-# printing results
 from tabulate import tabulate
-
-from tespy import connections as con
-
-from tespy.components.basics import (sink, source, subsystem_interface,
-                                     cycle_closer)
+from tespy.components.basics import (
+    sink, source, subsystem_interface, cycle_closer)
 from tespy.components.combustion import combustion_chamber, combustion_engine
 from tespy.components.customs import orc_evaporator
 from tespy.components.heat_exchangers import heat_exchanger
 from tespy.components.nodes import drum, merge, splitter
 from tespy.components.reactors import water_electrolyzer
-
+from tespy import connections as con
 from tespy.tools import data_containers as dc
 from tespy.tools import fluid_properties as fp
 from tespy.tools import helpers as hlp
-
 from tespy.tools.global_vars import molar_masses, gas_constants, err
 # calculation time
 from time import time
@@ -1679,7 +1667,7 @@ class network:
         if not self.progress:
             msg = ('The solver does not seem to make any progress, aborting '
                    'calculation. Residual value is '
-                   '{:.2e}'.format(norm(self.vec_res)) + '. This frequently '
+                   '{:.2e}'.format(norm(self.residual)) + '. This frequently '
                    'happens, if the solver pushes the fluid properties out of '
                    'their feasible range.')
             logging.warning(msg)
@@ -1719,9 +1707,9 @@ class network:
 
         for self.iter in range(self.max_iter):
 
-            self.vec_z_filter = np.absolute(self.vec_z) < err ** 2
+            self.increment_filter = np.absolute(self.increment) < err ** 2
             self.solve_control()
-            self.res = np.append(self.res, norm(self.vec_res))
+            self.res = np.append(self.res, norm(self.residual))
 
             if self.iterinfo:
                 self.print_iterinfo('solving')
@@ -1743,7 +1731,7 @@ class network:
         if self.iter == self.max_iter - 1:
             msg = ('Reached maximum iteration count (' + str(self.max_iter) +
                    '), calculation stopped. Residual value is '
-                   '{:.2e}'.format(norm(self.vec_res)))
+                   '{:.2e}'.format(norm(self.residual)))
             logging.warning(msg)
 
     def solve_determination(self):
@@ -1786,9 +1774,9 @@ class network:
         self.num_vars = (self.num_conn_vars * len(self.conns.index) +
                          self.num_comp_vars)
 
-        self.vec_res = np.zeros([self.num_vars])
-        self.vec_z = np.ones([self.num_vars])
-        self.mat_deriv = np.zeros((self.num_vars, self.num_vars))
+        self.residual = np.zeros([self.num_vars])
+        self.increment = np.ones([self.num_vars])
+        self.jacobian = np.zeros((self.num_vars, self.num_vars))
 
         msg = 'Total number of variables: ' + str(self.num_vars) + '.'
         logging.debug(msg)
@@ -1832,10 +1820,10 @@ class network:
[…]
