### 2020-04-08 0ec4059
Fix PEP8 issues
