+++ doc/whats_new/v0-3-0.rst
@@ -39,8 +39,9 @@ Bug fixes
 
 Other changes
 #############
-- Improve naming in networks and connections modules. A connection's target is
-  now labeled :code:`target` instead of :code:`t`, for instance.
+- Improve naming in networks, components and connections modules
+  (`PR #172 <https://github.com/oemof/tespy/pull/172>`_,
+   `PR #175 <https://github.com/oemof/tespy/pull/175>`_).
 
 Changes in the API
 ##################
