### 2020-04-08 51a685d
Remodel *_mix_ph and *_mix_ps functions for improved efficiency
