+++ tespy/tools/fluid_properties.py
@@ -176,21 +176,7 @@ class tespy_fluid:
 
         self.fluids_backends = OrderedDict()
 
-        for f in sorted(list(self.fluid.keys()) + [self.alias]):
-            try:
-                data = f.split('::')
-                backend = data[0]
-                fluid = data[1]
-            except IndexError:
-                backend = 'HEOS'
-                fluid = f
-
-            if f != self.alias:
-                self.fluids_backends[f] = backend
-            else:
-                self.fluids_backends[f] = 'TESPy'
-
-        memorise.add_fluids(self.fluids_backends)
+        memorise.add_fluids({self.alias: 'TESPy'})
 
         params = {}
 
@@ -544,39 +530,36 @@ def T_mix_ph(flow, T0=300):
     """
     # check if fluid properties have been calculated before
     fl = tuple(flow[3].keys())
-    a = memorise.T_ph[fl][:, 0:-1]
-    b = np.array([flow[1], flow[2]] + list(flow[3].values()))
-    ix = np.where(np.all(abs(a - b) <= err, axis=1))[0]
+    memorisation = fl in memorise.T_ph.keys()
+    if memorisation is True:
+        a = memorise.T_ph[fl][:, :-1]
+        b = np.array([flow[1], flow[2]] + list(flow[3].values()))
+        ix = np.where(np.all(abs(a - b) <= err, axis=1))[0]
 
-    if ix.size == 1:
-        # known fluid properties
-        T = memorise.T_ph[fl][ix, -1][0]
-        memorise.T_ph_f[fl] += [T]
-        return T
+        if ix.size == 1:
+            # known fluid properties
+            T = memorise.T_ph[fl][ix, -1][0]
+            memorise.T_ph_f[fl] += [T]
+            return T
+
+    # unknown fluid properties
+    fluid = single_fluid(flow[3])
+    if fluid is None:
+        # calculate the fluid properties for fluid mixtures
+        if T0 < 70:
+            T0 = 300
+        val = newton(h_mix_pT, dh_mix_pdT, flow, flow[2], val0=T0,
+                     valmin=70, valmax=3000, imax=10)
     else:
-        # unknown fluid properties
-        if num_fluids(flow[3]) > 1:
-            # calculate the fluid properties for fluid mixtures
-            if T0 < 70:
-                T0 = 300
-            val = newton(h_mix_pT, dh_mix_pdT, flow, flow[2], val0=T0,
-                         valmin=70, valmax=3000, imax=10)
-            new = np.array([[flow[1], flow[2]] +
-                            list(flow[3].values()) + [val]])
-            # memorise the newly calculated value
-            memorise.T_ph[fl] = np.append(memorise.T_ph[fl], new, axis=0)
-            return val
-        else:
-            # calculate fluid property for pure fluids
-            for fluid, x in flow[3].items():
-                if x > err:
-                    val = T_ph(flow[1], flow[2], fluid)
-                    new = np.array([[flow[1], flow[2]] +
-                                    list(flow[3].values()) + [val]])
-                    # memorise the newly calculated value
-                    memorise.T_ph[fl] = np.append(
-                            memorise.T_ph[fl], new, axis=0)
-                    return val
+        # calculate fluid property for pure fluids
+        val = T_ph(flow[1], flow[2], fluid)
+
+    if memorisation is True:
+        # memorise the newly calculated value
+        new = np.asarray([[flow[1], flow[2]] + list(flow[3].values()) + [val]])
+        memorise.T_ph[fl] = np.append(memorise.T_ph[fl], new, axis=0)
+
+    return val
 
 
 def T_ph(p, h, fluid):
@@ -693,8 +676,8 @@ def dT_mix_ph_dfluid(flow, T0=300):
[…]
