### 2020-04-08 bacd161
Let single_fluid() return None in case of mixtures
