### 2020-04-08 fefae9c
Add fluid components of a tespy fluid to the memorisation class
