### 2020-04-08 fda153f
Move heat_pump_model test
