+++ tests/test_models/test_heat_pump_model.py
@@ -5,7 +5,7 @@
 This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
 by the contributors recorded in the version control history of the file,
 available from its original location
-tests/test_heat_pump_model.py
+tests/test_models/test_heat_pump_model.py
 
 SPDX-License-Identifier: MIT
 """
@@ -29,15 +29,12 @@ import logging
 class TestHeatPump:
 
     def setup(self):
-        # %% network
-
+        # %% network setup
         self.nw = network(fluids=['water', 'NH3'], T_unit='C', p_unit='bar',
                           h_unit='kJ / kg', m_unit='kg / s')
 
         # %% components
-
         # sources & sinks
-
         c_in = source('coolant in')
         cb = source('consumer back flow')
         cf = sink('consumer feed flow')
@@ -49,13 +46,11 @@ class TestHeatPump:
         c_out = sink('coolant out')
 
         # consumer system
-
         cd = heat_exchanger('condenser')
         rp = pump('recirculation pump')
         cons = heat_exchanger_simple('consumer')
 
         # evaporator system
-
         va = valve('valve')
         dr = drum('drum')
         ev = heat_exchanger('evaporator')
@@ -63,13 +58,11 @@ class TestHeatPump:
         pu = pump('pump evaporator')
 
         # compressor-system
-
         cp1 = compressor('compressor 1')
         cp2 = compressor('compressor 2')
         he = heat_exchanger('intercooler')
 
         # busses
-
         self.power = bus('total compressor power')
         self.power.add_comps({'c': cp1}, {'c': cp2})
         self.heat = bus('total delivered heat')
@@ -77,9 +70,7 @@ class TestHeatPump:
         self.nw.add_busses(self.power, self.heat)
 
         # %% connections
-
         # consumer system
-
         c_in_cd = connection(c_in, 'out1', cd, 'in1')
 
         cb_rp = connection(cb, 'out1', rp, 'in1')
@@ -90,13 +81,11 @@ class TestHeatPump:
         self.nw.add_conns(c_in_cd, cb_rp, rp_cd, self.cd_cons, cons_cf)
 
         # connection condenser - evaporator system
-
         cd_va = connection(cd, 'out1', va, 'in1')
 
         self.nw.add_conns(cd_va)
 
         # evaporator system
-
         va_dr = connection(va, 'out1', dr, 'in1')
         dr_pu = connection(dr, 'out1', pu, 'in1')
         pu_ev = connection(pu, 'out1', ev, 'in2')
@@ -112,13 +101,11 @@ class TestHeatPump:
         self.nw.add_conns(self.amb_in_su, su_ev, ev_amb_out)
 
         # connection evaporator system - compressor system
-
         su_cp1 = connection(su, 'out2', cp1, 'in1')
 
         self.nw.add_conns(su_cp1)
 
         # compressor-system
-
         cp1_he = connection(cp1, 'out1', he, 'in1')
         he_cp2 = connection(he, 'out1', cp2, 'in1')
         cp2_c_out = connection(cp2, 'out1', c_out, 'in1')
@@ -129,41 +116,40 @@ class TestHeatPump:
         self.nw.add_conns(cp1_he, he_cp2, ic_in_he, he_ic_out, cp2_c_out)
 
[…]
