### 2020-04-08 7a89a86
Remove residuals from nose tests
