+++ tests/test_tools/test_fluid_properties.py
@@ -9,9 +9,6 @@ tests/test_tools/test_fluid_properties.py
 
 SPDX-License-Identifier: MIT
 """
-
-from nose.tools import eq_
-
 from tespy.tools import fluid_properties as fp
 from tespy.tools.global_vars import molar_masses, gas_constants
 from CoolProp.CoolProp import PropsSI as CP
@@ -84,7 +81,7 @@ class TestFluidProperties:
                                str(round(p, 0)) + ', T=' + str(round(T, 0)) +
                                ' for function ' + name + ', should be < ' +
                                str(d_rel_max) + '.')
-                        eq_(d_rel < d_rel_max, True, self.errormsg + msg)
+                        assert d_rel < d_rel_max, self.errormsg + msg
                     elif p < 5e6 and T < 500:
                         d_rel_max = 0.05
                         msg = ('Relative deviation is ' +
@@ -92,7 +89,7 @@ class TestFluidProperties:
                                str(round(p, 0)) + ', T=' + str(round(T, 0)) +
                                ' for function ' + name + ', should be < ' +
                                str(d_rel_max) + '.')
-                        eq_(d_rel < d_rel_max, True, self.errormsg + msg)
+                        assert d_rel < d_rel_max, self.errormsg + msg
                     elif p < 5e6 and T < 1000:
                         d_rel_max = 0.04
                         msg = ('Relative deviation is ' +
@@ -100,7 +97,7 @@ class TestFluidProperties:
                                str(round(p, 0)) + ', T=' + str(round(T, 0)) +
                                ' for function ' + name + ', should be < ' +
                                str(d_rel_max) + '.')
-                        eq_(d_rel < d_rel_max, True, self.errormsg + msg)
+                        assert d_rel < d_rel_max, self.errormsg + msg
                     elif p < 5e6 and T < 1500:
                         d_rel_max = 0.03
                         msg = ('Relative deviation is ' +
@@ -108,7 +105,7 @@ class TestFluidProperties:
                                str(round(p, 0)) + ', T=' + str(round(T, 0)) +
                                ' for function ' + name + ', should be < ' +
                                str(d_rel_max) + '.')
-                        eq_(d_rel < d_rel_max, True, self.errormsg + msg)
+                        assert d_rel < d_rel_max, self.errormsg + msg
                     elif T < 500:
                         d_rel_max = 0.1
                         msg = ('Relative deviation is ' +
@@ -116,7 +113,7 @@ class TestFluidProperties:
                                str(round(p, 0)) + ', T=' + str(round(T, 0)) +
                                ' for function ' + name + ', should be < ' +
                                str(d_rel_max) + '.')
-                        eq_(d_rel < d_rel_max, True, self.errormsg + msg)
+                        assert d_rel < d_rel_max, self.errormsg + msg
                     elif T < 1000:
                         d_rel_max = 0.075
                         msg = ('Relative deviation is ' +
@@ -124,7 +121,7 @@ class TestFluidProperties:
                                str(round(p, 0)) + ', T=' + str(round(T, 0)) +
                                ' for function ' + name + ', should be < ' +
                                str(d_rel_max) + '.')
-                        eq_(d_rel < d_rel_max, True, self.errormsg + msg)
+                        assert d_rel < d_rel_max, self.errormsg + msg
                     else:
                         d_rel_max = 0.025
                         msg = ('Relative deviation is ' +
@@ -132,7 +129,7 @@ class TestFluidProperties:
                                str(round(p, 0)) + ', T=' + str(round(T, 0)) +
                                ' for function ' + name + ', should be < ' +
[…]
