### 2020-04-08 e3f3062
Add gas turbine model test
