+++ tests/test_models/test_gas_turbine_models.py
@@ -1,0 +1,149 @@
+# -*- coding: utf-8 -*-
+
+"""Module for testing two tespy simulation against each other.
+
+This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
+by the contributors recorded in the version control history of the file,
+available from its original location
+tests/test_models/test_gasturbine_model.py
+
+SPDX-License-Identifier: MIT
+"""
+import logging
+import numpy as np
+import shutil
+from tespy.components.basics import sink, source
+from tespy.components.combustion import (
+    combustion_chamber, combustion_chamber_stoich)
+from tespy.components.turbomachinery import compressor, turbine
+from tespy.connections import connection
+from tespy.networks.networks import network
+
+
+class TestGasturbine:
+
+    def setup_combustion_chamber_model(self):
+        """Set up the model using the combustion chamber."""
+
+        # %% network setup
+        fluid_list = ['Ar', 'N2', 'O2', 'CO2', 'CH4', 'H2O']
+        self.nw1 = network(
+            fluids=fluid_list, p_unit='bar', T_unit='C',
+            p_range=[0.5, 20], T_range=[10, 2000])
+
+        # %% components
+        amb = source('ambient')
+        sf = source('fuel')
+        cc = combustion_chamber('combustion')
+        cp = compressor('compressor')
+        gt = turbine('turbine')
+        fg = sink('flue gas outlet')
+
+        # %% connections
+        amb_cp = connection(amb, 'out1', cp, 'in1')
+        cp_cc = connection(cp, 'out1', cc, 'in1')
+        sf_cc = connection(sf, 'out1', cc, 'in2')
+        cc_gt = connection(cc, 'out1', gt, 'in1', label='flue gas after cc')
+        gt_fg = connection(gt, 'out1', fg, 'in1', label='flue gas after gt')
+
+        self.nw1.add_conns(amb_cp, cp_cc, sf_cc, cc_gt, gt_fg)
+
+        # %% component parameters
+        cc.set_attr(lamb=3)
+        cp.set_attr(eta_s=0.9, pr=15)
+        gt.set_attr(eta_s=0.9)
+
+        # %% connection parameters
+        amb_cp.set_attr(
+            T=20, p=1, m=100,
+            fluid={'Ar': 0.0129, 'N2': 0.7553, 'H2O': 0, 'CH4': 0,
+                   'CO2': 0.0004, 'O2': 0.2314})
+        sf_cc.set_attr(
+            T=20, fluid={'CO2': 0.04, 'Ar': 0, 'N2': 0,
+                         'O2': 0, 'H2O': 0, 'CH4': 0.96})
+        gt_fg.set_attr(p=1)
+
+        # %% solving
+        mode = 'design'
+        self.nw1.solve(mode=mode)
+
+    def setup_combustion_chamber_stoich_model(self):
+        """Set up the model using the stoichimetric combustion chamber."""
+
+        # %% network setup
+        fluid_list = ['myAir', 'myFuel', 'myFuel_fg']
+        self.nw2 = network(
+            fluids=fluid_list, p_unit='bar', T_unit='C',
+            p_range=[0.5, 20], T_range=[10, 2000])
+
+        # %% components
+        amb = source('ambient')
+        sf = source('fuel')
+        cc = combustion_chamber_stoich('combustion')
+        cp = compressor('compressor')
+        gt = turbine('turbine')
+        fg = sink('flue gas outlet')
+
+        # %% connections
+        amb_cp = connection(amb, 'out1', cp, 'in1')
+        cp_cc = connection(cp, 'out1', cc, 'in1')
+        sf_cc = connection(sf, 'out1', cc, 'in2')
[…]
