### 2020-04-08 bec2d5e
Remove residuals after testing
