+++ tespy/components/combustion.py
@@ -1290,7 +1290,7 @@ class combustion_chamber_stoich(combustion_chamber):
     >>> comb_fg.set_attr(T=1200)
     >>> nw.solve('design')
     >>> round(comb.lamb.val, 3)
-    2.01
+    2.016
     >>> comb.set_attr(lamb=2)
     >>> comb_fg.set_attr(T=np.nan)
     >>> nw.solve('design')
