### 2020-04-08 e06c007
Fix combustion chamber stoich doctest
