### 2020-04-08 3fbde51
Fix PEP8 issues
