+++ tests/test_models/test_heat_pump_model.py
@@ -171,8 +171,8 @@ class TestHeatPump:
              1.175, 1.2125, 1.2375, 1.25])
         y = np.array(
             [0.0076, 0.1390, 0.2731, 0.4003, 0.5185, 0.6263, 0.7224, 0.8056,
-            0.8754, 0.9312, 0.9729, 1.0006, 1.0203, 1.0158, 1.0051, 1.0000,
-            0.9746, 0.9289, 0.8832, 0.8376, 0.7843, 0.7614])
+             0.8754, 0.9312, 0.9729, 1.0006, 1.0203, 1.0158, 1.0051, 1.0000,
+             0.9746, 0.9289, 0.8832, 0.8376, 0.7843, 0.7614])
         pu.set_attr(eta_s=0.8, design=['eta_s'], offdesign=['eta_s_char'],
                     eta_s_char=dc_cc(func=char_line(x, y), param='m'))
 
@@ -216,8 +216,8 @@ class TestHeatPump:
         x = np.linspace(0, 2.5, 26)
         y = np.array(
             [0.000, 0.164, 0.283, 0.389, 0.488, 0.581, 0.670, 0.756, 0.840,
-            0.921, 1.000, 1.078, 1.154, 1.228, 1.302, 1.374, 1.446, 1.516,
-            1.585, 1.654, 1.722, 1.789, 1.855, 1.921, 1.986, 2.051])
+             0.921, 1.000, 1.078, 1.154, 1.228, 1.302, 1.374, 1.446, 1.516,
+             1.585, 1.654, 1.722, 1.789, 1.855, 1.921, 1.986, 2.051])
         kA_char2 = dc_cc(func=char_line(x, y), param='m')
 
         cd.set_attr(kA_char1=kA_char1, kA_char2=kA_char2, pr2=0.9998,
