### 2020-04-10 f927155
Update h_os method
