### 2020-04-10 80e9109
Remove debug printing
