### 2020-04-10 26be708
Update generic derivative calculation
