### 2020-04-10 1ce1949
Fix value comparison by rounding
