### 2020-04-11 f507738
Remove rounding from test
