### 2020-04-11 b92a3e0
Try to fix travis issue, the tests run locally
