+++ tests/test_tools/test_fluid_properties.py
@@ -265,7 +265,7 @@ class TestFluidPropertyBackEnds:
         """Setup a pipeline network."""
 
     @pytest.mark.skipif(
-        'TRAVIS' in os.environ,
+        os.environ.get('TRAVIS') == 'true',
         reason='Travis CI cannot handle the tabular CoolProp back ends, '
         'skipping this test. The test should run on your local machine.')
     def test_clausius_rankine(self):
