### 2020-04-13 #2109 @corentin-dev
Dear Kyle,
I accept to review this submission.
Best regards,
Corentin

Le 13 avr. 2020 à 18:45, à 18:45, Kyle Niemeyer <notifications@github.com> a écrit:
>Hi @arosen93 @corentin-dev and @Jgoldfar, would any of you be
>interested in reviewing this submission for JOSS?
>
>I'm not sure if you've all reviewed for JOSS before (or recently), but
>we carry out our checklist-driven reviews here in GitHub issues and
>follow these guidelines:
>https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
>The package we would ask you to review is ["TESPy: Thermal Engineering
>Systems in Python"](https://github.com/oemof/tespy) and there is a
>[short associated
>paper](https://github.com/openjournals/joss-papers/blob/joss.02109/joss.02109/10.21105.joss.02109.pdf)
>to read too.
>
>Thanks!
>
>--
>You are receiving this because you were mentioned.
>Reply to this email directly or view it on GitHub:
>https://github.com/openjournals/joss-reviews/issues/2109#issuecomment-612982014
