### 2020-04-13 #2109 @fwitte
Looking forward to your review, @corentin-dev and @arosen93.
@kyleniemeyer, can you update the version number to 0.2.2? I plan to publish 0.3.0 shortly, so maybe even to 0.3.0 if it is ready within the next days?
Best regards
Francesco
