### 2020-04-13 9a164bd
Add a test for INCOMP backend
