### 2020-04-13 c4dfbd2
Add try for CoolProp IF97 back end error test
