### 2020-04-13 5596086
Tmin and Tmax are available for all fluids
