### 2020-04-13 35d9e7f
Compare to None instead of str
