### 2020-04-13 7ae2135
Add missing removal of stoichiometric combustion chamber lookup table
