### 2020-04-13 907a490
Add missing removal of stoichiometric combustion chamber lookup table
