### 2020-04-13 574a906
Add Python<=3.7 requirement due to CoolProp compability
