+++ setup.py
@@ -19,7 +19,7 @@ setup(name='TESPy',
       include_package_data=True,
       data_files=[('tespy/data', ['tespy/data/char_lines.json',
                                   'tespy/data/char_maps.json'])],
-      python_requires='>=3',
+      python_requires='>=3, <=3.7',
       install_requires=['CoolProp>=6,<7',
                         'numpy>=1.13.3,<2',
                         'pandas>=0.19.2,!=1.0.0,<2',
