### 2020-04-13 e318fdb
Update installation instructions for Python 3.8
