+++ doc/installation.rst
@@ -17,7 +17,7 @@ operation systems.
 
     Due to Python 3.8 incompability of the fluid property software (CoolProp)
     in version 6.3.0, we have implemented a Python version restriction to
-    Python 3.7 at maximum. If you build CoolProp manually, it is possible
+    Python lower than 3.8. If you build CoolProp manually, it is possible
     to use Python 3.8. For more information on the manual CoolProp installation
     refer to these instructions:
     http://www.coolprop.org/coolprop/wrappers/Python/index.html#python.
