### 2020-04-14 5261d92
Add None as possibilty to unset values
