+++ tespy/connections.py
@@ -232,10 +232,10 @@ class connection:
             raise ValueError(msg)
 
         if inlet_id not in comp2.inlets():
-            msg = ('Error creating connection. Specified inlet_id (' +
-                   inlet_id + ') is not valid for component ' +
-                   comp2.component() + '. Valid ids are: ' +
-                   str(comp2.inlets()) + '.')
+            msg = (
+                'Error creating connection. Specified inlet_id (' + inlet_id +
+                ') is not valid for component ' + comp2.component() +
+                '. Valid ids are: ' + str(comp2.inlets()) + '.')
             logging.error(msg)
             raise ValueError(msg)
 
@@ -258,11 +258,9 @@ class connection:
         self.printout = True
 
         # set default values for kwargs
-        var = self.attr()
-
-        for key in self.attr().keys():
-            self.__dict__.update({key: var[key]})
-
+        self.variables = self.attr()
+        self.variables0 = [x + '0' for x in self.variables.keys()]
+        self.__dict__.update(self.variables)
         self.set_attr(**kwargs)
 
         msg = (
@@ -362,26 +360,20 @@ class connection:
           adjust the enthalpy values of that connection for the first
           iterations in order to meet the state requirement.
         """
-        var = self.attr()
-        var0 = [x + '0' for x in var.keys()]
-
         # set specified values
         for key in kwargs:
-            if key in var.keys() or key in var0:
+            if key in self.variables.keys() or key in self.variables0:
+                # fluid specification
                 if 'fluid' in key:
-                    # fluid specification
                     if isinstance(kwargs[key], dict):
                         # starting values
-                        if key in var0:
-                            self.get_attr(key.replace('0', '')).set_attr(
-                                    val0=kwargs[key]
-                                    )
+                        if key in self.variables0:
+                            self.fluid.set_attr(val0=kwargs[key])
                         # specified parameters
                         else:
-                            self.get_attr(key).set_attr(val=kwargs[key].copy())
-                            for f in kwargs[key]:
-                                kwargs[key][f] = True
-                            self.get_attr(key).set_attr(val_set=kwargs[key])
+                            self.fluid.set_attr(val=kwargs[key])
+                            self.fluid.set_attr(
+                                val_set={f: True for f in kwargs[key].keys()})
 
                     elif isinstance(kwargs[key], dc_flu):
                         # data container for fluids
@@ -389,8 +381,9 @@ class connection:
 
                     else:
                         # bad datatype
-                        msg = ('Bad datatype for connection keyword ' + key +
-                               '.')
+                        msg = (
+                            'Datatype for fluid vector specification must be '
+                            'tespy.tools.data_containers.dc_flu or dict.')
                         logging.error(msg)
                         raise TypeError(msg)
 
@@ -399,53 +392,58 @@ class connection:
                         self.state.set_attr(val=kwargs[key], is_set=True)
                     elif isinstance(kwargs[key], dc_simple):
                         self.state = kwargs[key]
-                    else:
-                        if (isinstance(kwargs[key], float) or
-                                isinstance(kwargs[key], np.float64) or
-                                isinstance(kwargs[key], np.int64) or
-                                isinstance(kwargs[key], int)):
-                            if np.isnan(kwargs[key]):
-                                self.state.set_attr(
-                                        val=kwargs[key], is_set=False
-                                        )
-                            else:
-                                msg = ('Datatype for keyword argument ' +
-                                       key + ' must be str.')
-                                logging.error(msg)
-                                raise TypeError(msg)
+                    elif kwargs[key] is None:
+                        self.get_attr(key).set_attr(is_set=False)
+                    elif (isinstance(kwargs[key], float) or
+                          isinstance(kwargs[key], np.float64) or
+                          isinstance(kwargs[key], np.int64) or
[…]
