### 2020-04-14 6ff5d8a
Merge dc_simple and dc_cp handling
