### 2020-04-14 70e5a32
Replace more of the attr() calls with variables attribute
