### 2020-04-14 3a240d3
Fix some PEP issues
