### 2020-04-14 d60b66b
Add error test for NotImplemented referencing
