### 2020-04-14 d46f652
Add unsetting with None example
