### 2020-04-14 c016f46
Fix some typos
