+++ doc/tespy_modules/fluid_properties.rst
@@ -4,25 +4,25 @@ The basic fluid properties are handled by
 `CoolProp <http://www.coolprop.org/>`_. All available fluids can be found on
 their homepage. Also see :cite:`Bell2014`.
 
-CoolProp back-ends
+CoolProp back ends
 ------------------
-CoolProp provides multiple back-ends for fluid property calculation. The
-back-ends vary in calculation speed and calculation accuracy. It is possible
-to choose from the following back-ends:
+CoolProp provides multiple back ends for fluid property calculation. The
+back ends vary in calculation speed and calculation accuracy. It is possible
+to choose from the following back ends:
 
 - :code:`HEOS`: Helmhotz Equation Of State with highest accuracy and lowest
-  calculation speed. **This is the default back-end!**
-- :code:`BICUBIC`: Tabular back-end with high accuracy and very high
+  calculation speed. **This is the default back end!**
+- :code:`BICUBIC`: Tabular back end with high accuracy and very high
   calculation speed.
-- :code:`TTSE`: Tabular back-end with lowest accuracy and very high calculation
+- :code:`TTSE`: Tabular back end with lowest accuracy and very high calculation
   speed.
-- :code:`INCOMP`: Back-end for incompressible fluids.
-- ~~:code:`IF97`: Back-end for the IAPWS-IF97 of water, very accurate and much
+- :code:`INCOMP`: Back end for incompressible fluids.
+- ~~:code:`IF97`: Back end for the IAPWS-IF97 of water, very accurate and much
   higher calculation speed than :code:`HEOS`.~~ Due to a bug in the CoolProp
   back end this option is not available at the moment, for more information
   see this `github issue <https://github.com/CoolProp/CoolProp/issues/1918/>`_.
 
-For more information on the back-ends please visit the CoolProp online
+For more information on the Back ends please visit the CoolProp online
 documentation.
 
 Pure and pseudo-pure fluids
