### 2020-04-14 b4e5e1e
Fix doctest
