### 2020-04-14 93e3dfb
Test unsetting of dc_simple keyword
