### 2020-04-14 b7a02ca
Add bus evaluation methods
