### 2020-04-14 b511857
Modify bus function handling
