+++ tespy/components/heat_exchangers.py
@@ -636,17 +636,9 @@ class heat_exchanger_simple(component):
         """
         i = self.inl[0].to_flow()
         o = self.outl[0].to_flow()
-
         val = i[0] * (o[2] - i[2])
-        if np.isnan(bus.P_ref):
-            expr = 1
-        else:
-            expr = abs(val / bus.P_ref)
 
-        if calc_efficiency is True:
-            return bus.char.evaluate(expr)
-        else:
-            return val * bus.char.evaluate(expr)
+        return self.bus_func_handler(val, bus, calc_efficiency)
 
     def bus_deriv(self, bus):
         r"""
@@ -1747,17 +1739,9 @@ class heat_exchanger(component):
         """
         i = self.inl[0].to_flow()
         o = self.outl[0].to_flow()
-
         val = i[0] * (o[2] - i[2])
-        if np.isnan(bus.P_ref):
-            expr = 1
-        else:
-            expr = abs(val / bus.P_ref)
 
-        if calc_efficiency is True:
-            return bus.char.evaluate(expr)
-        else:
-            return val * bus.char.evaluate(expr)
+        return self.bus_func_handler(val, bus, calc_efficiency)
 
     def bus_deriv(self, bus):
         r"""
