### 2020-04-14 7b7ad84
Add generic bus function handler
