### 2020-04-14 8aea4ba
Add base field to bus class
