+++ tespy/connections.py
@@ -18,14 +18,16 @@ SPDX-License-Identifier: MIT
 
 import numpy as np
 import pandas as pd
-
 import logging
-
+from tespy.components.components import component
 from tespy.tools.characteristics import char_line
 from tespy.tools.data_containers import dc_cp, dc_flu, dc_prop, dc_simple
 from tespy.tools.helpers import TESPyConnectionError
+import warnings
 
-from tespy.components.components import component
+
+# pass the warning messages to the logger
+logging.captureWarnings(True)
 
 
 class connection:
@@ -724,12 +726,12 @@ class bus:
     >>> heat_bus.P.is_set
     False
 
-    >>> power_bus.add_comps({'c': chp, 'char': gen, 'p': 'P'},
-    ... {'c': pu, 'char': mot})
-    >>> heat_bus.add_comps({'c': chp, 'p': 'Q'},
-    ... {'c': fgc, 'char': -1})
-    >>> fuel_bus.add_comps({'c': chp, 'p': 'TI'},
-    ... {'c': pu, 'char': mot})
+    >>> power_bus.add_comps({'comp': chp, 'char': gen, 'param': 'P'},
+    ... {'comp': pu, 'char': mot})
+    >>> heat_bus.add_comps({'comp': chp, 'param': 'Q'},
+    ... {'comp': fgc, 'char': -1})
+    >>> fuel_bus.add_comps({'comp': chp, 'param': 'TI'},
+    ... {'comp': pu, 'char': mot})
     >>> nw.add_busses(power_bus, heat_bus, fuel_bus)
     >>> mode = 'design'
     >>> nw.solve(mode=mode)
@@ -769,7 +771,7 @@ class bus:
     def __init__(self, label, **kwargs):
 
         self.comps = pd.DataFrame(
-            columns=['param', 'P_ref', 'char', 'efficiency'])
+            columns=['param', 'P_ref', 'char', 'efficiency', 'base'])
 
         self.label = label
         self.P = dc_cp(val=np.nan, is_set=False)
@@ -860,17 +862,20 @@ class bus:
 
         Parameters
         ----------
-        c : dict
-            Dictionary containing the component information to be added to the
-            bus. These information are described in the notes!
+        *args : dict
+            Dictionaries containing the component information to be added to
+            the bus. The information are described below.
 
         Note
         ----
-        Keys for the dictionary c:
+        **Required Key**
 
-        - c (tespy.components.components.component): Component you want to add
-          to the bus.
-        - p (str): Bus parameter, optional.
+        - comp (tespy.components.components.component): Component you want to
+          add to the bus.
+
+        **Optional Keys**
+
+        - param (str): Bus parameter, optional.
 
             - You do not need to provide a parameter, if the component only has
               one option for the bus (turbomachines, heat exchangers,
@@ -891,45 +896,75 @@ class bus:
 
         - P_ref (float): Energy flow specification for reference case,
           :math:`P \text{/W}`, optional.
+        - base (str): Base value for characteristic line and efficiency
+          calculation. The base can either be :code:`'component'` (default) or
+          :code:`'bus'`.
+
+            - In case you choose :code:`'component'`, the characteristic line
+              input will follow the value of the component's bus function and
+              the efficiency definition is
+              :math:`\eta=\frac{P_\mathrm{bus}}{P_\mathrm{component}}`.
+            - In case you choose :code:`'bus'`, the characteristic line
+              input will follow the bus value of the component and the
+              efficiency definition is
+              :math:`\eta=\frac{P_\mathrm{component}}{P_\mathrm{bus}}`.
         """
         for c in args:
             if isinstance(c, dict):
-                if 'c' in c.keys():
-                    if isinstance(c['c'], component):
-                        self.comps.loc[c['c']] = [
[…]
