### 2020-04-14 911242f
Add base attribute to reader
