### 2020-04-14 8ae56ea
Adjust efficiency printout, add base export
