+++ tespy/networks/networks.py
@@ -834,7 +834,7 @@ class network:
         # unset design values for busses
         for b in self.busses.values():
             for cp in b.comps.index:
-                b.comps.loc[cp].P_ref = np.nan
+                b.comps.loc[cp, 'P_ref'] = np.nan
 
         series = pd.Series(dtype=np.float64)
         for cp in self.comps.index:
@@ -940,7 +940,7 @@ class network:
                 df.set_index('label', inplace=True)
                 # iter through all components of this type and set data
                 for c_label in df.index:
-                    comp = cp_sort.loc[c_label].comp
+                    comp = cp_sort.loc[c_label, 'comp']
                     # read data of components with individual design_path
                     if comp.design_path is not None:
                         path_c = hlp.modify_path_os(
@@ -1011,7 +1011,7 @@ class network:
         i = 0
         for b in data.busses:
             bus = self.busses[b].comps
-            bus.loc[component].P_ref = data.bus_P_ref[i]
+            bus.loc[component, 'P_ref'] = data.bus_P_ref[i]
             i += 1
 
     def init_conn_design_params(self, c, df):
@@ -1037,18 +1037,23 @@ class network:
         if len(conn.index) > 0:
             # read connection information
             conn_id = conn.index[0]
-            c.m.design = df.loc[conn_id].m * self.m[df.loc[conn_id].m_unit]
-            c.p.design = df.loc[conn_id].p * self.p[df.loc[conn_id].p_unit]
-            c.h.design = df.loc[conn_id].h * self.h[df.loc[conn_id].h_unit]
-            c.v.design = df.loc[conn_id].v * self.v[df.loc[conn_id].v_unit]
-            c.x.design = df.loc[conn_id].x * self.x[df.loc[conn_id].x_unit]
-            c.T.design = ((df.loc[conn_id]['T'] +
-                           self.T[df.loc[conn_id].T_unit][0]) *
-                          self.T[df.loc[conn_id].T_unit][1])
-            c.Td_bp.design = (df.loc[conn_id].Td_bp *
-                              self.T[df.loc[conn_id].T_unit][1])
+            c.m.design = (
+                df.loc[conn_id, 'm'] * self.m[df.loc[conn_id, 'm_unit']])
+            c.p.design = (
+                df.loc[conn_id, 'p'] * self.p[df.loc[conn_id, 'p_unit']])
+            c.h.design = (
+                df.loc[conn_id, 'h'] * self.h[df.loc[conn_id, 'h_unit']])
+            c.v.design = (
+                df.loc[conn_id, 'v'] * self.v[df.loc[conn_id, 'v_unit']])
+            c.x.design = (
+                df.loc[conn_id, 'x'] * self.x[df.loc[conn_id, 'x_unit']])
+            c.T.design = ((df.loc[conn_id, 'T'] +
+                           self.T[df.loc[conn_id, 'T_unit']][0]) *
+                          self.T[df.loc[conn_id, 'T_unit']][1])
+            c.Td_bp.design = (df.loc[conn_id, 'Td_bp'] *
+                              self.T[df.loc[conn_id, 'T_unit']][1])
             for fluid in self.fluids:
-                c.fluid.design[fluid] = df.loc[conn_id][fluid]
+                c.fluid.design[fluid] = df.loc[conn_id, fluid]
         else:
             # no matches in the connections of the network and the design files
             msg = (
@@ -1526,17 +1531,17 @@ class network:
                 # overwrite SI-values with values from init_file,
                 # except user specified values
                 if c.m.val_set is False:
-                    c.m.val_SI = (df.loc[conn_id].m *
-                                  self.m[df.loc[conn_id].m_unit])
[…]
