### 2020-04-14 e6ed80f
Add error test for new bus attribute base
