+++ tests/test_busses.py
@@ -1,0 +1,204 @@
+# -*- coding: utf-8
+
+"""Module for testing busses.
+
+This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
+by the contributors recorded in the version control history of the file,
+available from its original location
+tests/test_busses.py
+
+SPDX-License-Identifier: MIT
+"""
+import logging
+import numpy as np
+import shutil
+from tespy.components.basics import sink, source
+from tespy.components.combustion import combustion_chamber
+from tespy.components.turbomachinery import compressor, turbine
+from tespy.connections import connection, bus
+from tespy.networks.networks import network
+from tespy.tools.characteristics import char_line
+
+
+class TestBusses:
+
+    def setup(self):
+        """Set up the model."""
+        # %% network setup
+        fluid_list = ['Ar', 'N2', 'O2', 'CO2', 'CH4', 'H2O']
+        self.nw = network(
+            fluids=fluid_list, p_unit='bar', T_unit='C',
+            p_range=[0.5, 20], T_range=[10, 2000])
+
+        # %% components
+        amb = source('ambient')
+        sf = source('fuel')
+        cc = combustion_chamber('combustion')
+        cp = compressor('compressor')
+        gt = turbine('turbine')
+        fg = sink('flue gas outlet')
+
+        # %% connections
+        amb_cp = connection(amb, 'out1', cp, 'in1', label='ambient air flow')
+        cp_cc = connection(cp, 'out1', cc, 'in1')
+        sf_cc = connection(sf, 'out1', cc, 'in2')
+        cc_gt = connection(cc, 'out1', gt, 'in1')
+        gt_fg = connection(gt, 'out1', fg, 'in1')
+
+        self.nw.add_conns(amb_cp, cp_cc, sf_cc, cc_gt, gt_fg)
+
+        # %% component parameters
+        cc.set_attr(lamb=3)
+        cp.set_attr(eta_s=0.9, pr=15)
+        gt.set_attr(eta_s=0.9)
+
+        # %% connection parameters
+        amb_cp.set_attr(
+            T=20, p=1, m=100,
+            fluid={'Ar': 0.0129, 'N2': 0.7553, 'H2O': 0, 'CH4': 0,
+                   'CO2': 0.0004, 'O2': 0.2314})
+        sf_cc.set_attr(
+            T=20, fluid={'CO2': 0.04, 'Ar': 0, 'N2': 0,
+                         'O2': 0, 'H2O': 0, 'CH4': 0.96})
+        gt_fg.set_attr(p=1)
+
+        # motor efficiency
+        x = np.array(
+            [0, 0.05, 0.1, 0.15, 0.2, 0.25, 0.3, 0.35, 0.4, 0.45, 0.5, 0.55,
+             0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9, 0.95, 1, 1.05, 1.1, 1.15,
+             1.2, 10])
+        y = np.array(
+            [0.01, 0.3148, 0.5346, 0.6843, 0.7835, 0.8477, 0.8885, 0.9145,
+             0.9318, 0.9443, 0.9546, 0.9638, 0.9724, 0.9806, 0.9878, 0.9938,
+             0.9982, 0.999, 0.9995, 0.9999, 1, 0.9977, 0.9947, 0.9909, 0.9853,
+             0.9644]) * 0.975
+        self.motor_bus_based = char_line(x=x, y=y)
+        self.motor_comp_based = char_line(x=x, y=1 / y)
+
+        # generator efficiency
+        x = np.array(
[…]
