### 2020-04-15 77dd55b
Add bus efficiency calculation method and update docs
