### 2020-04-15 072a48b
Update code according to new bus calculation methods
