+++ tespy/components/customs.py
@@ -571,7 +571,7 @@ class orc_evaporator(component):
         o = self.outl[2].to_flow()
         val = i[0] * (o[2] - i[2])
 
-        return self.bus_func_handler(val, bus, calc_efficiency)
+        return val
 
     def bus_deriv(self, bus):
         r"""
