### 2020-04-15 e2da1d0
Fix bus error tests
