### 2020-04-15 94c57de
Update doctests
