### 2020-04-15 30424f6
Update bus_func documentation
