### 2020-04-15 eab8085
Rename bus_func_evaluation to calc_bus_value
