+++ tespy/components/turbomachinery.py
@@ -265,7 +265,7 @@ class turbomachine(component):
         -------
         val : float
             Value of energy transfer :math:`\dot{E}`. This value is passed to
-            :py:meth:`tespy.components.components.component.bus_func_evaluation`
+            :py:meth:`tespy.components.components.component.calc_bus_value`
             for value manipulation according to the specified characteristic
             line of the bus.
 
@@ -294,7 +294,7 @@ class turbomachine(component):
             Matrix of partial derivatives.
         """
         deriv = np.zeros((1, 2, self.num_nw_vars))
-        f = self.bus_func_evaluation
+        f = self.calc_bus_value
         deriv[0, 0, 0] = self.numeric_deriv(f, 'm', 0, bus=bus)
         deriv[0, 0, 2] = self.numeric_deriv(f, 'h', 0, bus=bus)
         deriv[0, 1, 2] = self.numeric_deriv(f, 'h', 1, bus=bus)
