+++ tespy/components/combustion.py
@@ -849,7 +849,7 @@ class combustion_chamber(component):
         -------
         val : float
             Value of energy transfer :math:`\dot{E}`. This value is passed to
-            :py:meth:`tespy.components.components.component.bus_func_evaluation`
+            :py:meth:`tespy.components.components.component.calc_bus_value`
             for value manipulation according to the specified characteristic
             line of the bus.
 
@@ -874,7 +874,7 @@ class combustion_chamber(component):
             Matrix of partial derivatives.
         """
         deriv = np.zeros((1, 3, self.num_nw_vars))
-        f = self.bus_func_evaluation
+        f = self.calc_bus_value
         for i in range(3):
             deriv[0, i, 0] = self.numeric_deriv(f, 'm', i, bus=bus)
             deriv[0, i, 3:] = self.numeric_deriv(f, 'fluid', i, bus=bus)
@@ -2757,7 +2757,7 @@ class combustion_engine(combustion_chamber):
         -------
         val : float
             Value of energy transfer :math:`\dot{E}`. This value is passed to
-            :py:meth:`tespy.components.components.component.bus_func_evaluation`
+            :py:meth:`tespy.components.components.component.calc_bus_value`
             for value manipulation according to the specified characteristic
             line of the bus.
 
@@ -2842,7 +2842,7 @@ class combustion_engine(combustion_chamber):
             Matrix of partial derivatives.
         """
         deriv = np.zeros((1, 7 + self.num_vars, self.num_nw_vars))
-        f = self.bus_func_evaluation
+        f = self.calc_bus_value
         b = bus.comps.loc[self]
 
         ######################################################################
