+++ tespy/components/customs.py
@@ -557,7 +557,7 @@ class orc_evaporator(component):
         -------
         val : float
             Value of energy transfer :math:`\dot{E}`. This value is passed to
-            :py:meth:`tespy.components.components.component.bus_func_evaluation`
+            :py:meth:`tespy.components.components.component.calc_bus_value`
             for value manipulation according to the specified characteristic
             line of the bus.
 
@@ -587,9 +587,10 @@ class orc_evaporator(component):
             Matrix of partial derivatives.
         """
         deriv = np.zeros((1, 6, self.num_nw_vars))
-        deriv[0, 2, 0] = self.numeric_deriv(self.bus_func, 'm', 2, bus=bus)
-        deriv[0, 2, 2] = self.numeric_deriv(self.bus_func, 'h', 2, bus=bus)
-        deriv[0, 5, 2] = self.numeric_deriv(self.bus_func, 'h', 5, bus=bus)
+        f = self.calc_bus_value
+        deriv[0, 2, 0] = self.numeric_deriv(f, 'm', 2, bus=bus)
+        deriv[0, 2, 2] = self.numeric_deriv(f, 'h', 2, bus=bus)
+        deriv[0, 5, 2] = self.numeric_deriv(f, 'h', 5, bus=bus)
         return deriv
 
     def initialise_source(self, c, key):
