### 2020-04-15 396703c
Remove plotting residuals, e. g. use get_attr('someproperty').val instead
