### 2020-04-15 cfd375c
Rename bus_func_evaluation calls
