+++ doc/whats_new/v0-3-0.rst
@@ -3,11 +3,10 @@ v0.3.0 - Someversion (somemonth, someday, 2020)
 
 New Features
 ############
-- Make entropy and specific volume available after simulation. Additionally,
-  the method :py:meth:`tespy.connections.connection.get_fluid_properties` may
-  be used, to return the full fluids state of a connection: Pressure,
-  specific enthalpy, temperature, specific volume, specific entropy and
-  vapor mass fraction (`PR #172 <https://github.com/oemof/tespy/pull/172>`_).
+- Make entropy :code:`myconn.s.val` or :code:`myconn.s.val_SI` and specific
+  volume :code:`myconn.vol.val` or :code:`myconn.s.val_SI` respectively
+  available after simulation
+  (`PR #172 <https://github.com/oemof/tespy/pull/172>`_).
 - Connections, components and busses are now easily accessible on every
   network. Connections now have a label, which can be specified individually,
   the default value of the label is given by the following logic.
@@ -43,11 +42,38 @@ New Features
       myconnection.set_attr(fluid={'DowQ': 1, 'water': 0, 'air': 0})
 
   (`PR #174 <https://github.com/oemof/tespy/pull/174>`_).
-
 - Unsetting a value is now also possible using :code:`None` instead of
   :code:`np.nan` (`PR #178 <https://github.com/oemof/tespy/pull/178>`_).
+- It is possible to choose a base for your components when adding them to a
+  bus:
 
+    - :code:`'component'`: This means, the efficiency definition of the bus is
 
+      .. math::
+
+          \eta=\frac{\dot{E}_\mathrm{bus}}{\dot{E}_\mathrm{component}}
+
+    - :code:`'bus'`: This means, the efficiency definition of the bus is
+
+      .. math::
+
+          \eta=\frac{\dot{E}_\mathrm{component}}{\dot{E}_\mathrm{bus}}
+
+  These definitions also apply to the characteristic line usage. The default
+  setting (if you do not specify the base) uses :code:`'base': 'component'`.
+  You therefore do not need do adapt your scripts regarding this change. For
+  an example of the bus usage see the examples section below
+  (`PR #157 <https://github.com/oemof/tespy/pull/157>`_).
+- Within the busses' printouts the efficiency of every component on the
+  respective bus is displayed. Additionally, it is possible to get the
+  bus value and the efficiency of a specific component on a specific bus.
+
+    - to get the bus value use :code:`mycomp.calc_bus_value(mybus)` and
+    - to get the efficiency use :code:`mycomp.calc_bus_efficiency(mybus)`
+
+  where :code:`mycomp` is a component object and :code:`mybus` is a bus the
+  component has been added to
+  (`PR #157 <https://github.com/oemof/tespy/pull/157>`_).
 
 Documentation
 #############
@@ -56,6 +82,9 @@ Documentation
 
 Parameter renaming
 ##################
+- Adding components to a bus now requires :code:`'comp'` instead of :code:`'c'`
+  and :code:`'param'` instead of :code:`'p'`
+  (`PR #157 <https://github.com/oemof/tespy/pull/157>`_).
 
 Testing
 #######
@@ -67,6 +96,8 @@ Testing
 - Add a gasturbine model test for comparison of the stoichiometric combustion
   chamber with the standard combustion chamber
   (`PR #174 <https://github.com/oemof/tespy/pull/174>`_).
+- Add bus tests for efficiency definitions
+  (`PR #157 <https://github.com/oemof/tespy/pull/157>`_).
 
 Bug fixes
 #########
@@ -78,13 +109,13 @@ Other changes
 #############
 - Improve naming in networks, components and connections modules
   (`PR #172 <https://github.com/oemof/tespy/pull/172>`_,
[…]
