### 2020-04-15 #2109 @fwitte
Hi guys,
I just wanted to let you know, how I plan my workflow for the review process:

- For any issues/proposals/... regarding the software, I will open a PR over at https://github.com/oemof/tespy and merge the changes into the dev branch.
- The documentation of the dev branch automatically updates and can be found at https://tespy.readthedocs.io/en/dev/index.html. I also append a short overview of the changes applied in every PR (https://tespy.readthedocs.io/en/dev/whats_new.html).
- If the review is successful, I will then merge the dev branch into the master branch and publish the version 0.3.0.

If you have any objections or suggestions, just let me know :). Thank you and stay healthy!
Best regards
Francesco
