### 2020-04-15 cc6570a
Add bus feature description and example
