+++ doc/tespy_modules/connections.rst
@@ -132,20 +132,20 @@ Busses
 ------
 
 Busses are energy flow connectors. You can sum the energy flow of different
-components and create relations between components regarding mass free energy
-transport.
+components and create relations between components regarding mass independent
+energy transport.
 
 Different use-cases for busses could be:
 
-- Easy post-processing.
-- Introduce motor or generator efficiencies.
-- Create relations of different components.
+- post-processing
+- introduce motor or generator efficiencies
+- create relations of different components
 
 The handling of busses is very similar to connections and components. You need
 to add components to your busses as a dictionary containing at least the
 instance of your component. Additionally you may provide a characteristic line,
-linking the ratio of actual value to referenced value (design case value) to a
-factor the actual value of the component is multiplied with on the bus. For
+linking the ratio of actual value to a referenced value (design case value) to
+an efficiency factor the component value of the bus is multiplied with. For
 instance, you can provide a characteristic line of an electrical generator or
 motor for a variable conversion efficiency. The referenced value is retrieved
 by the design point of your system. Offdesign calculations use the referenced
@@ -156,27 +156,56 @@ case, the ratio will always be 1.
 
     The available keywords for the dictionary are:
 
-    - 'c' for the component instance.
-    - 'p' for the parameter (the combustion engine has various parameters,
-      have a look at the :ref:`combustion engine example <combustion_engine_label>`).
-    - 'P_ref' for the reference value of the component.
-    - 'char' for the characteristic line.
+    - 'comp' for the component instance.
+    - 'param' for the parameter (e. g. the combustion engine has various
+      parameters, have a look at the
+      :ref:`combustion engine example <combustion_engine_label>`)
+    - 'char' for the characteristic line
+    - 'base' the base for efficiency definition
+    - 'P_ref' for the reference value of the component
 
     There are different specification possibilities:
 
-    - If you specify the component only, the parameter will be default
-      (not working with combustion engine) and the conversion factor of the
-      characteristic line will be 1 for every load.
-    - If you specify a numeric value for char, the conversion factor will be
-      equal to that value for every load.
+    - If you specify the component only, the parameter will be default and the
+      efficiency factor of the characteristic line will be 1 independent of
+      the load.
+    - If you specify a numeric value for char, the efficiency factor will be
+      equal to that value independet of the load.
     - If you want to specify a characteristic line, provide
       a :py:class:`char_line <tespy.tools.characteristics.char_line>`
       object.
+    - Specify :code:`'base': 'bus'` if you want to change from the default base
+      to the bus as base. This means, that the definition of the efficiency
+      factor will change according to your specification.
 
-The examples below shows the implementation of busses in your TESPy simulation.
+      .. math ::
 
-Create a pump that is powered by a turbine. The turbine's power output must
-therefore be equal to the pump's power consumption.
+          \eta = \begin{cases}
+          \frac{\dot{E}_\mathrm{component}}{\dot{E}_\mathrm{bus}} &
+          \text{'base': 'bus'}\\
[…]
