### 2020-04-16 9a9d771
Add bus efficiency and value calculation to documentation
