### 2020-04-16 #2109 @fwitte
Great, do not worry, I did/do not want to rush you.

Btw.: If you are testing code etc. you may want to install the dev version instead of the PyPi version (which is version 0.2.2) as the dev version already has some commits since 0.2.2 and breaks with the 0.2.x API. Just in case you do not know (you'll probably know :)), here is the workflow I usually go with:

0. If you have TESPy already installed remove it: `pip uninstall tespy`.
1. Clone the github repo using ssh `git@github.com:oemof/tespy.git` or using `https://github.com/oemof/tespy.git`
2. In a console activate your environment, navigate to the github clone and install manually `pip install -e ./`
3. If changes to dev are beeing made, you installation automatically updates if you go to your github folder and pull from origin `git pull origin dev`

Have a nice week, cheers!
Francesco
