### 2020-04-17 ac94a02
Update network documentation
