+++ CHANGELOG.rst
@@ -1,0 +1,8 @@
+
+Changelog
+=========
+
+0.3.0 dev (2020-04-17)
+----------------------
+
+* First release on PyPI.
