+++ README.rst
@@ -1,93 +1,116 @@
-.. image:: https://coveralls.io/repos/github/oemof/tespy/badge.svg?branch=dev
-    :target: https://coveralls.io/github/oemof/tespy?branch=dev
-.. image:: https://travis-ci.org/oemof/tespy.svg?branch=dev
+========
+Overview
+========
+
+.. start-badges
+
+.. list-table::
+    :stub-columns: 1
+
+    * - docs
+      - |docs|
+    * - tests
+      - | |travis| |appveyor|
+        | |coveralls| |codecov|
+        | |codacy| |codeclimate|
+    * - package
+      - | |version| |wheel| |supported-versions| |supported-implementations|
+        | |commits-since|
+.. |docs| image:: https://readthedocs.org/projects/tespy/badge/?style=flat
+    :target: https://readthedocs.org/projects/tespy
+    :alt: Documentation Status
+
+.. |travis| image:: https://api.travis-ci.org/oemof/tespy.svg?branch=master
+    :alt: Travis-CI Build Status
     :target: https://travis-ci.org/oemof/tespy
-.. image:: https://readthedocs.org/projects/tespy/badge/?version=dev
-    :target: https://tespy.readthedocs.io/en/dev/
-.. image:: https://badge.fury.io/py/TESPy.svg
-    :target: https://badge.fury.io/py/TESPy
-.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.2555866.svg
-   :target: https://doi.org/10.5281/zenodo.2555866
 
-TESPy stands for "Thermal Engineering Systems in Python" and provides a
-powerful simulation toolkit for thermal engineering plants such as power
-plants, district heating systems or heat pumps. It is an external extension
-module within the `Open Energy Modelling Framework <https://oemof.org/>`_ and
-can be used as a standalone package.
+.. |appveyor| image:: https://ci.appveyor.com/api/projects/status/github/oemof/tespy?branch=master&svg=true
+    :alt: AppVeyor Build Status
+    :target: https://ci.appveyor.com/project/oemof/tespy
 
-With the TESPy package you are able to calculate stationary operation in order
-to design the process of thermal energy systems. From that point it is possible
-to simulate the offdesign behaviour of your plant using underlying
-characteristics for each of the plants components. The package inlcudes basic
-components, such as turbines, pumps, compressors, heat exchangers, pipes,
-mixers and splitters as well as some advanced components (derivatives of heat
-exchangers, drum).
+.. |coveralls| image:: https://coveralls.io/repos/oemof/tespy/badge.svg?branch=master&service=github
+    :alt: Coverage Status
+    :target: https://coveralls.io/r/oemof/tespy
 
-Everybody is welcome to use and/or develop TESPy. Contribution is already
-possible on a low level by simply fixing typos in TESPy's documentation or
-rephrasing sections which are unclear. If you want to support us that way
-please fork the TESPy repository to your own github account and make changes
-as described in the github guidelines:
-https://guides.github.com/activities/hello-world/
+.. |codecov| image:: https://codecov.io/gh/oemof/tespy/branch/master/graphs/badge.svg?branch=master
+    :alt: Coverage Status
+    :target: https://codecov.io/github/oemof/tespy
+
+.. |codacy| image:: https://img.shields.io/codacy/grade/CODACY_TOKEN.svg
+    :target: https://www.codacy.com/app/oemof/tespy
+    :alt: Codacy Code Quality Status
+
+.. |codeclimate| image:: https://codeclimate.com/github/oemof/tespy/badges/gpa.svg
+   :target: https://codeclimate.com/github/oemof/tespy
+   :alt: CodeClimate Quality Status
+
[…]
