+++ tox.ini
@@ -1,0 +1,90 @@
+[tox]
+envlist =
+    clean,
+    check,
+    docs,
+    py36,
+    py37,
+    report
+
+[testenv]
+basepython =
+    docs: {env:TOXPYTHON:python3.6}
+    {bootstrap,clean,check,report,codecov,coveralls}: {env:TOXPYTHON:python3}
+setenv =
+    PYTHONPATH={toxinidir}/tests
+    PYTHONUNBUFFERED=yes
+passenv =
+    *
+deps =
+    pytest
+    pytest-travis-fold
+    pytest-cov
+commands =
+    {posargs:pytest --cov --cov-report=term-missing -vv --ignore=src}
+
+[testenv:bootstrap]
+deps =
+    jinja2
+    matrix
+skip_install = true
+commands =
+    python ci/bootstrap.py --no-env
+
+[testenv:check]
+deps =
+    docutils
+    check-manifest
+    flake8
+    readme-renderer
+    pygments
+    isort
+skip_install = true
+commands =
+    python setup.py check --strict --metadata --restructuredtext
+    check-manifest {toxinidir}
+    flake8
+    isort --verbose --check-only --diff --recursive
+
+
+[testenv:docs]
+usedevelop = true
+deps =
+    -r{toxinidir}/docs/requirements.txt
+commands =
+    sphinx-build {posargs:-E} -b html docs dist/docs
+    sphinx-build -b linkcheck docs dist/docs
+
+[testenv:coveralls]
+deps =
+    coveralls
+skip_install = true
+commands =
+    coveralls []
+
+
+
+[testenv:codecov]
+deps =
+    codecov
+skip_install = true
+commands =
+    codecov []
+
+[testenv:report]
+deps = coverage
+skip_install = true
+commands =
+    coverage report
+    coverage html
+
+[testenv:clean]
+commands = coverage erase
+skip_install = true
+deps = coverage
+
+[testenv:py36]
+basepython = {env:TOXPYTHON:python3.6}
+
+[testenv:py37]
+basepython = {env:TOXPYTHON:python3.7}
