+++ docs/requirements.txt
@@ -1,2 +1,4 @@
+sphinx>=1.3
+sphinx-rtd-theme
 matplotlib
 sphinxcontrib-bibtex
