+++ docs/requirements.txt
@@ -2,3 +2,4 @@ sphinx>=1.3
 sphinx-rtd-theme
 matplotlib
 sphinxcontrib-bibtex
+tespy
