### 2020-04-17 89ce142
Remove deprecated nose import
