### 2020-04-17 c21f637
Include md, yml, bib and tutorial files
