### 2020-04-17 7a6667b
Update version number according to PEP 440
