+++ tests/test_models/test_heat_pump_model.py
@@ -9,21 +9,19 @@ tests/test_models/test_heat_pump_model.py
 
 SPDX-License-Identifier: MIT
 """
+import numpy as np
+import shutil
 from tespy.components.basics import sink, source
-from tespy.components.heat_exchangers import (heat_exchanger,
-                                              heat_exchanger_simple)
+from tespy.components.heat_exchangers import (
+    heat_exchanger, heat_exchanger_simple
+)
 from tespy.components.nodes import drum
 from tespy.components.piping import valve
 from tespy.components.turbomachinery import compressor, pump
 from tespy.connections import connection, bus, ref
 from tespy.networks.networks import network
-from tespy.tools.data_containers import dc_cc
 from tespy.tools.characteristics import char_line
-from tespy.tools.characteristics import load_default_char as ldc
-
-import numpy as np
-import shutil
-import logging
+from tespy.tools.data_containers import dc_cc
 
 
 class TestHeatPump:
