+++ tests/test_components/test_piping.py
@@ -9,16 +9,13 @@ tests/test_components/test_piping.py
 
 SPDX-License-Identifier: MIT
 """
-
+import numpy as np
 from tespy.components.basics import sink, source
 from tespy.components.piping import valve, pipe
 from tespy.connections import connection
 from tespy.networks.networks import network
 from tespy.tools.characteristics import char_line
 from tespy.tools.data_containers import dc_cc
-
-import numpy as np
-import shutil
 
 
 def convergence_check(lin_dep):
