+++ tests/test_components/test_heat_exchangers.py
@@ -9,19 +9,15 @@ tests/test_components/test_heat_exchangers.py
 
 SPDX-License-Identifier: MIT
 """
-
+import numpy as np
+import shutil
 from tespy.components.basics import sink, source
-from tespy.components.heat_exchangers import (heat_exchanger_simple,
-                                              solar_collector,
-                                              heat_exchanger, condenser)
+from tespy.components.heat_exchangers import (
+    heat_exchanger_simple, solar_collector, heat_exchanger, condenser
+)
 from tespy.connections import connection, bus
 from tespy.networks.networks import network
 from tespy.tools.fluid_properties import T_bp_p
-
-import logging
-
-import numpy as np
-import shutil
 
 
 def convergence_check(lin_dep):
@@ -312,7 +308,6 @@ class TestHeatExchangers:
         # condenser the temperature of the condensing fluid is relevant.
         ttd_u = round(T_bp_p(self.c1.to_flow()) - self.c4.T.val_SI, 1)
         p = round(self.c1.p.val_SI, 5)
-        kA = instance.kA.val
         msg = ('Value of terminal temperature difference must be ' +
                str(round(instance.ttd_u.val, 1)) + ', is ' +
                str(ttd_u) + '.')
