+++ CHANGELOG.rst
@@ -1,8 +1,0 @@
-
-Changelog
-=========
-
-0.3.0 dev (2020-04-17)
-----------------------
-
-* First release on PyPI.
