+++ setup.cfg
@@ -3,7 +3,7 @@ universal = 1
 
 [flake8]
 max-line-length = 140
-exclude = .tox,ci/templates
+exclude = */migrations/*, docs/conf.py
 
 [options]
 # tests_require is a list of dependencies that are *absolutely required*
