+++ tox.ini
@@ -43,7 +43,7 @@ skip_install = true
 commands =
     python setup.py check --strict --metadata --restructuredtext
     check-manifest {toxinidir}
-    flake8
+    flake8 src tests setup.py
     isort --verbose --check-only --diff --recursive
 
 
@@ -61,8 +61,6 @@ deps =
 skip_install = true
 commands =
     coveralls []
-
-
 
 [testenv:codecov]
 deps =
@@ -85,6 +83,22 @@ deps = coverage
 
 [testenv:py36]
 basepython = {env:TOXPYTHON:python3.6}
+setenv =
+    {[testenv]setenv}
+usedevelop = true
+commands =
+    {posargs:pytest --cov --cov-report=term-missing -vv}
+deps =
+    {[testenv]deps}
+    pytest-cov
 
 [testenv:py37]
 basepython = {env:TOXPYTHON:python3.7}
+setenv =
+    {[testenv]setenv}
+usedevelop = true
+commands =
+    {posargs:pytest --cov --cov-report=term-missing -vv}
+deps =
+    {[testenv]deps}
+    pytest-cov
