### 2020-04-17 cb1caed
Fix flag positioning
