+++ src/tespy/__init__.py
@@ -3,16 +3,16 @@
 __version__ = '0.3.0 - dev'
 
 # tespy networks imports
-from .networks import (
+from .networks import (  # noqa: F401
     network_reader,
     networks
-)  # noqa: F401
+)
 
 # tespy connection imports
 from . import connections  # noqa: F401
 
 # tespy components imports
-from .components import (
+from .components import (  # noqa: F401
     basics,
     combustion,
     components,
@@ -22,17 +22,17 @@ from .components import (
     reactors,
     subsystems,
     turbomachinery
-)  # noqa: F401
+)
 
 # tespy data
 from . import data  # noqa: F401
 
 # tespy tools imports
-from .tools import (
+from .tools import (  # noqa: F401
     characteristics,
     data_containers,
     fluid_properties,
     global_vars,
     helpers,
     logger
-)  # noqa: F401
+)
