+++ src/tespy/tools/__init__.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8
 
-from .data_containers import (
+from .data_containers import (  # noqa: F401
     dc_cc, dc_cm, dc_cp, dc_flu, dc_gcp, dc_prop, dc_simple
-    )
+)
 
-from .characteristics import (
+from .characteristics import (  # noqa: F401
     char_line,
     char_map,
     load_custom_char,
     load_default_char
-    )
+)
