### 2020-04-17 331bcff
Adjustments according to lgtm issues
