+++ tutorial/step_3.py
@@ -30,14 +30,6 @@ cd = condenser('condenser')
 rp = pump('recirculation pump')
 cons = heat_exchanger_simple('consumer')
 
-cp1 = sink('compressor 1')
-
-# consumer system
-
-cd = condenser('condenser')
-rp = pump('recirculation pump')
-cons = heat_exchanger_simple('consumer')
-
 # evaporator system
 
 va = valve('valve')
