+++ setup.cfg
@@ -75,7 +75,7 @@ known_first_party = tespy
 default_section = THIRDPARTY
 forced_separate = test_tespy
 not_skip = __init__.py
-skip = .tox,ci/templates
+skip = migrations
 
 [matrix]
 # This is the configuration for the `./bootstrap.py` script.
@@ -108,6 +108,10 @@ dependencies =
 #    1.6: Django==1.6.8
 #    1.7: Django==1.7.1 !python_versions[py26]
 # Deps commented above are provided as examples. That's what you would use in a Django project.
+
+coverage_flags =
+    cover: true
+    nocov: false
 
 environment_variables =
     -
