+++ src/tespy/components/piping.py
@@ -16,12 +16,15 @@ SPDX-License-Identifier: MIT
 """
 
 import numpy as np
+
 from tespy.components.components import component
 from tespy.components.heat_exchangers import heat_exchanger_simple
-from tespy.tools.data_containers import dc_cc, dc_cp, dc_simple
-from tespy.tools.fluid_properties import s_mix_ph, v_mix_ph
+from tespy.tools.data_containers import dc_cc
+from tespy.tools.data_containers import dc_cp
+from tespy.tools.data_containers import dc_simple
+from tespy.tools.fluid_properties import s_mix_ph
+from tespy.tools.fluid_properties import v_mix_ph
 from tespy.tools.global_vars import err
-
 
 # %%
 
