+++ src/tespy/components/reactors.py
@@ -14,14 +14,22 @@ available from its original location tespy/components/reactors.py
 SPDX-License-Identifier: MIT
 """
 
-import CoolProp.CoolProp as CP
 import logging
+
+import CoolProp.CoolProp as CP
 import numpy as np
+
 from tespy.components.components import component
-from tespy.tools.data_containers import dc_cc, dc_cp, dc_simple
-from tespy.tools.fluid_properties import (
-    h_mix_pT, T_mix_ph, dT_mix_dph, dT_mix_pdh, v_mix_ph)
-from tespy.tools.global_vars import molar_masses, err
+from tespy.tools.data_containers import dc_cc
+from tespy.tools.data_containers import dc_cp
+from tespy.tools.data_containers import dc_simple
+from tespy.tools.fluid_properties import T_mix_ph
+from tespy.tools.fluid_properties import dT_mix_dph
+from tespy.tools.fluid_properties import dT_mix_pdh
+from tespy.tools.fluid_properties import h_mix_pT
+from tespy.tools.fluid_properties import v_mix_ph
+from tespy.tools.global_vars import err
+from tespy.tools.global_vars import molar_masses
 from tespy.tools.helpers import TESPyComponentError
 
 # %%
