+++ tests/test_tools/test_characteristics.py
@@ -10,13 +10,17 @@ tests/test_tools/test_characteristics.py
 SPDX-License-Identifier: MIT
 """
 import json
-import numpy as np
 import os
-from pkg_resources import resource_filename
 import shutil
-from tespy.tools.characteristics import (
-    char_line, char_map, compressor_map, load_default_char, load_custom_char
-)
+
+import numpy as np
+from pkg_resources import resource_filename
+
+from tespy.tools.characteristics import char_line
+from tespy.tools.characteristics import char_map
+from tespy.tools.characteristics import compressor_map
+from tespy.tools.characteristics import load_custom_char
+from tespy.tools.characteristics import load_default_char
 from tespy.tools.helpers import extend_basic_path
 
 
