+++ src/tespy/components/__init__.py
@@ -1,15 +1,26 @@
 # -*- coding: utf-8
 
-from .basics import cycle_closer, sink, source, subsystem_interface  # noqa: F401
-from .combustion import (  # noqa: F401
-    combustion_chamber, combustion_chamber_stoich, combustion_engine
-)
-from .heat_exchangers import (  # noqa: F401
-    heat_exchanger_simple, solar_collector,
-    condenser, desuperheater, heat_exchanger
-)
-from .nodes import drum, merge, node, separator, splitter  # noqa: F401
-from .piping import pipe, valve  # noqa: F401
+from .basics import cycle_closer  # noqa: F401
+from .basics import sink  # noqa: F401
+from .basics import source  # noqa: F401
+from .basics import subsystem_interface  # noqa: F401
+from .combustion import combustion_chamber  # noqa: F401
+from .combustion import combustion_chamber_stoich  # noqa: F401
+from .combustion import combustion_engine  # noqa: F401
+from .heat_exchangers import condenser  # noqa: F401
+from .heat_exchangers import desuperheater  # noqa: F401
+from .heat_exchangers import heat_exchanger  # noqa: F401
+from .heat_exchangers import heat_exchanger_simple  # noqa: F401
+from .heat_exchangers import solar_collector  # noqa: F401
+from .nodes import drum  # noqa: F401
+from .nodes import merge  # noqa: F401
+from .nodes import node  # noqa: F401
+from .nodes import separator  # noqa: F401
+from .nodes import splitter  # noqa: F401
+from .piping import pipe  # noqa: F401
+from .piping import valve  # noqa: F401
 from .reactors import water_electrolyzer  # noqa: F401
 from .subsystems import subsystem  # noqa: F401
-from .turbomachinery import compressor, pump, turbine  # noqa: F401
+from .turbomachinery import compressor  # noqa: F401
+from .turbomachinery import pump  # noqa: F401
+from .turbomachinery import turbine  # noqa: F401
