+++ tests/test_busses.py
@@ -9,12 +9,17 @@ tests/test_busses.py
 
 SPDX-License-Identifier: MIT
 """
-import numpy as np
 import shutil
-from tespy.components.basics import sink, source
+
+import numpy as np
+
+from tespy.components.basics import sink
+from tespy.components.basics import source
 from tespy.components.combustion import combustion_chamber
-from tespy.components.turbomachinery import compressor, turbine
-from tespy.connections import connection, bus
+from tespy.components.turbomachinery import compressor
+from tespy.components.turbomachinery import turbine
+from tespy.connections import bus
+from tespy.connections import connection
 from tespy.networks.networks import network
 from tespy.tools.characteristics import char_line
 

[…]
