### 2020-04-17 64df1fe
Only run py36 and py37 for appveyor
