### 2020-04-19 bf872b3
Add codecoverage triggers
