+++ tox.ini
@@ -19,9 +19,8 @@ passenv =
 deps =
     pytest
     pytest-travis-fold
-    pytest-cov
 commands =
-    {posargs:pytest --cov --cov-report=term-missing -vv --ignore=src}
+    {posargs:pytest -vv --ignore=src}
 
 [testenv:bootstrap]
 deps =
@@ -44,7 +43,7 @@ commands =
     python setup.py check --strict --metadata --restructuredtext
     check-manifest {toxinidir}
     flake8 src tests setup.py
-    isort --verbose --check-only --diff --recursive
+    isort --verbose --check-only --diff --recursive src tests setup.py
 
 
 [testenv:docs]
@@ -81,7 +80,7 @@ commands = coverage erase
 skip_install = true
 deps = coverage
 
-[testenv:py36]
+[testenv:py36-cover]
 basepython = {env:TOXPYTHON:python3.6}
 setenv =
     {[testenv]setenv}
