+++ tox.ini
@@ -80,7 +80,7 @@ commands = coverage erase
 skip_install = true
 deps = coverage
 
-[testenv:py36-cover]
+[testenv:py36]
 basepython = {env:TOXPYTHON:python3.6}
 setenv =
     {[testenv]setenv}
