### 2020-04-19 oemof/tespy#180 @arosen93
Comments on JOSS Manuscript
This is not really the ideal place for me to put this, but I'd rather share my comments now than wait for the JOSS review to formally be opened. I will link to this Issue when the review opens.

With regards to the manuscript hosted under the [pre-review](https://github.com/openjournals/joss-reviews/issues/2109), I have the following comments:
- It is obvious to most using this package what the variables mean, but I think it'd still be nice for them to be defined. One of the great things about making an open-source package like this is it makes it accessible to people who otherwise aren't experts in the field. For instance, a chemical engineer might initially interpret P to be pressure rather than power, which of course is inappropriate. Similarly, L, H, V_j, theta_{log}, etc. should be defined (all variables, actually).
- Since there will be people using this package from all backgrounds, could you explicitly define in the text the Q convention used? Just to prevent confusion.
- The equations are not numbered but still referenced with numbers. This should be adjusted.
- You may wish to write the general expression for the open-system energy balance before Equation 3, then in the text it can be made clear which terms are neglected overall (e.g. differences in potential energy due to height) as well as which terms are neglected for specific process equipment highlighted in the text (e.g. adiabatic turbomachinery). This would make it easier for the reader and of course more general.
- I think it would be helpful in the manuscript to briefly describe the layout of the code, highlighting the modular nature. For instance, on the TESPy modules page, you describe the Networks, Components, component groups, Connections, and so on. It would be worth briefly highlighting the different modules that make up TESPy in the manuscript.
- Are there any other attempts at open-source alternatives to Aspen Plus in Python or other languages? If so, these should be referenced and distinguishing qualities of TESPy highlighted.
- Several software packages are cited under the acknowledgments. SciPy now has a formal citation [here](https://www.nature.com/articles/s41592-019-0686-2?luicode=10000011&lfid=1008082086c7dfebc09fc300733002ea997ba2_-_feed&u=https%3A%2F%2Fwww.nature.com%2Farticles%2Fs41592-019-0686-2), which you may consider adding.
- Please update the pandas citation to include the journal, as mentioned [here](https://pandas.io/about/citing.html).
- Several other contributors are listed on the GitHub contributor list, although the contributions look to be somewhat minor. Per the JOSS review criteria, I feel obligated to ask you to confirm that no additional authors should be added to the current review.

Very minor typographical comments that do not influence the quality of the submitted manuscript:
- Please capitalize Python under the Motivation section
- Please change E. g. to not have a space between the abbreviated letters
- Please italicize fl in the second line of the second page to match the variable in the equations
- I was expecting the Example Implementations section to be like a mini tutorial of how the code is working. Perhaps it would be better to rename the section something more like Previous Implementations.
