### 2020-04-20 0722319
Implement first part of suggestions
