+++ paper.md
@@ -45,7 +45,7 @@ available in this field, for example EBSILON Professional or Aspen Plus, mainly
 used in the industrial environment. In order to open this kind of software to a
 wide (scientific) audience, an open source solution is very promising. Its
 prevalence in the scientific field and the availability of interfaces to other
-programming languages plead for a python implementation.
+programming languages plead for a Python implementation.
 
 # Method
 
@@ -60,9 +60,10 @@ representing the plant's topology and its parametrisation. By formulating the
 equations implicitly parameters and results are generally interchangeable
 offering high flexibility in the user specifications. The system of equations is
 numerically solved with an inbuilt solver applying the multi-dimensional
-Newton-Raphson method to determine the mass flow, pressure, enthalpy and fluid
-composition of every connection. This way, it is possible to solve for both
-thermal as well as hydraulic state of the plant.
+Newton-Raphson method to determine the mass flow $\dot{m}$, pressure $p$,
+enthalpy $h$ and fluid composition (defined by mass fraction $x$ of each fluid)
+of every connection. This way, it is possible to solve for both thermal as well
+as hydraulic state of the plant.
 
 To achieve this, TESPy implements balance equations (based on standard
 literature, for example [@Baehr; @Epple]) for every component regarding
@@ -73,35 +74,35 @@ literature, for example [@Baehr; @Epple]) for every component regarding
 
 In steady state, the total mass flow into a component must be equal to the total
 mass flow leaving the component (eq. 1). Additionally, the mass balance of a
-specific fluid fl is applied (eq. 2). The energy balance of all components is
+specific fluid $fl$ is applied (eq. 2). The energy balance of all components is
 derived from the stationary energy balance of open systems with multiple inlets
 and outlets. Differences in flow velocity as well as height are neglected as
 these are relatively small compared to change in enthalpy in thermal engineering
-applications. The values of heat and power transferred, depend on the individual
-component properties. For example, a pipe does not transfer power, thus only
-heat may be transferred. In contrast, turbomachinery is considered adiabatic,
-thus only transfers power (eq. 3). If chemical reactions take place, the
-corresponding chemical mass balance is taken into account instead of equation 2.
+applications. The values of heat $\dot{Q}$ and power $P$ transferred,
+depend on the individual component properties. For example, a pipe does not
+transfer power, thus only heat may be transferred. In contrast, turbomachinery
+is considered adiabatic, thus only transfers power (eq. 3). If chemical
+reactions take place, the corresponding chemical mass balance is taken into
+account instead of equation 2.
 On top of that, the energy balance is different, as the reaction enthalpy has to
-be considered, too. Furthermore, it is necessary to compensate for the different
-zero point definitions of enthalpy in the fluid properties of the reaction
-components by defining a reference state. E. g. for a combustion chamber
-equation 4 is implemented.
+be considered by using the fuel's lower heating value $LHV$. Furthermore, it is
+necessary to compensate for the different zero point definitions of enthalpy in
[…]
