### 2020-04-20 a998f3a
Implement suggested changes
