+++ paper.md
@@ -42,16 +42,34 @@ parametrisation.
 Thermal process simulation is a fundamental discipline in energy engineering.
 Consequently, there are several well known commercial software solutions
 available in this field, for example EBSILON Professional or Aspen Plus, mainly
-used in the industrial environment. In order to open this kind of software to a
-wide (scientific) audience, an open source solution is very promising. Its
-prevalence in the scientific field and the availability of interfaces to other
-programming languages plead for a Python implementation.
+used in the industrial environment. A similar approach was carried out by
+[@ThermoPower] with the open source library "ThermoPower" written in the
+Modelica language. The software provides dynamic modeling and focuses on steam
+power plants, but has not been updated since 2014 [@ThermoPowerOnline].
+
+Due to the fact that ThermoPower is not updated anymore and limited on
+one field of thermal engineering a new solution is proposed. In order to open
+the software to a wide (scientific) audience, an open source solution
+implemented in a widespread programming language like Python, is very promising.
+
+# Package Architecture
+
+TESPy is built up in modular structure with three main modules:
+
+• networks: The networks module represents the container for every simulation.
+  The solution process, pre- and postprocessing are handled by the network
+  class.
+• components: All components are part of the components module. The package
+  provides basic components, for example different types of turbomachinery, heat
+  exchangers, pipes, valves and mixers or splitters. On top of that, advanced
+  components like a drum, an internal combustion engine or a combustion chamber
+  are available. The individual properties of each component are defined in the respective class.
+• connections: Connections are used to connect components and hold the fluid
+  property information. Thus, they represent the plant's topology and its state.
 
 # Method
 
-The package is built of basic components, for example different types of
-turbomachinery, heat exchangers, pipes, valves and mixers or splitters. To
-simulate a specific plant an individual model is created connecting the
+To simulate a specific plant an individual model is created connecting the
 components to form a topological network. Steady state operation of the plant is
 determined by the fluid's state on every connection (and therefore its change
 within components) between the plant's individual components. Thus, based on the
@@ -69,55 +87,91 @@ To achieve this, TESPy implements balance equations (based on standard
 literature, for example [@Baehr; @Epple]) for every component regarding
 
 • mass flow and fluid composition as well as
-
 • energy (covering thermal and hydraulic properties).
 
 In steady state, the total mass flow into a component must be equal to the total
-mass flow leaving the component (eq. 1). Additionally, the mass balance of a
-specific fluid $fl$ is applied (eq. 2). The energy balance of all components is
-derived from the stationary energy balance of open systems with multiple inlets
-and outlets. Differences in flow velocity as well as height are neglected as
-these are relatively small compared to change in enthalpy in thermal engineering
-applications. The values of heat $\dot{Q}$ and power $P$ transferred,
-depend on the individual component properties. For example, a pipe does not
[…]
