+++ paper.bib
@@ -38,6 +38,21 @@
     doi = {10.1038/s41592-019-0686-2},
 }
 
+@article{ThermoPower,
+    author = {Casella, Francesco and Leva, Alberto},
+    year = {2003},
+    month = {11},
+    pages = {41-50},
+    title = {Modelica open library for power plant simulation: Design and experimental validation},
+    journal = {Proc. 3rd Modelica Conference}
+}
+
+@misc{ThermoPowerOnline,
+    title = {ThermoPower},
+    howpublished = {\url{https://build.openmodelica.org/Documentation/ThermoPower.html}},
+    note = {Accessed: 2020-04-20}
+}
+
 @article{NumPy,
     doi = {10.1109/mcse.2011.37},
     year = {2011},
