### 2020-04-20 9df5ecd
Fix LaTeX issue in eq 8
