+++ paper.md
@@ -106,11 +106,11 @@ can be simplified (eq. \ref{eq:4}).
 0=\underset{i}{\sum}\dot{m}_{\mathrm{in,}i}\cdot x_{fl\mathrm{,in,}i}-
 \underset{o}{\sum}\dot{m}_{\mathrm{out,}o}\cdot x_{fl\mathrm{,out,}o}
 \ \forall fl\in\mathrm{network\ fluids} \label{eq:2}\\
-0=\underset{o}{\sum}\dot{m}_{\mathrm{out,}o}\cdot \left(
-h_{\mathrm{out,}o} + \mathrm{g} \cdot z_{\mathrm{out,}o} +
+0=\underset{o}{\sum}\dot{m}_{\mathrm{out,}o}\left(
+h_{\mathrm{out,}o} + \mathrm{g} z_{\mathrm{out,}o} +
 \frac{c_{\mathrm{out,}o}^2}{2}\right)-
-\underset{i}{\sum}\dot{m}_{\mathrm{in,}i}\cdot \left(
-h_{\mathrm{in,}i} + \mathrm{g} \cdot z_{\mathrm{in,}i} +
+\underset{i}{\sum}\dot{m}_{\mathrm{in,}i} \left(
+h_{\mathrm{in,}i} + \mathrm{g} z_{\mathrm{in,}i} +
 \frac{c_{\mathrm{in,}i}^2}{2}\right)-P-\dot{Q} \label{eq:3}\\
 0=\underset{o}{\sum}\dot{m}_{\mathrm{out,}o}\cdot h_{\mathrm{out,}o}-
 \underset{i}{\sum}\dot{m}_{\mathrm{in,}i}\cdot h_{\mathrm{in,}i}-P-\dot{Q}
@@ -160,11 +160,10 @@ this, design specific component parameters are calculated in the design case,
 for example: the area independent heat transfer coefficient $kA$ of heat
 exchangers. The heat transfer at a different operation point is calculated from
 the $kA$ value and the logarithmic temperature difference
-$\Delta\vartheta_\mathrm{log}$ in equation \ref{eg:8}.
+$\Delta\vartheta_\mathrm{log}$ in equation \ref{eq:8}.
 
  \begin{equation}
- 0=\dot{Q}-kA\cdot\Delta\vartheta_\mathrm{log}
- \label{eq:8}
+ 0=\dot{Q}-kA\cdot\Delta\vartheta_\mathrm{log} \label{eq:8}
  \end{equation}
 
 In general, the design parameters ($kA$ in case of the heat exchanger) can be
