### 2020-04-20 f82fcf4
Fix equation alignment
