### 2020-04-20 ef71d86
Update sources and polish some lists
