+++ paper.bib
@@ -48,9 +48,10 @@
 }
 
 @misc{ThermoPowerOnline,
-    title = {ThermoPower},
-    howpublished = {\url{https://build.openmodelica.org/Documentation/ThermoPower.html}},
-    note = {Accessed: 2020-04-20}
+  author = {​{Politecnico di Milano}},
+  title = {ThermoPower},
+  year = {2014},
+  url = {​https://build.openmodelica.org/Documentation/ThermoPower.html}
 }
 
 @article{NumPy,
