+++ paper.bib
@@ -51,7 +51,7 @@
   author = {​{Politecnico di Milano}},
   title = {ThermoPower},
   year = {2014},
-  url = {​https://build.openmodelica.org/Documentation/ThermoPower.html}
+  howpublished = {\url{https://build.openmodelica.org/Documentation/ThermoPower.html}}
 }
 
 @article{NumPy,
