### 2020-04-20 fbe9c20
Define c and z
