### 2020-04-20 oemof/tespy#180 @arosen93
@fwitte -- These changes look great! I think this really helps improve the readability of the equations, particularly for those who may not deal with mass/energy balances as often but want to use your code. It's a nice writeup! I'll post remaining comments on the main JOSS page.
