### 2020-04-27 oemof/tespy#185 @arosen93
Doc bugfix
[Here](https://tespy.readthedocs.io/en/master/first_steps.html), `from tespy.components import (sink, source, pipe, valve, heat_excanger_simple)` has a misspelled `heat_exchanger_simple`.

[Here](https://tespy.readthedocs.io/en/master/tutorials_examples.html#heat-pump-tutorial-label) where it says "Enjoy fiddling around with the source code of the cogeneration unit in the examples repository!", the source code link leads to a 404 error.

[Here](https://tespy.readthedocs.io/en/master/tutorials_examples.html), it says "You can download the full code from the tespy_examples repository." but the link is dead. Same for "The example for the stoichiometric combustion chamber can as well be taken from the tespy_examples repository."
