### 2020-04-27 #2178 @arosen93
@fwitte: I can confirm that, in my view, TESPy fulfills all the required checkboxes for the review. In particular, I am very impressed with the extensive documentation and examples provided with the code. I also think the manuscript is greatly improved, now that it more clearly specifies the variables used and takes a more general approach with regards to the equations. I can confirm that the code works as-advertised, and the modular nature of the package is self-evident. I still would like to test out the code for a few more days (and will provide any minor remaining comments then -- probably just clarifications) but will certainly complete the review by the 2-week suggested deadline.

As for right now, I think there is one easy improvement that can be made. One of the checkbox items states "Documentation: Are there automated tests or manual steps described so that the functionality of the software can be verified?" It is clear there is actually quite extensive automated tests, for which I applaud the developers. Still, I think it would be helpful to link to the tests being carried out (and briefly describe where the expected results are) at the corresponding [docs page](https://tespy.readthedocs.io/en/master/developing_tespy.html#tests). Nothing more than a sentence or two is really needed. Also, very minor, but when you say "If you have further questions regarding the tests, do not bother to contact us", it makes it sound like you should _not_ be contacted if there are questions. The phrase that makes more sense here is "do not hesitate to contact us." I also left a few minor comments on the docs [here](https://github.com/oemof/tespy/issues/185).
