### 2020-04-27 8b3a3cd
Address issues from JOSS review
Fix "e.g." typo in first_steps.rst and update the broken
links to the combustion chamber and combustion chamber stoich
examples. Rephrase sentence in the tests section of
developing_tespy.rst.
