### 2020-04-27 oemof/tespy#185 @fwitte
Thank you for your comments!

> [Here](https://tespy.readthedocs.io/en/master/first_steps.html), `from tespy.components import (sink, source, pipe, valve, heat_excanger_simple)` has a misspelled `heat_exchanger_simple`.

Someone else found this just before you, so it was already fixed on the dev branch (https://github.com/oemof/tespy/commit/99abd84449dbdf0d24113b02ad08dd65f03dcddb).

> [Here](https://tespy.readthedocs.io/en/master/tutorials_examples.html#heat-pump-tutorial-label) where it says "Enjoy fiddling around with the source code of the cogeneration unit in the examples repository!", the source code link leads to a 404 error.

This one has also been fixed before (https://github.com/oemof/tespy/commit/08b7d8f5627cd06399910848f3294cdbe018217b#diff-247a44d286b6ee8af451624b963a2bf5L44-R46).

> [Here](https://tespy.readthedocs.io/en/master/tutorials_examples.html), it says "You can download the full code from the tespy_examples repository." but the link is dead. Same for "The example for the stoichiometric combustion chamber can as well be taken from the tespy_examples repository."

Resolved at #181 (https://github.com/oemof/tespy/pull/181/commits/8b3a3cd57dec92462ccd668bcb57bdb691c88e8f).

Cheers!
