### 2020-04-27 #2178 @fwitte
@arosen93: Thank you very much for your comments, take your time and have fun :).

> Very minor comment. When you say "If you have further questions regarding the tests, do not bother to contact us", it makes it sound like you should _not_ be contacted if there are questions. The phrase that makes more sense here is "do not hesitate to contact us."

Haha, I like this one. When there is too much work on the desk, you must find your priorities... Thank you very much for the hint! I addressed this over at https://github.com/oemof/tespy/pull/181 (https://tespy--181.org.readthedocs.build/en/181/developing_tespy.html#tests).

>  I also left a few minor comments on the docs [here](https://github.com/oemof/tespy/issues/185).

Addressed these comments as well.

Have a nice week!
