+++ tespy/components/nodes.py
@@ -453,6 +453,361 @@ class node(component):
         elif key == 'h':
             return 5e5
 
+    # %%
+
+
+class droplet_separator(component):
+    r"""
+    Separate liquid phase from gas phase of a single fluid.
+
+    Equations
+
+        **mandatory equations**
+
+        - :func:`tespy.components.nodes.droplet_separator.fluid_func`
+        - :func:`tespy.components.components.component.mass_flow_func`
+
+        .. math::
+
+            0 = \dot{m}_{in} \cdot h_{in} -
+            \sum_j \left(\dot{m}_{j,out} \cdot h_{j,out} \right)\\
+            \forall j \in outlet
+
+            0 = p_{in} - p_{out,i}\\
+            \forall i \in \mathrm{outlets}
+
+            0 = h_{1,out} - h\left(p, x=1 \right)
+
+            0 = h_{2,out} - h\left(p, x=0 \right)\\
+            x: \text{vapour mass fraction}
+
+    Inlets/Outlets
+
+        - in1
+        - out1, out2 (index 1: gas phase, index 2: liquid phase)
+
+    Image
+
+        .. image:: _images/droplet_separator.svg
+           :scale: 100 %
+           :alt: alternative text
+           :align: center
+
+    Parameters
+    ----------
+    label : str
+        The label of the component.
+
+    design : list
+        List containing design parameters (stated as String).
+
+    offdesign : list
+        List containing offdesign parameters (stated as String).
+
+    design_path: str
+        Path to the components design case.
+
+    local_offdesign : boolean
+        Treat this component in offdesign mode in a design calculation.
+
+    local_design : boolean
+        Treat this component in design mode in an offdesign calculation.
+
+    char_warnings: boolean
+        Ignore warnings on default characteristics usage for this component.
+
+    printout: boolean
+        Include this component in the network's results printout.
+
+    Note
+    ----
+    If you are using a drum in a network with multiple fluids, it is likely
+    the fluid propagation causes trouble. If this is the case, try to
+    specify the fluid composition at another connection of your network.
+
+    This component assumes, that the fluid composition between outlet 1 and
+    inlet 2 does not change, thus there is no equation for the fluid mass
+    fraction at the inlet 2!
+
+    Example
+    -------
+    The droplet separator separates gas from liquid phase. From a stream of
+    water the liquid phase will be separated.
+
+    >>> from tespy.components import sink, source, droplet_separator
+    >>> from tespy.connections import connection
+    >>> from tespy.networks import network
+    >>> from tespy.tools.fluid_properties import Q_ph
+    >>> import shutil
+    >>> nw = network(fluids=['water'], T_unit='C', p_unit='bar',
+    ... h_unit='kJ / kg', iterinfo=False)
+    >>> so = source('two phase inflow')
+    >>> sig = sink('gas outflow')
+    >>> sil = sink('liquid outflow')
+    >>> ds = droplet_separator('droplet separator')
+    >>> ds.component()
+    'droplet separator'
+    >>> so_ds = connection(so, 'out1', ds, 'in1')
+    >>> ds_sig = connection(ds, 'out1', sig, 'in1')
+    >>> ds_sil = connection(ds, 'out2', sil, 'in1')
+    >>> nw.add_conns(so_ds, ds_sig, ds_sil)
+
+    We specify the fluid's state at the inlet. At the gas outflow saturated
+    gas enthalpy is expected, at the liquid gas outflow saturated liquid
+    enthalpy. The mass flow at the outlets is expected to split according to
+    the vapor mass fraction:
+
+    .. math::
+
[…]
