+++ tespy/components/__init__.py
@@ -8,7 +8,7 @@ from .heat_exchangers import (
     heat_exchanger_simple, solar_collector,
     condenser, desuperheater, heat_exchanger
     )
-from .nodes import drum, merge, node, separator, splitter
+from .nodes import droplet_separator, drum, merge, node, separator, splitter
 from .piping import pipe, valve
 from .reactors import water_electrolyzer
 from .subsystems import subsystem
