### 2020-04-28 2bb9009
Add one more check to droplet separator example
