### 2020-04-28 e1dd5e5
Add droplet separator network reader's component list
