### 2020-05-03 #2178 @arosen93
@fwitte, @kyleniemeyer -- my review is completed. The code is well-documented with an extensive set of (working) examples, there are robust unit-tests, and there is a clear need for such a program in the community. I am happy to recommend TESPy for publication in JOSS.
