### 2020-05-04 #2178 @fwitte
Hi @corentin-dev,
looking forward to your review. Just to let you know: I am keeping track of the changes made within the review process on this PR: https://github.com/oemof/tespy/pull/181. You might want to have a look at the corresponding documentation (https://tespy--181.org.readthedocs.build/en/181/) instead of the documentation on the master branch (https://tespy.readthedocs.io) as some typos etc. have already been fixed.
Have a nice week, best regards
Francesco
