+++ tespy/components/heat_exchangers.py
@@ -4,12 +4,12 @@
 
 Components in this module:
 
-    - :func:`tespy.components.heat_exchangers.heat_exchanger_simple`
-      (parent class)
-    - :func:`tespy.components.heat_exchangers.heat_exchanger` (parent class)
-    - :func:`tespy.components.heat_exchangers.condenser`
-    - :func:`tespy.components.heat_exchangers.desuperheater`
-    - :func:`tespy.components.heat_exchangers.solar_collector`
+- :func:`tespy.components.heat_exchangers.heat_exchanger`
+- :func:`tespy.components.heat_exchangers.heat_exchanger_simple`
+- :func:`tespy.components.heat_exchangers.condenser`
+- :func:`tespy.components.heat_exchangers.desuperheater`
+- :func:`tespy.components.heat_exchangers.solar_collector`
+- :func:`tespy.components.heat_exchangers.concentrating_solar`
 
 
 This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
@@ -776,9 +776,406 @@ class heat_exchanger_simple(component):
 # %%
 
 
+class parabolic_trough(heat_exchanger_simple):
+    r"""
+    The parabolic trough calculates heat output from irradiance.
+
+    Equations
+
+        **mandatory equations**
+
+        - :func:`tespy.components.components.component.fluid_func`
+        - :func:`tespy.components.components.component.mass_flow_func`
+
+        **optional equations**
+
+        .. math::
+
+            0 = \dot{m}_{in} \cdot \left(h_{out} - h_{in} \right) -
+            \dot{Q}
+
+            0 = p_{in} \cdot pr - p_{out}
+
+        - :func:`tespy.components.components.component.zeta_func`
+
+        - :func:`tespy.components.heat_exchangers.heat_exchanger_simple.darcy_func`
+          or :func:`tespy.components.heat_exchangers.heat_exchanger_simple.hw_func`
+
+        **additional equations**
+
+        - :func:`tespy.components.heat_exchangers.parabolic_trough.additional_equations`
+
+    Inlets/Outlets
+
+        - in1
+        - out1
+
+    Image
+
+        .. image:: _images/parabolic_trough.svg
+           :scale: 100 %
+           :alt: alternative text
+           :align: center
+
+    Parameters
+    ----------
+    label : str
+        The label of the component.
+
+    design : list
+        List containing design parameters (stated as String).
+
+    offdesign : list
+        List containing offdesign parameters (stated as String).
+
+    design_path: str
+        Path to the components design case.
+
+    local_offdesign : boolean
+        Treat this component in offdesign mode in a design calculation.
+
+    local_design : boolean
+        Treat this component in design mode in an offdesign calculation.
+
+    char_warnings: boolean
+        Ignore warnings on default characteristics usage for this component.
+
+    printout: boolean
+        Include this component in the network's results printout.
+
+    Q : str/float/tespy.tools.data_containers.dc_cp
+        Heat transfer, :math:`Q/\text{W}`.
+
+    pr : str/float/tespy.tools.data_containers.dc_cp
+        Outlet to inlet pressure ratio, :math:`pr/1`.
+
+    zeta : str/float/tespy.tools.data_containers.dc_cp
+        Geometry independent friction coefficient,
+        :math:`\frac{\zeta}{D^4}/\frac{1}{\text{m}^4}`.
+
+    D : str/float/tespy.tools.data_containers.dc_cp
+        Diameter of the absorber tube, :math:`D/\text{m}`.
+
+    L : str/float/tespy.tools.data_containers.dc_cp
[…]
