### 2020-05-04 718089a
Add parabolic trough to network reader
