+++ tespy/components/reactors.py
@@ -4,7 +4,7 @@
 
 Components in this module:
 
-    - :func:`tespy.components.reactors.water_electrolyzer`
+- :func:`tespy.components.reactors.water_electrolyzer`
 
 
 This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
