+++ tespy/components/combustion.py
@@ -4,9 +4,9 @@
 
 Components in this module:
 
-    - :func:`tespy.components.combustion.combustion_chamber`
-    - :func:`tespy.components.combustion.combustion_chamber_stoich`
-    - :func:`tespy.components.combustion.combustion_engine`
+- :func:`tespy.components.combustion.combustion_chamber`
+- :func:`tespy.components.combustion.combustion_chamber_stoich`
+- :func:`tespy.components.combustion.combustion_engine`
 
 
 This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
