### 2020-05-04 7df751a
Some modifications for better readabilty of the parabolic example
