### 2020-05-04 8e98244
Improve readability of energy balance equations for solar components
