### 2020-05-04 4efdbd1
Fix typo in example's results
