### 2020-05-04 #2178 @corentin-dev
Sorry, but my invitation to joss-review has been revoked (I took too much time to accept I guess). Can you renew it please ?
```
 Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account.
```
