### 2020-05-04 eb1a232
Fix equation for energy func
