+++ tespy/components/heat_exchangers.py
@@ -1132,9 +1132,11 @@ class parabolic_trough(heat_exchanger_simple):
             T_m = & \frac{T_{out} + T_{in}}{2}\\
             iam = & 1 - iam_1 \cdot |aoi| - iam_2 \cdot aoi^2\\
             0 = & \dot{m} \cdot \left( h_{out} - h_{in} \right)\\
-            & - A \cdot \left[E \cdot \eta_{opt} \cdot doc^1.5 \cdot iam\\
-            & - c_1 \cdot \left(T_m - T_{amb} \right) - c_2 \cdot
-            \left(T_m - T_{amb}\right)^2 \right]
+            & - A \cdot \left[E \cdot \eta_{opt} \cdot doc^{1.5} \cdot
+            iam \right. \\
+            & \left. - c_1 \cdot \left(T_m - T_{amb} \right) -
+            c_2 \cdot \left(T_m - T_{amb}\right)^2
+            \vphantom{ \eta_{opt} \cdot doc^{1.5}} \right]
             \end{split}
 
         Reference: :cite:`Janotte2014`.
