### 2020-05-04 7e920eb
Add Cauchy-Schwarz inequality for linear dependency finding
