### 2020-05-04 a8045cd
Fix typos in characteristic map docs
