### 2020-05-05 d9f8a48
Fix ZeroDivisionError in iteration speed calculation
