+++ doc/introduction.rst
@@ -58,6 +58,27 @@ model a heat pump in TESPy.
 The :ref:`TESPy modules section <tespy_modules_label>` provides you
 with all important information on the different modules of TESPy.
 
+Citation
+========
+
+The scope and functionalities of TESPy have been documented in a paper
+published in the Journal of Open Source Software (link to paper) with an
+OpenAccess license.
+
+As TESPy is a free software, we kindly ask that you add a reference to TESPy if
+you use the software for your scientific work. Please cite the article with the
+BibTeX citation below.
+
+Additionally, you have the possibility to cite a specific version of TESPy to
+make your work reproducible. The source code of every version is published on
+zenodo. Find your version here: https://doi.org/10.5281/zenodo.2555866.
+
+BibTeX citation::
+
+    @article{
+        here goes some data
+    }
+
 License
 =======
 
