+++ doc/api.rst
@@ -1,10 +1,9 @@
-~~~~~~~~~~~~~~~~~~~~~~
+~~~
 API
-~~~~~~~~~~~~~~~~~~~~~~
+~~~
 
 .. toctree::
    :maxdepth: 1
    :glob:
 
    api/*
-
