### 2020-05-12 e3e2ae6
Add min_iter property to network class
