### 2020-05-12 aa228c0
Update docs according to min_iter network property
