### 2020-05-12 55d3868
Use the copy of a dictionary instead of the actual dictionary passed for setting fluid composition values
