### 2020-05-13 0993774
Add kA_char parameter next to kA parameter
