+++ tespy/components/heat_exchangers.py
@@ -165,7 +165,7 @@ class heat_exchanger_simple(component):
     >>> heat_sink.component()
     'heat exchanger simple'
     >>> heat_sink.set_attr(Tamb=10, pr=0.95, design=['pr'],
-    ... offdesign=['zeta', 'kA'])
+    ... offdesign=['zeta', 'kA_char'])
     >>> inc = connection(so1, 'out1', heat_sink, 'in1')
     >>> outg = connection(heat_sink, 'out1', si1, 'in1')
     >>> nw.add_conns(inc, outg)
@@ -206,17 +206,18 @@ class heat_exchanger_simple(component):
 
     @staticmethod
     def attr():
-        return {'Q': dc_cp(),
-                'pr': dc_cp(min_val=1e-4, max_val=1),
-                'zeta': dc_cp(min_val=0),
-                'D': dc_cp(min_val=1e-2, max_val=2, d=1e-4),
-                'L': dc_cp(min_val=1e-1, d=1e-3),
-                'ks': dc_cp(val=1e-4, min_val=1e-7, max_val=1e-3, d=1e-8),
-                'kA': dc_cp(min_val=0, d=1),
-                'Tamb': dc_cp(),
-                'kA_char': dc_cc(param='m'),
-                'SQ1': dc_simple(), 'SQ2': dc_simple(), 'Sirr': dc_simple(),
-                'hydro_group': dc_gcp(), 'kA_group': dc_gcp()}
+        return {
+            'Q': dc_cp(),
+            'pr': dc_cp(min_val=1e-4, max_val=1), 'zeta': dc_cp(min_val=0),
+            'D': dc_cp(min_val=1e-2, max_val=2, d=1e-4),
+            'L': dc_cp(min_val=1e-1, d=1e-3),
+            'ks': dc_cp(val=1e-4, min_val=1e-7, max_val=1e-3, d=1e-8),
+            'kA': dc_cp(min_val=0, d=1),
+            'kA_char': dc_cc(param='m'), 'Tamb': dc_cp(),
+            'SQ1': dc_simple(), 'SQ2': dc_simple(), 'Sirr': dc_simple(),
+            'hydro_group': dc_gcp(), 'kA_group': dc_gcp(),
+            'kA_char_group': dc_gcp()
+        }
 
     @staticmethod
     def inlets():
@@ -249,15 +250,16 @@ class heat_exchanger_simple(component):
                 method = 'Hazen-Williams equation'
             else:
                 method = 'darcy friction factor'
-            msg = ('Pressure loss calculation from pipe dimensions method is '
-                   'set to ' + method + '.')
+            msg = (
+                'Pressure loss calculation from pipe dimensions method is set '
+                'to ' + method + '.')
             logging.debug(msg)
 
         elif self.hydro_group.is_set:
-            msg = ('All parameters of the component group have to be '
-                   'specified! This component group uses the following '
-                   'parameters: L, ks, D at ' + self.label + '. '
-                   'Group will be set to False.')
+            msg = (
+                'All parameters of the component group have to be specified! '
+                'This component group uses the following parameters: L, ks, D '
+                'at ' + self.label + '. Group will be set to False.')
             logging.warning(msg)
             self.hydro_group.set_attr(is_set=False)
         else:
@@ -266,29 +268,43 @@ class heat_exchanger_simple(component):
         # parameters for kA group
         self.kA_group.set_attr(elements=[self.kA, self.Tamb])
 
-        is_set = True
-        for e in self.kA_group.elements:
-            if not e.is_set:
-                is_set = False
+        is_set = self.kA.is_set and self.Tamb.is_set
 
[…]
