### 2020-05-13 da2ec41
Adapt tests to parameter changes in heat_exchagners
