### 2020-05-13 d1d5b80
Fix drum example
