### 2020-05-13 56d1d7f
Adjust tests to work with CoolProp 6.4.0 reference state of NH3
