### 2020-05-13 5d8596f
Remove unused variables
