### 2020-05-13 1e75868
Some typo fixes
