### 2020-05-13 9bee504
Remove possiblity to specify variable ambient temperature
