+++ tespy/components/heat_exchangers.py
@@ -135,9 +135,8 @@ class heat_exchanger_simple(component):
     kA_char : tespy.tools.charactersitics.char_line/tespy.tools.data_containers.dc_cc
         Characteristic line for heat transfer coefficient.
 
-    Tamb : str/float/tespy.tools.data_containers.dc_cp
-        Ambient temperature, provide parameter in network's temperature
-        unit.
+    Tamb : float/tespy.tools.data_containers.dc_simple
+        Ambient temperature, provide parameter in network's temperature unit.
 
     kA_group : tespy.tools.data_containers.dc_gcp
         Parametergroup for heat transfer calculation from ambient temperature
@@ -213,7 +212,7 @@ class heat_exchanger_simple(component):
             'L': dc_cp(min_val=1e-1, d=1e-3),
             'ks': dc_cp(val=1e-4, min_val=1e-7, max_val=1e-3, d=1e-8),
             'kA': dc_cp(min_val=0, d=1),
-            'kA_char': dc_cc(param='m'), 'Tamb': dc_cp(),
+            'kA_char': dc_cc(param='m'), 'Tamb': dc_simple(),
             'SQ1': dc_simple(), 'SQ2': dc_simple(), 'Sirr': dc_simple(),
             'hydro_group': dc_gcp(), 'kA_group': dc_gcp(),
             'kA_char_group': dc_gcp()
@@ -239,10 +238,7 @@ class heat_exchanger_simple(component):
         # parameters for hydro group
         self.hydro_group.set_attr(elements=[self.L, self.ks, self.D])
 
-        is_set = True
-        for e in self.hydro_group.elements:
-            if not e.is_set:
-                is_set = False
+        is_set = self.L.is_set and self.ks.is_set and self.D.is_set
 
         if is_set:
             self.hydro_group.set_attr(is_set=True)
@@ -495,11 +491,9 @@ class heat_exchanger_simple(component):
                 self.jacobian[k, 1, 1] = self.numeric_deriv(f, 'p', 1)
             if not increment_filter[1, 2]:
                 self.jacobian[k, 1, 2] = self.numeric_deriv(f, 'h', 1)
-            # variable Tamb or kA
-            for var in self.kA_group.elements:
-                if var.is_var:
-                    self.jacobian[k, 2 + var.var_pos, 0] = (
-                        self.numeric_deriv(f, self.vars[var], 2))
+            if self.kA.is_var:
+                self.jacobian[k, 2 + self.kA.var_pos, 0] = (
+                    self.numeric_deriv(f, self.vars[self.kA], 2))
             k += 1
 
         ######################################################################
@@ -517,9 +511,6 @@ class heat_exchanger_simple(component):
                 self.jacobian[k, 1, 1] = self.numeric_deriv(f, 'p', 1)
             if not increment_filter[1, 2]:
                 self.jacobian[k, 1, 2] = self.numeric_deriv(f, 'h', 1)
-            if self.Tamb.is_var:
-                self.jacobian[k, 2 + self.Tamb.var_pos, 0] = (
-                    self.numeric_deriv(f, self.vars[self.Tamb], 2))
             k += 1
 
     def darcy_func(self):
@@ -637,6 +628,8 @@ class heat_exchanger_simple(component):
             td_log = (ttd_2 - ttd_1) / np.log(ttd_2 / ttd_1)
         else:
             td_log = 0
+
+        print(ttd_1, ttd_2)
 
         return i[0] * (o[2] - i[2]) + self.kA.val * td_log
 
[…]
