### 2020-05-13 12bade5
Add kA specification tests for condenser and heat exchanger
