### 2020-05-13 93765d2
Remove heat exchanger Tamb.design value
