### 2020-05-13 7f3a89a
Exclude Tamb from variable parameters
