### 2020-05-13 3f432a4
Add warning message for kA implementation
