+++ tespy/components/heat_exchangers.py
@@ -29,6 +29,7 @@ from tespy.tools.fluid_properties import (
         )
 from tespy.tools.global_vars import err
 from tespy.tools.helpers import lamb, single_fluid
+import warnings
 
 # %%
 
