### 2020-05-13 73c7c57
Remove Python 3.7 restriction and adapt CoolProp requirement
