### 2020-05-13 6bc047a
Fix Python 3.8 specific RuntimeError
