### 2020-05-15 ee05ab4
Prompt warning message if the calculated temperature value in postprocessing is infeasible
