### 2020-05-15 0982177
Add isnan check for T0 of T_mix_ph and T_mix_ps
