### 2020-05-15 a2f0370
Reduce accuracy requirement for mixture temperature finding
