### 2020-05-17 #2178 @corentin-dev
**Scientific paper**

The scientific paper is clear but short. It clearly states the motivations, the architecture and the main equations solved. It does not state any major contribution to the field and mainly aims at providing an up to date open source package written in python and using state-of-the-art procedures.

The paper could list component models, fluid models, connections, etc... It does not need to detail everything, but pointing to the right papers / documentation would be nice. The documentation already covers a lot of those details, and it should not take too long to have the paper updated.

Typo : pre- in page 1
Consider using Re with curved `\mathrm{Re}` for Reynolds number even if it is purely author's choice.

**Documentation, usability**

The package is very well documented. It is easy to use. Many examples, tutorials are available and cover many aspects of the software.

The source code respects current standard developpement in python.

**Global opinion**

The sofware is suitable to be published in JOSS. The paper can be improved, but overall everything is in the documentation which is extensive and provides lots of well written tutorials and examples.
