### 2020-05-18 9edf4c9
Remove T_range from network class
