### 2020-05-18 bf9d571
Remove convergence check for heat exchangers
