### 2020-05-18 eb5606e
Clean up value range handling in set_attr
