### 2020-05-18 8eccb7e
Remove T_range from docs
