### 2020-05-18 3e23322
Remove T_range, add T_mix_ps error test
