### 2020-05-18 255eac9
Remove temperature range requiremetns
