### 2020-05-19 #2178 @fwitte
Hi,
thank you for the review. I'll address your comments until tomorrow and will prepare the release thereafter. Some quick questions/remarks:

> The paper could list component models, fluid models, connections, etc... It does not need to detail everything, but pointing to the right papers / documentation would be nice. The documentation already covers a lot of those details, and it should not take too long to have the paper updated.

I wanted to stay within the 1000 word limit. As adding more formulas of the components available (e.g. parababolic trough and solar collector energy balances, stodolas cone law applied for turbines, compressor maps, ...) would add a lot of content to the paper, is a link with a short description, where the respective formulas are located in the documentation sufficient?

> Typo : pre- in page 1

Is it "pre and postprocessing" in english?

> Consider using Re with curved `\mathrm{Re}` for Reynolds number even if it is purely author's choice.

Is that not the opposite of curved? See https://latex.wikia.org/wiki/Mathrm_(LaTeX_symbol).
