### 2020-05-19 6c2adb1
Update the JOSS paper after review of corentin-dev
Add links to TESPy online documentation and API-docs.
Add fluid property handling information.
Address typo and suggested changes in Reynolds
