+++ paper.bib
@@ -54,6 +54,13 @@
   howpublished = {\url{https://build.openmodelica.org/Documentation/ThermoPower.html}}
 }
 
+@misc{TESPy,
+  author = {Witte, Francesco}
+  title = {TESPy},
+  year = {2020},
+  howpublished = {\url{https://tespy.readthedocs.io/}}
+}
+
 @article{NumPy,
     doi = {10.1109/mcse.2011.37},
     year = {2011},
@@ -144,4 +151,14 @@
   author = {Wolf Tilmann Pfeiffer and Christof Beyer and Sebastian Bauer},
   title = {Hydrogen storage in a heterogeneous sandstone formation: dimensioning and induced hydraulic effects},
   journal = {Petroleum Geoscience}
+}
+
+@article{Herning1936,
+    author = {Herning, Fritz and Zipperer, L},
+    title = {Calculation of the Viscosity of Technical Gas Mixtures From the
+        Viscosity of the Individual Gases},
+    journal = {Gas-und Wasserfach},
+    volume = {79},
+    pages = {69-73},
+    year = {2015}
 }
