+++ paper.bib
@@ -48,17 +48,17 @@
 }
 
 @misc{ThermoPowerOnline,
-  author = {​{Politecnico di Milano}},
-  title = {ThermoPower},
-  year = {2014},
-  howpublished = {\url{https://build.openmodelica.org/Documentation/ThermoPower.html}}
+    author = {​{Politecnico di Milano}},
+    title = {ThermoPower},
+    year = {2014},
+    howpublished = {\url{https://build.openmodelica.org/Documentation/ThermoPower.html}}
 }
 
 @misc{TESPy,
-  author = {Witte, Francesco}
-  title = {TESPy},
-  year = {2020},
-  howpublished = {\url{https://tespy.readthedocs.io/}}
+    author = {Witte, Francesco},
+    title = {TESPy},
+    year = {2020},
+    howpublished = {\url{https://tespy.readthedocs.io/}}
 }
 
 @article{NumPy,
