### 2020-05-19 #2178 @fwitte
@corentin-dev, thank you for your fast reply. I updated the manuscript and hope it suits your remarks well.

There is only one issue, that I am not sure of: What about the syllabification of "initialisation"? Is it correct or do I need to apply something manually here?
