### 2020-05-19 #2178 @kyleniemeyer
Hi @fwitte, I have a few editorial comments on the paper:

- [ ] For a few citations (such as Witte 2020 in the Summery and Casella & Leva 2003 in the Motivation paragraph, I think you want to change the citations from the `[@author:year]` format to the `@author:year` format, which will make the citation look like "Author et al. (year)".
- [ ] In the second paragraph of Motivation, the first sentence could be improved a bit; perhaps something like "Here we propose a new solution for simulating thermal processes, since ThermoPower is not currently maintained and also limited to one field of thermal engineering." Then, in the second sentence, the "is very promising" at the end seems out of place, and I think should just be removed.
- [ ] In the second paragraph of Method, you have nested parentheses with the citations; perhaps just bring the "based on standard literature..." to the main level of the sentence.
- [ ] Just after that, "component specific" needs a hyphen, I think.
- [ ] Just after _that_, I don't think you mean that CoolProp is "implemented", but rather that it is used. Perhaps the sentence could be edited to something like "For calculating fluid properties, TESPy uses CoolProp (Bell ...), which supports a wide range of pure and incompressible fluids."
- [ ] Just above equation 5, a sentence starts with "E.g."; this is a bit awkward, and perhaps could just be "For example, for an adiabatic ..."
- [ ] The number "7" appears right after "pressure ratio pr" just before equation 6. I think this sentence should be something like "A simpler approach is to specify the pressure ratio pr:" (no need to write the equation number when the equation appears next)
