### 2020-05-20 f6439ef
Update paper according to comments of kyleniemeyer
