### 2020-05-20 #2178 @fwitte
Thank you for your comments and suggestions @kyleniemeyer, I have addressed all of them over here https://github.com/oemof/tespy/pull/181.
