### 2020-05-20 350ec3a
Update current state to cookiecutter skeleton
