+++ tespy/tools/fluid_properties.py
@@ -1,1762 +1,0 @@
-# -*- coding: utf-8
-
-"""Module for fluid property integration.
-
-TESPy uses the CoolProp python interface for all fluid property functions. The
-tespy_fluid class allows the creation of lookup table for custom fluid
-mixtures.
-
-
-This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
-by the contributors recorded in the version control history of the file,
-available from its original location tespy/tools/fluid_properties.py
-
-SPDX-License-Identifier: MIT
-"""
-
-from tespy.tools.helpers import (
-        num_fluids, molar_mass_flow, single_fluid,
-        newton, reverse_2d, reverse_2d_deriv
-        )
-
-from tespy.tools.global_vars import (
-        err, molar_masses, gas_constants
-        )
-
-import CoolProp as CP
-from CoolProp.CoolProp import PropsSI as CPPSI
-
-import math
-import numpy as np
-from scipy import interpolate
-import pandas as pd
-import os
-from collections import OrderedDict
-
-import logging
-
-# %%
-
-
-class tespy_fluid:
-    r"""
-    Allow the creation of lookup tables for specified mixtures.
-
-    The created fluid property lookup table adress an ideal mixture of real
-    fluids.
-
-    Parameters
-    ----------
-    alias : str
-        Alias for the fluid. Please note: The alias of a tespy_fluid class
-        object will always start with :code:`TESPy::`. See the example for more
-        information!
-
-    fluid : dict
-        Fluid vector specifying the fluid composition of the TESPy fluid.
-
-    p_range : list/ndarray
-        Pressure range for the new fluid lookup table.
-
-    path : str
-        Path to importing tespy fluid from.
-
-    Note
-    ----
-    Creates lookup tables for
-
-    - enthalpy (h),
-    - entropy (s),
-    - density (d) and
-    - viscoity (visc)
-
-    from pressure and temperature. Additionally molar mass and gas constant
-    will be calculated. Inverse functions, e.g. entropy from pressure and
-    enthalpy are calculated via newton algorithm from these tables.
-
-    Example
-    -------
-    Create a custom fluid from specified composition within defined pressure
-    and temperature limits. We define dry air component wise as our custom
-    fluid.
-
-    >>> from tespy.connections import connection
-    >>> from tespy.components.basics import sink, source
-    >>> from tespy.tools.fluid_properties import (tespy_fluid, h_mix_pT,
-    ... s_mix_pT, v_mix_pT, visc_mix_pT)
-    >>> from tespy.networks.networks import network
-    >>> import shutil
-    >>> fluidvec = {'N2': 0.7552, 'O2': 0.2314, 'CO2': 0.0005, 'Ar': 0.0129}
-    >>> p_arr = np.array([0.1, 10]) * 1e5
-    >>> myfluid = tespy_fluid('dry air', fluid=fluidvec, p_range=p_arr,
-    ... T_range=[300, 1200])
-
-    Check if the fluid creation was successful and compare some fluid
-    properties to the CoolProp air implementation. We have to add the CoolProp
-    air implementation to the memorise class first. The relative deviation
-    should be very small (< 0.01), we check for enthalpy, volume, entropy and
-    viscosity. Specific volume and viscosity are absolute values, thus no
-    difference is calculated.
-
-    >>> memorise.add_fluids({'air': 'HEOS'})
-
-    >>> type(myfluid)
-    <class 'tespy.tools.fluid_properties.tespy_fluid'>
-    >>> p = 3e5
-    >>> fluid_props = [0, p, 0, {myfluid.alias: 1}]
-    >>> T1 = 400
-    >>> T2 = 1000
-    >>> delta_h_tespy = h_mix_pT(fluid_props, T2) - h_mix_pT(fluid_props, T1)
[…]
