+++ src/tespy/components/basics.py
@@ -1,0 +1,507 @@
+# -*- coding: utf-8
+
+"""Module for basic components.
+
+Components in this module:
+
+- :func:`tespy.components.basics.source`
+- :func:`tespy.components.basics.sink`
+- :func:`tespy.components.basics.subsystem_interface`
+- :func:`tespy.components.basics.cycle_closer`
+
+
+This file is part of project TESPy (github.com/oemof/tespy). It's copyrighted
+by the contributors recorded in the version control history of the file,
+available from its original location tespy/components/basics.py
+
+SPDX-License-Identifier: MIT
+"""
+
+import numpy as np
+
+from tespy.components.components import component
+from tespy.tools.data_containers import dc_cp
+from tespy.tools.data_containers import dc_simple
+
+# %%
+
+
+class cycle_closer(component):
+    r"""
+    Component for closing cycles.
+
+    Equations
+
+        **mandatory equations**
+
+        .. math::
+
+            0 = p_{in} - p_{out}
+
+            0 = h_{in} - h_{out}
+
+    Image not available
+
+    Parameters
+    ----------
+    label : str
+        The label of the component.
+
+    design : list
+        List containing design parameters (stated as String).
+
+    offdesign : list
+        List containing offdesign parameters (stated as String).
+
+    design_path: str
+        Path to the components design case.
+
+    local_offdesign : boolean
+        Treat this component in offdesign mode in a design calculation.
+
+    local_design : boolean
+        Treat this component in design mode in an offdesign calculation.
+
+    char_warnings: boolean
+        Ignore warnings on default characteristics usage for this component.
+
+    printout: boolean
+        Include this component in the network's results printout.
+
+    Note
+    ----
+    This component can be used to close a cycle process. The system of
+    equations describing your plant will overdetermined, if you close a cycle
+    without this component or a cut the cycle with a sink and a source at
+    some point of the cycle. This component can be used instead of cutting
+    the cycle.
+
+    Example
+    -------
+    Create a cycle containing a pump and a pipe. The pump increases pressure
+    the pipe cools the liquid and destroys the pressure rise. The heat
+    extracted at the pipe must be the same value of the power input at the
+    pump (but negative), as there is no other in- or outputs of energy in the
+    system.
+
+    >>> from tespy.components import cycle_closer, pipe, pump
+    >>> from tespy.connections import connection
+    >>> from tespy.networks import network
+    >>> nw = network(['water'], p_unit='bar', T_unit='C', iterinfo=False)
+    >>> pi = pipe('pipe')
+    >>> pu = pump('pump')
+    >>> cc = cycle_closer('cycle closing component')
+    >>> cc.component()
+    'cycle closer'
+    >>> pu_pi = connection(pu, 'out1', pi, 'in1')
+    >>> pi_cc = connection(pi, 'out1', cc, 'in1')
+    >>> cc_pu = connection(cc, 'out1', pu, 'in1')
+    >>> nw.add_conns(pu_pi, pi_cc, cc_pu)
+    >>> pi_cc.set_attr(p=1, T=20, fluid={'water': 1})
+    >>> pu_pi.set_attr(p=10)
+    >>> pu.set_attr(eta_s=0.8, P=1000)
+    >>> nw.solve('design')
+    >>> round(pi.Q.val, 1) == -round(pu.P.val, 1)
+    True
+    """
+
+    @staticmethod
+    def component():
+        return 'cycle closer'
+
+    @staticmethod
+    def attr():
+        return {
+            'mass_deviation': dc_cp(val=0, max_val=1e-3, printout=False),
+            'fluid_deviation': dc_cp(val=0, max_val=1e-5, printout=False)
+        }
+
[…]
