### 2020-05-20 3588cab
Fix typos in badges
