### 2020-05-20 1e0d844
Fix some PEP issues
