+++ src/tespy/networks/networks.py
@@ -29,8 +29,6 @@ from tabulate import tabulate
 
 from tespy import connections as con
 from tespy.components.basics import cycle_closer
-from tespy.components.basics import sink
-from tespy.components.basics import source
 from tespy.components.basics import subsystem_interface
 from tespy.components.combustion import combustion_chamber
 from tespy.components.combustion import combustion_engine
@@ -508,7 +506,7 @@ class network:
     def check_conns(self):
         r"""Check connections for multiple usage of inlets or outlets."""
         dub = self.conns.loc[
-            self.conns.duplicated(['source', 'source_id']) == True]  # noqa: F712
+            self.conns.duplicated(['source', 'source_id']) == True]  # noqa: E712
         for c in dub.index:
             targets = ''
             for conns in self.conns[
@@ -525,7 +523,7 @@ class network:
             raise hlp.TESPyNetworkError(msg)
 
         dub = self.conns.loc[
-            self.conns.duplicated(['target', 'target_id']) == True]  # noqa: F712
+            self.conns.duplicated(['target', 'target_id']) == True]  # noqa: E712
         for c in dub.index:
             sources = ''
             for conns in self.conns[
@@ -1232,7 +1230,7 @@ class network:
             outc['source'] = self.conns['source'] == c.target
             outc['source_id'] = (
                 self.conns['source_id'] == c.target_id.replace('in', 'out'))
-            conn, cid = outc['source'] == True, outc['source_id'] == True  # noqa: F712
+            conn, cid = outc['source'] == True, outc['source_id'] == True  # noqa: E712
             outconn = outc.index[conn & cid][0]
 
             for fluid, x in c.fluid.val.items():
@@ -1309,7 +1307,7 @@ class network:
             inc['target'] = self.conns['target'] == c.source
             inc['target_id'] = (
                 self.conns['target_id'] == c.source_id.replace('out', 'in'))
-            conn, cid = inc['target'] == True, inc['target_id'] == True  # noqa: F712
+            conn, cid = inc['target'] == True, inc['target_id'] == True  # noqa: E712
             inconn = inc.index[conn & cid][0]
 
             for fluid, x in c.fluid.val.items():
@@ -1818,7 +1816,7 @@ class network:
 
             else:
                 if np.isnan(norm(self.residual)):
-                    msg += '\t|      nan'.format(norm(self.residual))
+                    msg += '\t|      nan'
                 else:
                     msg += '\t| ' + '{:.2e}'.format(norm(self.residual))
                 msg += ' |      nan'
