### 2020-05-20 4ca45cf
Fix equation rendering
