+++ docs/api/tespy.data.rst
@@ -1,181 +1,0 @@
-tespy.data module
-=================
-
-Module contents
----------------
-
-.. automodule:: tespy.data
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-Default characteristics
------------------------
-
-Characteristic lines
-^^^^^^^^^^^^^^^^^^^^
-**turbine**
-
-.. figure:: _images/turbine_eta_s_char_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "eta_s_char".
-    :align: center
-
-    Reference: Generic data.
-
-.. figure:: _images/turbine_eta_s_char_TRAUPEL.svg
-    :scale: 100 %
-    :alt: Characteristic line "TRAUPEL" for parameter "eta_s_char".
-    :align: center
-
-    Reference: :cite:`Traupel2001`.
-
-**compressor**
-
-.. figure:: _images/compressor_eta_s_char_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "eta_s_char".
-    :align: center
-
-    Reference: Generic data.
-
-**pump**
-
-.. figure:: _images/pump_eta_s_char_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "eta_s_char".
-    :align: center
-
-    Reference: Generic data.
-
-**combustion engine**
-
-.. figure:: _images/combustion_engine_tiP_char_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "tiP_char".
-    :align: center
-
-    Reference: Generic data.
-
-.. figure:: _images/combustion_engine_Q1_char_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "Q1_char".
-    :align: center
-
-    Reference: Generic data.
-
-.. figure:: _images/combustion_engine_Q2_char_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "Q2_char".
-    :align: center
-
-    Reference: Generic data.
-
-.. figure:: _images/combustion_engine_Qloss_char_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "Qloss_char".
-    :align: center
-
-    Reference: Generic data.
-
-**heat exchanger**
-
-.. figure:: _images/heat_exchanger_kA_char1_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "kA_char1".
-    :align: center
-
-    Reference: Generic data.
-
-.. figure:: _images/heat_exchanger_kA_char1_CONDENSING_FLUID.svg
-    :scale: 100 %
-    :alt: Characteristic line "CONDENSING FLUID" for parameter "kA_char1".
-    :align: center
-
-    Reference: Generic data.
-
-.. figure:: _images/heat_exchanger_kA_char2_DEFAULT.svg
-    :scale: 100 %
-    :alt: Characteristic line "DEFAULT" for parameter "kA_char2".
-    :align: center
-
-    Reference: Generic data.
-
-.. figure:: _images/heat_exchanger_kA_char2_EVAPORATING_FLUID.svg
-    :scale: 100 %
-    :alt: Characteristic line "EVAPORATING FLUID" for parameter "kA_char2".
-    :align: center
-
-    Reference: Generic data.
-
-**condenser**
-
-.. figure:: _images/condenser_kA_char1_DEFAULT.svg
-    :scale: 100 %
[…]
