+++ setup.cfg
@@ -104,6 +104,7 @@ skip = migrations
 python_versions =
     py36
     py37
+    py38
 
 dependencies =
 #    1.4: Django==1.4.16 !python_versions[py3*]
