### 2020-05-20 #2178 @kyleniemeyer
@fwitte almost done! The references are just missing a few details here and there:

- add volume and pages to the Chen et al. 2019 reference
- I think the year for the Herning and Zipperer reference should be 1936, not 2015
- For the IEA technology roadmap, please provide a URL or other information
- The Virtanen 2020 citation (for SciPy) is missing some info (volume, page numbers)

When you've addressed these, we are ready to accept! Please merge the JOSS review branch into master, archive the repo (e.g., with Zenodo), and let me know the DOI.
