### 2020-05-20 #2178 @fwitte
@kyleniemeyer, great! I will do so tomorrow morning. Just one question: Is it possible to add the bibtex entry and the doi of the JOSS paper to the online documentation prior to the release? If so, I would not need to make a second release after the paper is published (as I usually only commit to master in case I release a new version).
Cheers
