### 2020-05-20 #2178 @kyleniemeyer
@fwitte sure, the DOI will be what shows up in the paper draft, though it does not yet resolve.
