+++ paper.bib
@@ -4,7 +4,7 @@
     journal = {Industrial \& Engineering Chemistry Research},
     volume = {53},
     number = {6},
-    pages = {2498-2508},
+    pages = {2498--2508},
     year = {2014},
     doi = {10.1021/ie4033999}
 }
@@ -34,15 +34,17 @@
     title = {SciPy 1.0: Fundamental Algorithms for Scientific
              Computing in Python},
     journal = {Nature Methods},
-    year = "2020",
+    year = {2020},
+    volume={17},
+    pages={261--272},
     doi = {10.1038/s41592-019-0686-2},
 }
 
 @article{ThermoPower,
     author = {Casella, Francesco and Leva, Alberto},
     year = {2003},
-    month = {11},
-    pages = {41-50},
+    month = nov,
+    pages = {41--50},
     title = {Modelica open library for power plant simulation: Design and experimental validation},
     journal = {Proc. 3rd Modelica Conference}
 }
@@ -75,15 +77,15 @@
 }
 
 @article{BTES,
-    author = {Chen, Shuang and Witte, Francesco and Kolditz, Olaf and Shao,
-    Haibing},
-    year = {2019},
-    month = {12},
-    pages = {},
-    title = {Shifted thermal extraction rates in large Borehole Heat Exchanger
-    array -a numerical experiment},
-    journal = {Applied Thermal Engineering},
-    doi = {10.1016/j.applthermaleng.2019.114750}
+  doi = {10.1016/j.applthermaleng.2019.114750},
+  year = {2020},
+  month = feb,
+  publisher = {Elsevier {BV}},
+  volume = {167},
+  author = {Chen, Shuang and Witte, Francesco and Kolditz, Olaf and Shao,
+  Haibing},
+  title = {Shifted thermal extraction rates in large Borehole Heat Exchanger array {\textendash} A numerical experiment},
+  journal = {Applied Thermal Engineering}
 }
 
 @article{CAES,
@@ -133,11 +135,12 @@
     title = {Simulation von Kraftwerken und Feuerungen}
 }
 
-@book{IEA,
+@techreport{IEA,
     year = {2014},
-    publisher = {IEA},
-    editor = {{International Energy Agency}},
-    title = {Technology Roadmap - Energy Storage}
+    author = {{International Energy Agency - IEA}},
+    title = {Technology Roadmap - Energy Storage},
+    month = mar,
+    note = {\url{https://www.iea.org/reports/technology-roadmap-energy-storage}}
 }
 
 @article{Pfeiffer,
@@ -159,6 +162,6 @@
         Viscosity of the Individual Gases},
     journal = {Gas-und Wasserfach},
     volume = {79},
-    pages = {69-73},
-    year = {2015}
+    pages = {69--73},
+    year = {1936}
 }
