+++ docs/references.bib
@@ -49,7 +49,7 @@
     journal = {Gas-und Wasserfach},
     volume = {79},
     pages = {69-73},
-    year = {2015}
+    year = {1936}
 }
 
 @Inbook{Nirschl2018,
