+++ paper.bib
@@ -9,12 +9,14 @@
     doi = {10.1021/ie4033999}
 }
 
-@article{pandas,
-    title={pandas: a foundational Python library for data analysis and statistics},
-    author={Wes McKinney},
-    journal={Python for High Performance and Scientific Computing},
-    volume={14},
-    year={2011}
+@InProceedings{pandas,
+    author = {{M}c{K}inney, Wes},
+    title = {{D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython},
+    booktitle = {{P}roceedings of the 9th {P}ython in {S}cience {C}onference},
+    pages = {56--61},
+    year = {2010},
+    editor = {{S}t\'efan van der {W}alt and {J}arrod {M}illman},
+    doi = {10.25080/Majora-92bf1922-00a}
 }
 
 @article{SciPy,
@@ -77,15 +79,15 @@
 }
 
 @article{BTES,
-  doi = {10.1016/j.applthermaleng.2019.114750},
-  year = {2020},
-  month = feb,
-  publisher = {Elsevier {BV}},
-  volume = {167},
-  author = {Chen, Shuang and Witte, Francesco and Kolditz, Olaf and Shao,
-  Haibing},
-  title = {Shifted thermal extraction rates in large Borehole Heat Exchanger array {\textendash} A numerical experiment},
-  journal = {Applied Thermal Engineering}
+    doi = {10.1016/j.applthermaleng.2019.114750},
+    year = {2020},
+    month = feb,
+    publisher = {Elsevier {BV}},
+    volume = {167},
+    author = {Chen, Shuang and Witte, Francesco and Kolditz, Olaf and Shao,
+    Haibing},
+    title = {Shifted thermal extraction rates in large Borehole Heat Exchanger array {\textendash} A numerical experiment},
+    journal = {Applied Thermal Engineering}
 }
 
 @article{CAES,
@@ -107,23 +109,30 @@
 }
 
 @article{oemof,
-  doi = {10.1016/j.esr.2018.07.001},
-  year = {2018},
-  month = nov,
-  publisher = {Elsevier {BV}},
-  volume = {22},
-  pages = {16--25},
-  author = {Hilpert, Simon and Kaldemeyer, Cord and Krien, Uwe and G\"{u}nther, Stephan and Wingenbach, Clemens and Plessmann, Guido},
-  title = {The Open Energy Modelling Framework (oemof) - A new approach to facilitate open science in energy system modelling},
-  journal = {Energy Strategy Reviews}
+    doi = {10.1016/j.esr.2018.07.001},
+    year = {2018},
+    month = nov,
+    publisher = {Elsevier {BV}},
+    volume = {22},
+    pages = {16--25},
+    author = {Hilpert, Simon and Kaldemeyer, Cord and Krien, Uwe and G\"{u}nther, Stephan and Wingenbach, Clemens and Plessmann, Guido},
+    title = {The Open Energy Modelling Framework (oemof) - A new approach to facilitate open science in energy system modelling},
+    journal = {Energy Strategy Reviews}
 }
 
 @misc{ogs,
-  doi = {10.5281/ZENODO.591265},
[…]
