+++ paper.bib
@@ -148,7 +148,6 @@
     year = {2014},
     author = {{International Energy Agency - IEA}},
     title = {Technology Roadmap - Energy Storage},
-    month = mar,
     howpublished = {\url{https://www.iea.org/reports/technology-roadmap-energy-storage}}
 }
 
