### 2020-05-21 #2178 @fwitte
>* add volume and pages to the Chen et al. 2019 reference

I have looked this up, there is no pages information available. Only an article number (114750): https://www.sciencedirect.com/journal/applied-thermal-engineering/vol/167/suppl/C.

>* I think the year for the Herning and Zipperer reference should be 1936, not 2015

Adjusted, thank you!

>* For the IEA technology roadmap, please provide a URL or other information

Done.

>* The Virtanen 2020 citation (for SciPy) is missing some info (volume, page numbers)

Also added the missing info.

> When you've addressed these, we are ready to accept! Please merge the JOSS review branch into master, archive the repo (e.g., with Zenodo), and let me know the DOI.

I will prepare the release and let you know, when it is done.
