+++ paper.bib
@@ -84,6 +84,7 @@
     month = feb,
     publisher = {Elsevier {BV}},
     volume = {167},
+    pages = {114750},
     author = {Chen, Shuang and Witte, Francesco and Kolditz, Olaf and Shao,
     Haibing},
     title = {Shifted thermal extraction rates in large Borehole Heat Exchanger array {\textendash} A numerical experiment},
