### 2020-05-21 e62ccb6
Add citation info to README and online documentation
