+++ README.rst
@@ -118,9 +118,9 @@ Citation
 ========
 
 The scope and functionalities of TESPy have been documented in a paper
-published in the Journal of Open Source Software
-(https://doi.org/10.21105/joss.02178) with an OpenAccess license. As TESPy is a
-free software, we kindly ask that you add a reference to TESPy if you use the
+published in the Journal of Open Source Software with an OpenAccess license.
+Download the paper from https://doi.org/10.21105/joss.02178. As TESPy is a free
+software, we kindly ask that you add a reference to TESPy if you use the
 software for your scientific work. Please cite the article with the BibTeX
 citation below.
 
