### 2020-05-21 db585fe
Update BibTeX and add version alias
