+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.3.0 dev
+current_version = 0.3.0
 commit = True
 tag = True
 
