### 2020-05-21 a6b30f2
Simplify some printout/logging methods
