+++ src/tespy/networks/networks.py
@@ -1677,8 +1677,8 @@ class network:
         self.start_time = time()
         self.progress = True
 
-        if self.iterinfo:
-            self.print_iterinfo('start')
+        if self.iterinfo is True:
+            self.print_iterinfo_head()
 
         for self.iter in range(self.max_iter):
 
@@ -1686,8 +1686,8 @@ class network:
             self.solve_control()
             self.res = np.append(self.res, norm(self.residual))
 
-            if self.iterinfo:
-                self.print_iterinfo('solving')
+            if self.iterinfo is True:
+                self.print_iterinfo_body()
 
             if ((self.iter >= self.min_iter and self.res[-1] < err ** 0.5) or
                     self.lin_dep):
@@ -1701,7 +1701,7 @@ class network:
 
         self.end_time = time()
 
-        self.print_iterinfo('end')
+        self.print_iterinfo_tail()
 
         if self.iter == self.max_iter - 1:
             msg = ('Reached maximum iteration count (' + str(self.max_iter) +
@@ -1775,81 +1775,73 @@ class network:
             logging.error(msg)
             raise hlp.TESPyNetworkError(msg)
 
-    def print_iterinfo(self, position):
-        """Print out the convergence progress."""
-        if position == 'start':
-            if self.num_comp_vars == 0:
-                # iterinfo printout without any custom variables
-                msg = ('iter\t| residual | massflow | pressure | enthalpy | '
-                       'fluid\n')
-                msg += ('--------+----------+----------+----------+----------'
-                        '+---------')
+    def print_iterinfo_head(self):
+        """Print head of convergence progress."""
+        if self.num_comp_vars == 0:
+            # iterinfo printout without any custom variables
+            msg = (
+                'iter\t| residual | massflow | pressure | enthalpy | fluid\n')
+            msg += '-' * 8 + '+----------' * 4 + '+' + '-' * 9
 
+        else:
+            # iterinfo printout with custom variables in network
+            msg = ('iter\t| residual | massflow | pressure | enthalpy | '
+                   'fluid    | custom\n')
+            msg += '-' * 8 + '+----------' * 5 + '+' + '-' * 9
+
+        print(msg)
+
+    def print_iterinfo_body(self):
+        """Print convergence progress."""
+        vec = self.increment[0:-(self.num_comp_vars + 1)]
+        msg = (str(self.iter + 1))
+
+        if self.lin_dep is False and not np.isnan(norm(self.residual)):
+            msg += '\t| ' + '{:.2e}'.format(norm(self.residual))
+            msg += ' | ' + '{:.2e}'.format(norm(vec[0::self.num_conn_vars]))
+            msg += ' | ' + '{:.2e}'.format(norm(vec[1::self.num_conn_vars]))
+            msg += ' | ' + '{:.2e}'.format(norm(vec[2::self.num_conn_vars]))
+
+            ls = []
+            for f in range(len(self.fluids)):
+                ls += vec[3 + f::self.num_conn_vars].tolist()
+
+            msg += ' | ' + '{:.2e}'.format(norm(ls))
+
+            if self.num_comp_vars > 0:
+                msg += ' | ' + '{:.2e}'.format(norm(
+                    self.increment[-self.num_comp_vars:]))
+
+        else:
+            if np.isnan(norm(self.residual)):
+                msg += '\t|      nan'
             else:
-                # iterinfo printout with custom variables in network
-                msg = ('iter\t| residual | massflow | pressure | enthalpy | '
-                       'fluid    | custom\n')
-                msg += ('--------+----------+----------+----------+----------'
-                        '+----------+---------')
-
-            print(msg)
-
[…]
