### 2020-05-21 5e706bd
More adjustments to loggings
