### 2020-05-21 3260f65
Fix logging message for solver properties
