### 2020-05-21 445f605
Delete authors.rst
