+++ AUTHORS.rst
@@ -1,5 +1,0 @@
-
-Authors
-=======
-
-* Francesco Witte - https://oemof.org
