### 2020-05-21 8ae8875
Remove unnecessary variable definition
