### 2020-05-21 #2178 @fwitte
Hi @kyleniemeyer,
version has been published, archive to be found here: https://zenodo.org/record/3837555.
Thank to all of you for your efforts and support :tada: :beers:!
Best
Francesco
